Nginx_Nctionality

Documentation

accept_mutex :: [Text] -> ConfStatement

If accept_mutex is enabled, worker processes will accept new connections by turn. Otherwise, all worker processes will be notified about new connections, and if volume of new connections is low, some of the worker processes may just waste system resources. Prior to version 1.11.3, the default value was on.

accept_mutex_s :: [Text] -> Accept_Mutex

Safe version of accept_mutex

newtype Accept_Mutex

Constructors

Accept_Mutex ConfStatement

accept_mutex_delay :: [Text] -> ConfStatement

If is enabled, specifies the maximum time during which a worker process will try to restart accepting new connections if another worker process is currently accepting new connections.

accept_mutex_delay_s :: [Text] -> Accept_Mutex_Delay

Safe version of accept_mutex_delay

newtype Accept_Mutex_Delay

Constructors

Accept_Mutex_Delay ConfStatement

daemon :: [Text] -> ConfStatement

Determines whether nginx should become a daemon. Mainly used during development.

daemon_s :: [Text] -> Daemon

Safe version of daemon

newtype Daemon

Constructors

Daemon ConfStatement

debug_connection :: [Text] -> ConfStatement

Enables debugging log for selected client connections. Other connections will use logging level set by the directive. Debugged connections are specified by IPv4 or IPv6 (1.3.0, 1.2.1) address or network. A connection may also be specified using a hostname. For connections using UNIX-domain sockets (1.3.0, 1.2.1), debugging log is enabled by the “unix:” parameter. events { debug_connection 127.0.0.1; debug_connection localhost; debug_connection 192.0.2.024; debug_connection ::1; debug_connection 2001:0db8::32; debug_connection unix:; ... } For this directive to work, nginx needs to be built with --with-debug, see “”.

debug_connection_s :: [Text] -> Debug_Connection

Safe version of debug_connection

newtype Debug_Connection

Constructors

Debug_Connection ConfStatement

debug_points :: [Text] -> ConfStatement

This directive is used for debugging.

debug_points_s :: [Text] -> Debug_Points

Safe version of debug_points

newtype Debug_Points

Constructors

Debug_Points ConfStatement

error_log :: [Text] -> ConfStatement

Configures logging. Several logs can be specified on the same level (1.5.2).

error_log_s :: [Text] -> Error_Log

Safe version of error_log

newtype Error_Log

Constructors

Error_Log ConfStatement

env :: [Text] -> ConfStatement

By default, nginx removes all environment variables inherited from its parent process except the TZ variable. This directive allows preserving some of the inherited variables, changing their values, or creating new environment variables. These variables are then: inherited during a live upgrade of an executable file; used by the ngx_http_perl_module module; used by worker processes. One should bear in mind that controlling system libraries in this way is not always possible as it is common for libraries to check variables only during initialization, well before they can be set using this directive. An exception from this is an above mentioned live upgrade of an executable file.

env_s :: [Text] -> Env

Safe version of env

newtype Env

Constructors

Env ConfStatement

events :: [Text] -> [ConfStatement] -> ConfStatement

Provides the configuration file context in which the directives that affect connection processing are specified.

events_s :: (HasContextEvents f, ToConfStatement f) => [Text] -> [f] -> ConfStatement

Safe version of events

include :: [Text] -> ConfStatement

Includes another file, or files matching the specified mask, into configuration. Included files should consist of syntactically correct directives and blocks.

include_s :: [Text] -> Include

Safe version of include

newtype Include

Constructors

Include ConfStatement

load_module :: [Text] -> ConfStatement

Loads a dynamic module.

load_module_s :: [Text] -> Load_Module

Safe version of load_module

newtype Load_Module

Constructors

Load_Module ConfStatement

lock_file :: [Text] -> ConfStatement

nginx uses the locking mechanism to implement and serialize access to shared memory. On most systems the locks are implemented using atomic operations, and this directive is ignored. On other systems the “lock file” mechanism is used. This directive specifies a prefix for the names of lock files.

lock_file_s :: [Text] -> Lock_File

Safe version of lock_file

newtype Lock_File

Constructors

Lock_File ConfStatement

master_process :: [Text] -> ConfStatement

Determines whether worker processes are started. This directive is intended for nginx developers.

master_process_s :: [Text] -> Master_Process

Safe version of master_process

newtype Master_Process

Constructors

Master_Process ConfStatement

multi_accept :: [Text] -> ConfStatement

If multi_accept is disabled, a worker process will accept one new connection at a time. Otherwise, a worker process will accept all new connections at a time. The directive is ignored if connection processing method is used, because it reports the number of new connections waiting to be accepted.

multi_accept_s :: [Text] -> Multi_Accept

Safe version of multi_accept

newtype Multi_Accept

Constructors

Multi_Accept ConfStatement

pcre_jit :: [Text] -> ConfStatement

Enables or disables the use of “just-in-time compilation” (PCRE JIT) for the regular expressions known by the time of configuration parsing.

pcre_jit_s :: [Text] -> Pcre_Jit

Safe version of pcre_jit

newtype Pcre_Jit

Constructors

Pcre_Jit ConfStatement

pid :: [Text] -> ConfStatement

Defines a file that will store the process ID of the main process.

pid_s :: [Text] -> Pid

Safe version of pid

newtype Pid

Constructors

Pid ConfStatement

ssl_engine :: [Text] -> ConfStatement

Defines the name of the hardware SSL accelerator.

ssl_engine_s :: [Text] -> Ssl_Engine

Safe version of ssl_engine

newtype Ssl_Engine

Constructors

Ssl_Engine ConfStatement

thread_pool :: [Text] -> ConfStatement

Defines named thread pools used for multi-threaded reading and sending of files without blocking worker processes.

thread_pool_s :: [Text] -> Thread_Pool

Safe version of thread_pool

newtype Thread_Pool

Constructors

Thread_Pool ConfStatement

timer_resolution :: [Text] -> ConfStatement

Reduces timer resolution in worker processes, thus reducing the number of gettimeofday system calls made. By default, gettimeofday is called each time a kernel event is received. With reduced resolution, gettimeofday is only called once per specified interval.

timer_resolution_s :: [Text] -> Timer_Resolution

Safe version of timer_resolution

newtype Timer_Resolution

Constructors

Timer_Resolution ConfStatement

use :: [Text] -> ConfStatement

Specifies the connection processing method to use. There is normally no need to specify it explicitly, because nginx will by default use the most efficient method.

use_s :: [Text] -> Use

Safe version of use

newtype Use

Constructors

Use ConfStatement

user :: [Text] -> ConfStatement

Defines user and group credentials used by worker processes. If group is omitted, a group whose name equals that of user is used.

user_s :: [Text] -> User

Safe version of user

newtype User

Constructors

User ConfStatement

worker_aio_requests :: [Text] -> ConfStatement

When using with the connection processing method, sets the maximum number of outstanding asynchronous I/O operations for a single worker process.

worker_aio_requests_s :: [Text] -> Worker_Aio_Requests

Safe version of worker_aio_requests

newtype Worker_Aio_Requests

Constructors

Worker_Aio_Requests ConfStatement

worker_connections :: [Text] -> ConfStatement

Sets the maximum number of simultaneous connections that can be opened by a worker process.

worker_connections_s :: [Text] -> Worker_Connections

Safe version of worker_connections

newtype Worker_Connections

Constructors

Worker_Connections ConfStatement

worker_cpu_affinity :: [Text] -> ConfStatement

Binds worker processes to the sets of CPUs. Each CPU set is represented by a bitmask of allowed CPUs. There should be a separate set defined for each of the worker processes. By default, worker processes are not bound to any specific CPUs.

worker_cpu_affinity_s :: [Text] -> Worker_Cpu_Affinity

Safe version of worker_cpu_affinity

newtype Worker_Cpu_Affinity

Constructors

Worker_Cpu_Affinity ConfStatement

worker_priority :: [Text] -> ConfStatement

Defines the scheduling priority for worker processes like it is done by the nice command: a negative number means higher priority. Allowed range normally varies from -20 to 20.

worker_priority_s :: [Text] -> Worker_Priority

Safe version of worker_priority

newtype Worker_Priority

Constructors

Worker_Priority ConfStatement

worker_processes :: [Text] -> ConfStatement

Defines the number of worker processes.

worker_processes_s :: [Text] -> Worker_Processes

Safe version of worker_processes

newtype Worker_Processes

Constructors

Worker_Processes ConfStatement

worker_rlimit_core :: [Text] -> ConfStatement

Changes the limit on the largest size of a core file (RLIMIT_CORE) for worker processes. Used to increase the limit without restarting the main process.

worker_rlimit_core_s :: [Text] -> Worker_Rlimit_Core

Safe version of worker_rlimit_core

newtype Worker_Rlimit_Core

Constructors

Worker_Rlimit_Core ConfStatement

worker_rlimit_nofile :: [Text] -> ConfStatement

Changes the limit on the maximum number of open files (RLIMIT_NOFILE) for worker processes. Used to increase the limit without restarting the main process.

worker_rlimit_nofile_s :: [Text] -> Worker_Rlimit_Nofile

Safe version of worker_rlimit_nofile

newtype Worker_Rlimit_Nofile

Constructors

Worker_Rlimit_Nofile ConfStatement

working_directory :: [Text] -> ConfStatement

Defines the current working directory for a worker process. It is primarily used when writing a core-file, in which case a worker process should have write permission for the specified directory.

working_directory_s :: [Text] -> Working_Directory

Safe version of working_directory

newtype Working_Directory

Constructors

Working_Directory ConfStatement