Nginx_Ngx_Http_Addition_Module

Documentation

add_before_body :: [Text] -> ConfStatement

Adds the text returned as a result of processing a given subrequest before the response body. An empty string ("") as a parameter cancels addition inherited from the previous configuration level.

add_before_body_s :: [Text] -> Add_Before_Body

Safe version of add_before_body

newtype Add_Before_Body

Constructors

Add_Before_Body ConfStatement

add_after_body :: [Text] -> ConfStatement

Adds the text returned as a result of processing a given subrequest after the response body. An empty string ("") as a parameter cancels addition inherited from the previous configuration level.

add_after_body_s :: [Text] -> Add_After_Body

Safe version of add_after_body

newtype Add_After_Body

Constructors

Add_After_Body ConfStatement

addition_types :: [Text] -> ConfStatement

Allows adding text in responses with the specified MIME types, in addition to “text/html”. The special value “*” matches any MIME type (0.8.29).

addition_types_s :: [Text] -> Addition_Types

Safe version of addition_types

newtype Addition_Types

Constructors

Addition_Types ConfStatement