Nginx_Ngx_Http_Auth_Request_Module

Documentation

auth_request :: [Text] -> ConfStatement

Enables authorization based on the result of a subrequest and sets the URI to which the subrequest will be sent.

auth_request_s :: [Text] -> Auth_Request

Safe version of auth_request

newtype Auth_Request

Constructors

Auth_Request ConfStatement

auth_request_set :: [Text] -> ConfStatement

Sets the request variable to the given value after the authorization request completes. The value may contain variables from the authorization request, such as $upstream_http_*.

auth_request_set_s :: [Text] -> Auth_Request_Set

Safe version of auth_request_set

newtype Auth_Request_Set

Constructors

Auth_Request_Set ConfStatement