Nginx_Ngx_Http_Autoindex_Module

Documentation

autoindex :: [Text] -> ConfStatement

Enables or disables the directory listing output.

autoindex_s :: [Text] -> Autoindex

Safe version of autoindex

newtype Autoindex

Constructors

Autoindex ConfStatement

autoindex_exact_size :: [Text] -> ConfStatement

For the HTML format, specifies whether exact file sizes should be output in the directory listing, or rather rounded to kilobytes, megabytes, and gigabytes.

autoindex_exact_size_s :: [Text] -> Autoindex_Exact_Size

Safe version of autoindex_exact_size

newtype Autoindex_Exact_Size

Constructors

Autoindex_Exact_Size ConfStatement

autoindex_format :: [Text] -> ConfStatement

Sets the format of a directory listing.

autoindex_format_s :: [Text] -> Autoindex_Format

Safe version of autoindex_format

newtype Autoindex_Format

Constructors

Autoindex_Format ConfStatement

autoindex_localtime :: [Text] -> ConfStatement

For the HTML format, specifies whether times in the directory listing should be output in the local time zone or UTC.

autoindex_localtime_s :: [Text] -> Autoindex_Localtime

Safe version of autoindex_localtime

newtype Autoindex_Localtime

Constructors

Autoindex_Localtime ConfStatement