Nginx_Ngx_Http_Browser_Module

Documentation

ancient_browser :: [Text] -> ConfStatement

If any of the specified substrings is found in the User-Agent request header field, the browser will be considered ancient. The special string “netscape4” corresponds to the regular expression “^Mozilla/[1-4]”.

ancient_browser_s :: [Text] -> Ancient_Browser

Safe version of ancient_browser

newtype Ancient_Browser

Constructors

Ancient_Browser ConfStatement

ancient_browser_value :: [Text] -> ConfStatement

Sets a value for the $ancient_browser variables.

ancient_browser_value_s :: [Text] -> Ancient_Browser_Value

Safe version of ancient_browser_value

newtype Ancient_Browser_Value

Constructors

Ancient_Browser_Value ConfStatement

modern_browser :: [Text] -> ConfStatement

Specifies a version starting from which a browser is considered modern. A browser can be any one of the following: msie, gecko (browsers based on Mozilla), opera, safari, or konqueror.

modern_browser_s :: [Text] -> Modern_Browser

Safe version of modern_browser

newtype Modern_Browser

Constructors

Modern_Browser ConfStatement

modern_browser_value :: [Text] -> ConfStatement

Sets a value for the $modern_browser variables.

modern_browser_value_s :: [Text] -> Modern_Browser_Value

Safe version of modern_browser_value

newtype Modern_Browser_Value

Constructors

Modern_Browser_Value ConfStatement