Nginx_Ngx_Http_Charset_Module

Documentation

charset :: [Text] -> ConfStatement

Adds the specified charset to the Content-Type response header field. If this charset is different from the charset specified in the directive, a conversion is performed.

charset_s :: [Text] -> Charset

Safe version of charset

newtype Charset

Constructors

Charset ConfStatement

charset_map :: [Text] -> [ConfStatement] -> ConfStatement

Describes the conversion table from one charset to another. A reverse conversion table is built using the same data. Character codes are given in hexadecimal. Missing characters in the range 80-FF are replaced with “?”. When converting from UTF-8, characters missing in a one-byte charset are replaced with “&#XXXX;”.

charset_map_s :: (HasContextCharset_Map f, ToConfStatement f) => [Text] -> [f] -> ConfStatement

Safe version of charset_map

charset_types :: [Text] -> ConfStatement

Enables module processing in responses with the specified MIME types in addition to “text/html”. The special value “*” matches any MIME type (0.8.29).

charset_types_s :: [Text] -> Charset_Types

Safe version of charset_types

newtype Charset_Types

Constructors

Charset_Types ConfStatement

override_charset :: [Text] -> ConfStatement

Determines whether a conversion should be performed for answers received from a proxied or a FastCGIuwsgiSCGI server when the answers already carry a charset in the Content-Type response header field. If conversion is enabled, a charset specified in the received response is used as a source charset. It should be noted that if a response is received in a subrequest then the conversion from the response charset to the main request charset is always performed, regardless of the override_charset directive setting.

override_charset_s :: [Text] -> Override_Charset

Safe version of override_charset

newtype Override_Charset

Constructors

Override_Charset ConfStatement

source_charset :: [Text] -> ConfStatement

Defines the source charset of a response. If this charset is different from the charset specified in the directive, a conversion is performed.

source_charset_s :: [Text] -> Source_Charset

Safe version of source_charset

newtype Source_Charset

Constructors

Source_Charset ConfStatement