Nginx_Ngx_Http_Core_Module

Documentation

aio :: [Text] -> ConfStatement

Enables or disables the use of asynchronous file IO (AIO) on FreeBSD and Linux: location video/ { aio on; output_buffers 1 64k; }

aio_s :: [Text] -> Aio

Safe version of aio

newtype Aio

Constructors

Aio ConfStatement

aio_write :: [Text] -> ConfStatement

If is enabled, specifies whether it is used for writing files. Currently, this only works when using aio threads and is limited to writing temporary files with data received from proxied servers.

aio_write_s :: [Text] -> Aio_Write

Safe version of aio_write

newtype Aio_Write

Constructors

Aio_Write ConfStatement

alias :: [Text] -> ConfStatement

Defines a replacement for the specified location. For example, with the following configuration location i { alias dataw3images; } on request of “itop.gif”, the file dataw3imagestop.gif will be sent.

alias_s :: [Text] -> Alias

Safe version of alias

newtype Alias

Constructors

Alias ConfStatement

chunked_transfer_encoding :: [Text] -> ConfStatement

Allows disabling chunked transfer encoding in HTTP/1.1. It may come in handy when using a software failing to support chunked encoding despite the standard’s requirement.

chunked_transfer_encoding_s :: [Text] -> Chunked_Transfer_Encoding

Safe version of chunked_transfer_encoding

newtype Chunked_Transfer_Encoding

Constructors

Chunked_Transfer_Encoding ConfStatement

client_body_buffer_size :: [Text] -> ConfStatement

Sets buffer size for reading client request body. In case the request body is larger than the buffer, the whole body or only its part is written to a temporary file. By default, buffer size is equal to two memory pages. This is 8K on x86, other 32-bit platforms, and x86-64. It is usually 16K on other 64-bit platforms.

client_body_buffer_size_s :: [Text] -> Client_Body_Buffer_Size

Safe version of client_body_buffer_size

newtype Client_Body_Buffer_Size

Constructors

Client_Body_Buffer_Size ConfStatement

client_body_in_file_only :: [Text] -> ConfStatement

Determines whether nginx should save the entire client request body into a file. This directive can be used during debugging, or when using the $request_body_file variable, or the $r->request_body_file method of the module ngx_http_perl_module.

client_body_in_file_only_s :: [Text] -> Client_Body_In_File_Only

Safe version of client_body_in_file_only

newtype Client_Body_In_File_Only

Constructors

Client_Body_In_File_Only ConfStatement

client_body_in_single_buffer :: [Text] -> ConfStatement

Determines whether nginx should save the entire client request body in a single buffer. The directive is recommended when using the $request_body variable, to save the number of copy operations involved.

client_body_in_single_buffer_s :: [Text] -> Client_Body_In_Single_Buffer

Safe version of client_body_in_single_buffer

newtype Client_Body_In_Single_Buffer

Constructors

Client_Body_In_Single_Buffer ConfStatement

client_body_temp_path :: [Text] -> ConfStatement

Defines a directory for storing temporary files holding client request bodies. Up to three-level subdirectory hierarchy can be used under the specified directory. For example, in the following configuration client_body_temp_path spoolnginxclient_temp 1 2; a path to a temporary file might look like this: spoolnginxclient_temp745/00000123457

client_body_temp_path_s :: [Text] -> Client_Body_Temp_Path

Safe version of client_body_temp_path

newtype Client_Body_Temp_Path

Constructors

Client_Body_Temp_Path ConfStatement

client_body_timeout :: [Text] -> ConfStatement

Defines a timeout for reading client request body. The timeout is set only for a period between two successive read operations, not for the transmission of the whole request body. If a client does not transmit anything within this time, the error is returned to the client.

client_body_timeout_s :: [Text] -> Client_Body_Timeout

Safe version of client_body_timeout

newtype Client_Body_Timeout

Constructors

Client_Body_Timeout ConfStatement

client_header_buffer_size :: [Text] -> ConfStatement

Sets buffer size for reading client request header. For most requests, a buffer of 1K bytes is enough. However, if a request includes long cookies, or comes from a WAP client, it may not fit into 1K. If a request line or a request header field does not fit into this buffer then larger buffers, configured by the directive, are allocated.

client_header_buffer_size_s :: [Text] -> Client_Header_Buffer_Size

Safe version of client_header_buffer_size

newtype Client_Header_Buffer_Size

Constructors

Client_Header_Buffer_Size ConfStatement

client_header_timeout :: [Text] -> ConfStatement

Defines a timeout for reading client request header. If a client does not transmit the entire header within this time, the error is returned to the client.

client_header_timeout_s :: [Text] -> Client_Header_Timeout

Safe version of client_header_timeout

newtype Client_Header_Timeout

Constructors

Client_Header_Timeout ConfStatement

client_max_body_size :: [Text] -> ConfStatement

Sets the maximum allowed size of the client request body, specified in the Content-Length request header field. If the size in a request exceeds the configured value, the error is returned to the client. Please be aware that browsers cannot correctly display this error. Setting size to 0 disables checking of client request body size.

client_max_body_size_s :: [Text] -> Client_Max_Body_Size

Safe version of client_max_body_size

newtype Client_Max_Body_Size

Constructors

Client_Max_Body_Size ConfStatement

connection_pool_size :: [Text] -> ConfStatement

Allows accurate tuning of per-connection memory allocations. This directive has minimal impact on performance and should not generally be used. By default, the size is equal to 256 bytes on 32-bit platforms and 512 bytes on 64-bit platforms. Prior to version 1.9.8, the default value was 256 on all platforms.

connection_pool_size_s :: [Text] -> Connection_Pool_Size

Safe version of connection_pool_size

newtype Connection_Pool_Size

Constructors

Connection_Pool_Size ConfStatement

default_type :: [Text] -> ConfStatement

Defines the default MIME type of a response. Mapping of file name extensions to MIME types can be set with the directive.

default_type_s :: [Text] -> Default_Type

Safe version of default_type

newtype Default_Type

Constructors

Default_Type ConfStatement

directio :: [Text] -> ConfStatement

Enables the use of the O_DIRECT flag (FreeBSD, Linux), the F_NOCACHE flag (Mac OS X), or the directio function (Solaris), when reading files that are larger than or equal to the specified size. The directive automatically disables (0.7.15) the use of for a given request. It can be useful for serving large files: directio 4m; or when using on Linux.

directio_s :: [Text] -> Directio

Safe version of directio

newtype Directio

Constructors

Directio ConfStatement

directio_alignment :: [Text] -> ConfStatement

Sets the alignment for . In most cases, a 512-byte alignment is enough. However, when using XFS under Linux, it needs to be increased to 4K.

directio_alignment_s :: [Text] -> Directio_Alignment

Safe version of directio_alignment

newtype Directio_Alignment

Constructors

Directio_Alignment ConfStatement

disable_symlinks :: [Text] -> ConfStatement

Determines how symbolic links should be treated when opening files: off Symbolic links in the pathname are allowed and not checked. This is the default behavior. on If any component of the pathname is a symbolic link, access to a file is denied. if_not_owner Access to a file is denied if any component of the pathname is a symbolic link, and the link and object that the link points to have different owners. from=part When checking symbolic links (parameters on and if_not_owner), all components of the pathname are normally checked. Checking of symbolic links in the initial part of the pathname may be avoided by specifying additionally the from=part parameter. In this case, symbolic links are checked only from the pathname component that follows the specified initial part. If the value is not an initial part of the pathname checked, the whole pathname is checked as if this parameter was not specified at all. If the value matches the whole file name, symbolic links are not checked. The parameter value can contain variables.

disable_symlinks_s :: [Text] -> Disable_Symlinks

Safe version of disable_symlinks

newtype Disable_Symlinks

Constructors

Disable_Symlinks ConfStatement

error_page :: [Text] -> ConfStatement

Defines the URI that will be shown for the specified errors. error_page directives are inherited from the previous level only if there are no error_page directives defined on the current level. A uri value can contain variables.

error_page_s :: [Text] -> Error_Page

Safe version of error_page

newtype Error_Page

Constructors

Error_Page ConfStatement

etag :: [Text] -> ConfStatement

Enables or disables automatic generation of the ETag response header field for static resources.

etag_s :: [Text] -> Etag

Safe version of etag

newtype Etag

Constructors

Etag ConfStatement

http :: [Text] -> [ConfStatement] -> ConfStatement

Provides the configuration file context in which the HTTP server directives are specified.

http_s :: (HasContextHttp f, ToConfStatement f) => [Text] -> [f] -> ConfStatement

Safe version of http

if_modified_since :: [Text] -> ConfStatement

Specifies how to compare modification time of a response with the time in the If-Modified-Since request header field: off the If-Modified-Since request header field is ignored (0.7.34); exact exact match; before modification time of a response is less than or equal to the time in the If-Modified-Since request header field.

if_modified_since_s :: [Text] -> If_Modified_Since

Safe version of if_modified_since

newtype If_Modified_Since

Constructors

If_Modified_Since ConfStatement

ignore_invalid_headers :: [Text] -> ConfStatement

Controls whether header fields with invalid names should be ignored. Valid names are composed of English letters, digits, hyphens, and possibly underscores (as controlled by the directive).

ignore_invalid_headers_s :: [Text] -> Ignore_Invalid_Headers

Safe version of ignore_invalid_headers

newtype Ignore_Invalid_Headers

Constructors

Ignore_Invalid_Headers ConfStatement

internal :: [Text] -> ConfStatement

Specifies that a given location can only be used for internal requests. For external requests, the client error is returned. Internal requests are the following: requests redirected by the , , , and directives; requests redirected by the X-Accel-Redirect response header field from an upstream server; subrequests formed by the “include virtual” command of the ngx_http_ssi_module module and by the ngx_http_addition_module module directives; requests changed by the directive.

internal_s :: [Text] -> Internal

Safe version of internal

newtype Internal

Constructors

Internal ConfStatement

keepalive_disable :: [Text] -> ConfStatement

Disables keep-alive connections with misbehaving browsers. The browser parameters specify which browsers will be affected. The value msie6 disables keep-alive connections with old versions of MSIE, once a POST request is received. The value safari disables keep-alive connections with Safari and Safari-like browsers on Mac OS X and Mac OS X-like operating systems. The value none enables keep-alive connections with all browsers. Prior to version 1.1.18, the value safari matched all Safari and Safari-like browsers on all operating systems, and keep-alive connections with them were disabled by default.

keepalive_disable_s :: [Text] -> Keepalive_Disable

Safe version of keepalive_disable

newtype Keepalive_Disable

Constructors

Keepalive_Disable ConfStatement

keepalive_requests :: [Text] -> ConfStatement

Sets the maximum number of requests that can be served through one keep-alive connection. After the maximum number of requests are made, the connection is closed.

keepalive_requests_s :: [Text] -> Keepalive_Requests

Safe version of keepalive_requests

newtype Keepalive_Requests

Constructors

Keepalive_Requests ConfStatement

keepalive_timeout :: [Text] -> ConfStatement

The first parameter sets a timeout during which a keep-alive client connection will stay open on the server side. The zero value disables keep-alive client connections. The optional second parameter sets a value in the Keep-Alive: timeout=time response header field. Two parameters may differ.

keepalive_timeout_s :: [Text] -> Keepalive_Timeout

Safe version of keepalive_timeout

newtype Keepalive_Timeout

Constructors

Keepalive_Timeout ConfStatement

large_client_header_buffers :: [Text] -> ConfStatement

Sets the maximum number and size of buffers used for reading large client request header. A request line cannot exceed the size of one buffer, or the error is returned to the client. A request header field cannot exceed the size of one buffer as well, or the error is returned to the client. Buffers are allocated only on demand. By default, the buffer size is equal to 8K bytes. If after the end of request processing a connection is transitioned into the keep-alive state, these buffers are released.

large_client_header_buffers_s :: [Text] -> Large_Client_Header_Buffers

Safe version of large_client_header_buffers

newtype Large_Client_Header_Buffers

Constructors

Large_Client_Header_Buffers ConfStatement

limit_except :: [Text] -> [ConfStatement] -> ConfStatement

Limits allowed HTTP methods inside a location. The method parameter can be one of the following: GET, HEAD, POST, PUT, DELETE, MKCOL, COPY, MOVE, OPTIONS, PROPFIND, PROPPATCH, LOCK, UNLOCK, or PATCH. Allowing the GET method makes the HEAD method also allowed. Access to other methods can be limited using the ngx_http_access_module and ngx_http_auth_basic_module modules directives: limit_except GET { allow 192.168.1.0/32; deny all; } Please note that this will limit access to all methods except GET and HEAD.

limit_except_s :: (HasContextLimit_Except f, ToConfStatement f) => [Text] -> [f] -> ConfStatement

Safe version of limit_except

limit_rate :: [Text] -> ConfStatement

Limits the rate of response transmission to a client. The rate is specified in bytes per second. The zero value disables rate limiting. The limit is set per a request, and so if a client simultaneously opens two connections, the overall rate will be twice as much as the specified limit.

limit_rate_s :: [Text] -> Limit_Rate

Safe version of limit_rate

newtype Limit_Rate

Constructors

Limit_Rate ConfStatement

limit_rate_after :: [Text] -> ConfStatement

Sets the initial amount after which the further transmission of a response to a client will be rate limited.

limit_rate_after_s :: [Text] -> Limit_Rate_After

Safe version of limit_rate_after

newtype Limit_Rate_After

Constructors

Limit_Rate_After ConfStatement

lingering_close :: [Text] -> ConfStatement

Controls how nginx closes client connections.

lingering_close_s :: [Text] -> Lingering_Close

Safe version of lingering_close

newtype Lingering_Close

Constructors

Lingering_Close ConfStatement

lingering_time :: [Text] -> ConfStatement

When is in effect, this directive specifies the maximum time during which nginx will process (read and ignore) additional data coming from a client. After that, the connection will be closed, even if there will be more data.

lingering_time_s :: [Text] -> Lingering_Time

Safe version of lingering_time

newtype Lingering_Time

Constructors

Lingering_Time ConfStatement

lingering_timeout :: [Text] -> ConfStatement

When is in effect, this directive specifies the maximum waiting time for more client data to arrive. If data are not received during this time, the connection is closed. Otherwise, the data are read and ignored, and nginx starts waiting for more data again. The “wait-read-ignore” cycle is repeated, but no longer than specified by the directive.

lingering_timeout_s :: [Text] -> Lingering_Timeout

Safe version of lingering_timeout

newtype Lingering_Timeout

Constructors

Lingering_Timeout ConfStatement

listen :: [Text] -> ConfStatement

Sets the address and port for IP, or the path for a UNIX-domain socket on which the server will accept requests. Both address and port, or only address or only port can be specified. An address may also be a hostname, for example: listen 127.0.0.1:8000; listen 127.0.0.1; listen 8000; listen *:8000; listen localhost:8000; IPv6 addresses (0.7.36) are specified in square brackets: listen [::]:8000; listen [::1]; UNIX-domain sockets (0.8.21) are specified with the “unix:” prefix: listen unix:varrun/nginx.sock;

listen_s :: [Text] -> Listen

Safe version of listen

newtype Listen

Constructors

Listen ConfStatement

location :: [Text] -> [ConfStatement] -> ConfStatement

Sets configuration depending on a request URI.

location_s :: (HasContextLocation f, ToConfStatement f) => [Text] -> [f] -> ConfStatement

Safe version of location

log_not_found :: [Text] -> ConfStatement

Enables or disables logging of errors about not found files into .

log_not_found_s :: [Text] -> Log_Not_Found

Safe version of log_not_found

newtype Log_Not_Found

Constructors

Log_Not_Found ConfStatement

log_subrequest :: [Text] -> ConfStatement

Enables or disables logging of subrequests into .

log_subrequest_s :: [Text] -> Log_Subrequest

Safe version of log_subrequest

newtype Log_Subrequest

Constructors

Log_Subrequest ConfStatement

max_ranges :: [Text] -> ConfStatement

Limits the maximum allowed number of ranges in byte-range requests. Requests that exceed the limit are processed as if there were no byte ranges specified. By default, the number of ranges is not limited. The zero value disables the byte-range support completely.

max_ranges_s :: [Text] -> Max_Ranges

Safe version of max_ranges

newtype Max_Ranges

Constructors

Max_Ranges ConfStatement

merge_slashes :: [Text] -> ConfStatement

Enables or disables compression of two or more adjacent slashes in a URI into a single slash.

merge_slashes_s :: [Text] -> Merge_Slashes

Safe version of merge_slashes

newtype Merge_Slashes

Constructors

Merge_Slashes ConfStatement

msie_padding :: [Text] -> ConfStatement

Enables or disables adding comments to responses for MSIE clients with status greater than 400 to increase the response size to 512 bytes.

msie_padding_s :: [Text] -> Msie_Padding

Safe version of msie_padding

newtype Msie_Padding

Constructors

Msie_Padding ConfStatement

msie_refresh :: [Text] -> ConfStatement

Enables or disables issuing refreshes instead of redirects for MSIE clients.

msie_refresh_s :: [Text] -> Msie_Refresh

Safe version of msie_refresh

newtype Msie_Refresh

Constructors

Msie_Refresh ConfStatement

open_file_cache :: [Text] -> ConfStatement

Configures a cache that can store: open file descriptors, their sizes and modification times; information on existence of directories; file lookup errors, such as “file not found”, “no read permission”, and so on. Caching of errors should be enabled separately by the directive.

open_file_cache_s :: [Text] -> Open_File_Cache

Safe version of open_file_cache

newtype Open_File_Cache

Constructors

Open_File_Cache ConfStatement

open_file_cache_errors :: [Text] -> ConfStatement

Enables or disables caching of file lookup errors by .

open_file_cache_errors_s :: [Text] -> Open_File_Cache_Errors

Safe version of open_file_cache_errors

newtype Open_File_Cache_Errors

Constructors

Open_File_Cache_Errors ConfStatement

open_file_cache_min_uses :: [Text] -> ConfStatement

Sets the minimum number of file accesses during the period configured by the inactive parameter of the directive, required for a file descriptor to remain open in the cache.

open_file_cache_min_uses_s :: [Text] -> Open_File_Cache_Min_Uses

Safe version of open_file_cache_min_uses

newtype Open_File_Cache_Min_Uses

Constructors

Open_File_Cache_Min_Uses ConfStatement

open_file_cache_valid :: [Text] -> ConfStatement

Sets a time after which elements should be validated.

open_file_cache_valid_s :: [Text] -> Open_File_Cache_Valid

Safe version of open_file_cache_valid

newtype Open_File_Cache_Valid

Constructors

Open_File_Cache_Valid ConfStatement

output_buffers :: [Text] -> ConfStatement

Sets the number and size of the buffers used for reading a response from a disk. Prior to version 1.9.5, the default value was 1 32k.

output_buffers_s :: [Text] -> Output_Buffers

Safe version of output_buffers

newtype Output_Buffers

Constructors

Output_Buffers ConfStatement

port_in_redirect :: [Text] -> ConfStatement

Enables or disables specifying the port in redirects issued by nginx.

port_in_redirect_s :: [Text] -> Port_In_Redirect

Safe version of port_in_redirect

newtype Port_In_Redirect

Constructors

Port_In_Redirect ConfStatement

postpone_output :: [Text] -> ConfStatement

If possible, the transmission of client data will be postponed until nginx has at least size bytes of data to send. The zero value disables postponing data transmission.

postpone_output_s :: [Text] -> Postpone_Output

Safe version of postpone_output

newtype Postpone_Output

Constructors

Postpone_Output ConfStatement

read_ahead :: [Text] -> ConfStatement

Sets the amount of pre-reading for the kernel when working with file.

read_ahead_s :: [Text] -> Read_Ahead

Safe version of read_ahead

newtype Read_Ahead

Constructors

Read_Ahead ConfStatement

recursive_error_pages :: [Text] -> ConfStatement

Enables or disables doing several redirects using the directive. The number of such redirects is limited.

recursive_error_pages_s :: [Text] -> Recursive_Error_Pages

Safe version of recursive_error_pages

newtype Recursive_Error_Pages

Constructors

Recursive_Error_Pages ConfStatement

request_pool_size :: [Text] -> ConfStatement

Allows accurate tuning of per-request memory allocations. This directive has minimal impact on performance and should not generally be used.

request_pool_size_s :: [Text] -> Request_Pool_Size

Safe version of request_pool_size

newtype Request_Pool_Size

Constructors

Request_Pool_Size ConfStatement

reset_timedout_connection :: [Text] -> ConfStatement

Enables or disables resetting timed out connections. The reset is performed as follows. Before closing a socket, the SO_LINGER option is set on it with a timeout value of 0. When the socket is closed, TCP RST is sent to the client, and all memory occupied by this socket is released. This helps avoid keeping an already closed socket with filled buffers in a FIN_WAIT1 state for a long time.

reset_timedout_connection_s :: [Text] -> Reset_Timedout_Connection

Safe version of reset_timedout_connection

newtype Reset_Timedout_Connection

Constructors

Reset_Timedout_Connection ConfStatement

resolver :: [Text] -> ConfStatement

Configures name servers used to resolve names of upstream servers into addresses, for example: resolver 127.0.0.1 [::1]:5353; An address can be specified as a domain name or IP address, and an optional port (1.3.1, 1.2.2). If port is not specified, the port 53 is used. Name servers are queried in a round-robin fashion. Before version 1.1.7, only a single name server could be configured. Specifying name servers using IPv6 addresses is supported starting from versions 1.3.1 and 1.2.2. By default, nginx will look up both IPv4 and IPv6 addresses while resolving. If looking up of IPv6 addresses is not desired, the ipv6=off parameter can be specified. Resolving of names into IPv6 addresses is supported starting from version 1.5.8. By default, nginx caches answers using the TTL value of a response. An optional valid parameter allows overriding it: resolver 127.0.0.1 [::1]:5353 valid=30s; Before version 1.1.9, tuning of caching time was not possible, and nginx always cached answers for the duration of 5 minutes.

resolver_s :: [Text] -> Resolver

Safe version of resolver

newtype Resolver

Constructors

Resolver ConfStatement

resolver_timeout :: [Text] -> ConfStatement

Sets a timeout for name resolution, for example: resolver_timeout 5s;

resolver_timeout_s :: [Text] -> Resolver_Timeout

Safe version of resolver_timeout

newtype Resolver_Timeout

Constructors

Resolver_Timeout ConfStatement

root :: [Text] -> ConfStatement

Sets the root directory for requests. For example, with the following configuration location i { root dataw3; } The dataw3itop.gif file will be sent in response to the “itop.gif” request.

root_s :: [Text] -> Root

Safe version of root

newtype Root

Constructors

Root ConfStatement

satisfy :: [Text] -> ConfStatement

Allows access if all (all) or at least one (any) of the ngx_http_access_module, ngx_http_auth_basic_module or ngx_http_auth_request_module modules allow access.

satisfy_s :: [Text] -> Satisfy

Safe version of satisfy

newtype Satisfy

Constructors

Satisfy ConfStatement

send_lowat :: [Text] -> ConfStatement

If the directive is set to a non-zero value, nginx will try to minimize the number of send operations on client sockets by using either NOTE_LOWAT flag of the method or the SO_SNDLOWAT socket option. In both cases the specified size is used.

send_lowat_s :: [Text] -> Send_Lowat

Safe version of send_lowat

newtype Send_Lowat

Constructors

Send_Lowat ConfStatement

send_timeout :: [Text] -> ConfStatement

Sets a timeout for transmitting a response to the client. The timeout is set only between two successive write operations, not for the transmission of the whole response. If the client does not receive anything within this time, the connection is closed.

send_timeout_s :: [Text] -> Send_Timeout

Safe version of send_timeout

newtype Send_Timeout

Constructors

Send_Timeout ConfStatement

sendfile :: [Text] -> ConfStatement

Enables or disables the use of sendfile.

sendfile_s :: [Text] -> Sendfile

Safe version of sendfile

newtype Sendfile

Constructors

Sendfile ConfStatement

sendfile_max_chunk :: [Text] -> ConfStatement

When set to a non-zero value, limits the amount of data that can be transferred in a single sendfile call. Without the limit, one fast connection may seize the worker process entirely.

sendfile_max_chunk_s :: [Text] -> Sendfile_Max_Chunk

Safe version of sendfile_max_chunk

newtype Sendfile_Max_Chunk

Constructors

Sendfile_Max_Chunk ConfStatement

server :: [Text] -> [ConfStatement] -> ConfStatement

Sets configuration for a virtual server. There is no clear separation between IP-based (based on the IP address) and name-based (based on the Host request header field) virtual servers. Instead, the directives describe all addresses and ports that should accept connections for the server, and the directive lists all server names. Example configurations are provided in the “” document.

server_s :: (HasContextServer f, ToConfStatement f) => [Text] -> [f] -> ConfStatement

Safe version of server

server_name :: [Text] -> ConfStatement

Sets names of a virtual server, for example: server { server_name example.com www.example.com; }

server_name_s :: [Text] -> Server_Name

Safe version of server_name

newtype Server_Name

Constructors

Server_Name ConfStatement

server_name_in_redirect :: [Text] -> ConfStatement

Enables or disables the use of the primary server name, specified by the directive, in redirects issued by nginx. When the use of the primary server name is disabled, the name from the Host request header field is used. If this field is not present, the IP address of the server is used.

server_name_in_redirect_s :: [Text] -> Server_Name_In_Redirect

Safe version of server_name_in_redirect

newtype Server_Name_In_Redirect

Constructors

Server_Name_In_Redirect ConfStatement

server_names_hash_bucket_size :: [Text] -> ConfStatement

Sets the bucket size for the server names hash tables. The default value depends on the size of the processor’s cache line. The details of setting up hash tables are provided in a separate document.

server_names_hash_bucket_size_s :: [Text] -> Server_Names_Hash_Bucket_Size

Safe version of server_names_hash_bucket_size

newtype Server_Names_Hash_Bucket_Size

Constructors

Server_Names_Hash_Bucket_Size ConfStatement

server_names_hash_max_size :: [Text] -> ConfStatement

Sets the maximum size of the server names hash tables. The details of setting up hash tables are provided in a separate document.

server_names_hash_max_size_s :: [Text] -> Server_Names_Hash_Max_Size

Safe version of server_names_hash_max_size

newtype Server_Names_Hash_Max_Size

Constructors

Server_Names_Hash_Max_Size ConfStatement

server_tokens :: [Text] -> ConfStatement

Enables or disables emitting nginx version in error messages and in the Server response header field.

server_tokens_s :: [Text] -> Server_Tokens

Safe version of server_tokens

newtype Server_Tokens

Constructors

Server_Tokens ConfStatement

tcp_nodelay :: [Text] -> ConfStatement

Enables or disables the use of the TCP_NODELAY option. The option is enabled only when a connection is transitioned into the keep-alive state.

tcp_nodelay_s :: [Text] -> Tcp_Nodelay

Safe version of tcp_nodelay

newtype Tcp_Nodelay

Constructors

Tcp_Nodelay ConfStatement

tcp_nopush :: [Text] -> ConfStatement

Enables or disables the use of the TCP_NOPUSH socket option on FreeBSD or the TCP_CORK socket option on Linux. The options are enabled only when is used. Enabling the option allows sending the response header and the beginning of a file in one packet, on Linux and FreeBSD 4.*; sending a file in full packets.

tcp_nopush_s :: [Text] -> Tcp_Nopush

Safe version of tcp_nopush

newtype Tcp_Nopush

Constructors

Tcp_Nopush ConfStatement

try_files :: [Text] -> ConfStatement

Checks the existence of files in the specified order and uses the first found file for request processing; the processing is performed in the current context. The path to a file is constructed from the file parameter according to the and directives. It is possible to check directory’s existence by specifying a slash at the end of a name, e.g. “$uri”. If none of the files were found, an internal redirect to the uri specified in the last parameter is made. For example: location images { try_files $uri imagesdefault.gif; } location = imagesdefault.gif { expires 30s; } The last parameter can also point to a named location, as shown in examples below. Starting from version 0.7.51, the last parameter can also be a code: location { try_files $uri $uri/index.html $uri.html =404; }

try_files_s :: [Text] -> Try_Files

Safe version of try_files

newtype Try_Files

Constructors

Try_Files ConfStatement

types :: [Text] -> [ConfStatement] -> ConfStatement

Maps file name extensions to MIME types of responses. Extensions are case-insensitive. Several extensions can be mapped to one type, for example: types { applicationoctet-stream bin exe dll; applicationoctet-stream deb; application/octet-stream dmg; }

types_s :: (HasContextTypes f, ToConfStatement f) => [Text] -> [f] -> ConfStatement

Safe version of types

types_hash_bucket_size :: [Text] -> ConfStatement

Sets the bucket size for the types hash tables. The details of setting up hash tables are provided in a separate document. Prior to version 1.5.13, the default value depended on the size of the processor’s cache line.

types_hash_bucket_size_s :: [Text] -> Types_Hash_Bucket_Size

Safe version of types_hash_bucket_size

newtype Types_Hash_Bucket_Size

Constructors

Types_Hash_Bucket_Size ConfStatement

types_hash_max_size :: [Text] -> ConfStatement

Sets the maximum size of the types hash tables. The details of setting up hash tables are provided in a separate document.

types_hash_max_size_s :: [Text] -> Types_Hash_Max_Size

Safe version of types_hash_max_size

newtype Types_Hash_Max_Size

Constructors

Types_Hash_Max_Size ConfStatement

underscores_in_headers :: [Text] -> ConfStatement

Enables or disables the use of underscores in client request header fields. When the use of underscores is disabled, request header fields whose names contain underscores are marked as invalid and become subject to the directive.

underscores_in_headers_s :: [Text] -> Underscores_In_Headers

Safe version of underscores_in_headers

newtype Underscores_In_Headers

Constructors

Underscores_In_Headers ConfStatement

variables_hash_bucket_size :: [Text] -> ConfStatement

Sets the bucket size for the variables hash table. The details of setting up hash tables are provided in a separate document.

variables_hash_bucket_size_s :: [Text] -> Variables_Hash_Bucket_Size

Safe version of variables_hash_bucket_size

newtype Variables_Hash_Bucket_Size

Constructors

Variables_Hash_Bucket_Size ConfStatement

variables_hash_max_size :: [Text] -> ConfStatement

Sets the maximum size of the variables hash table. The details of setting up hash tables are provided in a separate document. Prior to version 1.5.13, the default value was 512.

variables_hash_max_size_s :: [Text] -> Variables_Hash_Max_Size

Safe version of variables_hash_max_size

newtype Variables_Hash_Max_Size

Constructors

Variables_Hash_Max_Size ConfStatement