Nginx_Ngx_Http_Dav_Module

Documentation

dav_access :: [Text] -> ConfStatement

Sets access permissions for newly created files and directories, e.g.: dav_access user:rw group:rw all:r;

dav_access_s :: [Text] -> Dav_Access

Safe version of dav_access

newtype Dav_Access

Constructors

Dav_Access ConfStatement

dav_methods :: [Text] -> ConfStatement

Allows the specified HTTP and WebDAV methods. The parameter off denies all methods processed by this module. The following methods are supported: PUT, DELETE, MKCOL, COPY, and MOVE.

dav_methods_s :: [Text] -> Dav_Methods

Safe version of dav_methods

newtype Dav_Methods

Constructors

Dav_Methods ConfStatement

create_full_put_path :: [Text] -> ConfStatement

The WebDAV specification only allows creating files in already existing directories. This directive allows creating all needed intermediate directories.

create_full_put_path_s :: [Text] -> Create_Full_Put_Path

Safe version of create_full_put_path

newtype Create_Full_Put_Path

Constructors

Create_Full_Put_Path ConfStatement

min_delete_depth :: [Text] -> ConfStatement

Allows the DELETE method to remove files provided that the number of elements in a request path is not less than the specified number. For example, the directive min_delete_depth 4; allows removing files on requests users0000name users0000namepic.jpg users0000page.html and denies the removal of users0000

min_delete_depth_s :: [Text] -> Min_Delete_Depth

Safe version of min_delete_depth

newtype Min_Delete_Depth

Constructors

Min_Delete_Depth ConfStatement