Nginx_Ngx_Http_Fastcgi_Module

Documentation

fastcgi_bind :: [Text] -> ConfStatement

Makes outgoing connections to a FastCGI server originate from the specified local IP address with an optional port (1.11.2). Parameter value can contain variables (1.3.12). The special value off (1.3.12) cancels the effect of the fastcgi_bind directive inherited from the previous configuration level, which allows the system to auto-assign the local IP address and port.

fastcgi_bind_s :: [Text] -> Fastcgi_Bind

Safe version of fastcgi_bind

newtype Fastcgi_Bind

Constructors

Fastcgi_Bind ConfStatement

fastcgi_buffer_size :: [Text] -> ConfStatement

Sets the size of the buffer used for reading the first part of the response received from the FastCGI server. This part usually contains a small response header. By default, the buffer size is equal to one memory page. This is either 4K or 8K, depending on a platform. It can be made smaller, however.

fastcgi_buffer_size_s :: [Text] -> Fastcgi_Buffer_Size

Safe version of fastcgi_buffer_size

newtype Fastcgi_Buffer_Size

Constructors

Fastcgi_Buffer_Size ConfStatement

fastcgi_buffering :: [Text] -> ConfStatement

Enables or disables buffering of responses from the FastCGI server.

fastcgi_buffering_s :: [Text] -> Fastcgi_Buffering

Safe version of fastcgi_buffering

newtype Fastcgi_Buffering

Constructors

Fastcgi_Buffering ConfStatement

fastcgi_buffers :: [Text] -> ConfStatement

Sets the number and size of the buffers used for reading a response from the FastCGI server, for a single connection. By default, the buffer size is equal to one memory page. This is either 4K or 8K, depending on a platform.

fastcgi_buffers_s :: [Text] -> Fastcgi_Buffers

Safe version of fastcgi_buffers

newtype Fastcgi_Buffers

Constructors

Fastcgi_Buffers ConfStatement

fastcgi_busy_buffers_size :: [Text] -> ConfStatement

When buffering of responses from the FastCGI server is enabled, limits the total size of buffers that can be busy sending a response to the client while the response is not yet fully read. In the meantime, the rest of the buffers can be used for reading the response and, if needed, buffering part of the response to a temporary file. By default, size is limited by the size of two buffers set by the and directives.

fastcgi_busy_buffers_size_s :: [Text] -> Fastcgi_Busy_Buffers_Size

Safe version of fastcgi_busy_buffers_size

newtype Fastcgi_Busy_Buffers_Size

Constructors

Fastcgi_Busy_Buffers_Size ConfStatement

fastcgi_cache :: [Text] -> ConfStatement

Defines a shared memory zone used for caching. The same zone can be used in several places. Parameter value can contain variables (1.7.9). The off parameter disables caching inherited from the previous configuration level.

fastcgi_cache_s :: [Text] -> Fastcgi_Cache

Safe version of fastcgi_cache

newtype Fastcgi_Cache

Constructors

Fastcgi_Cache ConfStatement

fastcgi_cache_bypass :: [Text] -> ConfStatement

Defines conditions under which the response will not be taken from a cache. If at least one value of the string parameters is not empty and is not equal to “0” then the response will not be taken from the cache: fastcgi_cache_bypass $cookie_nocache $arg_nocache$arg_comment; fastcgi_cache_bypass $http_pragma $http_authorization; Can be used along with the directive.

fastcgi_cache_bypass_s :: [Text] -> Fastcgi_Cache_Bypass

Safe version of fastcgi_cache_bypass

newtype Fastcgi_Cache_Bypass

Constructors

Fastcgi_Cache_Bypass ConfStatement

fastcgi_cache_key :: [Text] -> ConfStatement

Defines a key for caching, for example fastcgi_cache_key localhost:9000$request_uri;

fastcgi_cache_key_s :: [Text] -> Fastcgi_Cache_Key

Safe version of fastcgi_cache_key

newtype Fastcgi_Cache_Key

Constructors

Fastcgi_Cache_Key ConfStatement

fastcgi_cache_lock :: [Text] -> ConfStatement

When enabled, only one request at a time will be allowed to populate a new cache element identified according to the directive by passing a request to a FastCGI server. Other requests of the same cache element will either wait for a response to appear in the cache or the cache lock for this element to be released, up to the time set by the directive.

fastcgi_cache_lock_s :: [Text] -> Fastcgi_Cache_Lock

Safe version of fastcgi_cache_lock

newtype Fastcgi_Cache_Lock

Constructors

Fastcgi_Cache_Lock ConfStatement

fastcgi_cache_lock_age :: [Text] -> ConfStatement

If the last request passed to the FastCGI server for populating a new cache element has not completed for the specified time, one more request may be passed to the FastCGI server.

fastcgi_cache_lock_age_s :: [Text] -> Fastcgi_Cache_Lock_Age

Safe version of fastcgi_cache_lock_age

newtype Fastcgi_Cache_Lock_Age

Constructors

Fastcgi_Cache_Lock_Age ConfStatement

fastcgi_cache_lock_timeout :: [Text] -> ConfStatement

Sets a timeout for . When the time expires, the request will be passed to the FastCGI server, however, the response will not be cached. Before 1.7.8, the response could be cached.

fastcgi_cache_lock_timeout_s :: [Text] -> Fastcgi_Cache_Lock_Timeout

Safe version of fastcgi_cache_lock_timeout

newtype Fastcgi_Cache_Lock_Timeout

Constructors

Fastcgi_Cache_Lock_Timeout ConfStatement

fastcgi_cache_methods :: [Text] -> ConfStatement

If the client request method is listed in this directive then the response will be cached. “GET” and “HEAD” methods are always added to the list, though it is recommended to specify them explicitly. See also the directive.

fastcgi_cache_methods_s :: [Text] -> Fastcgi_Cache_Methods

Safe version of fastcgi_cache_methods

newtype Fastcgi_Cache_Methods

Constructors

Fastcgi_Cache_Methods ConfStatement

fastcgi_cache_min_uses :: [Text] -> ConfStatement

Sets the number of requests after which the response will be cached.

fastcgi_cache_min_uses_s :: [Text] -> Fastcgi_Cache_Min_Uses

Safe version of fastcgi_cache_min_uses

newtype Fastcgi_Cache_Min_Uses

Constructors

Fastcgi_Cache_Min_Uses ConfStatement

fastcgi_cache_path :: [Text] -> ConfStatement

Sets the path and other parameters of a cache. Cache data are stored in files. Both the key and file name in a cache are a result of applying the MD5 function to the proxied URL. The levels parameter defines hierarchy levels of a cache. For example, in the following configuration fastcgi_cache_path datanginxcache levels=1:2 keys_zone=one:10m; file names in a cache will look like this: datanginxcachec29/b7f54b2df7773722d382f4809d65029c

fastcgi_cache_path_s :: [Text] -> Fastcgi_Cache_Path

Safe version of fastcgi_cache_path

newtype Fastcgi_Cache_Path

Constructors

Fastcgi_Cache_Path ConfStatement

fastcgi_cache_purge :: [Text] -> ConfStatement

Defines conditions under which the request will be considered a cache purge request. If at least one value of the string parameters is not empty and is not equal to “0” then the cache entry with a corresponding cache key is removed. The result of successful operation is indicated by returning the response.

fastcgi_cache_purge_s :: [Text] -> Fastcgi_Cache_Purge

Safe version of fastcgi_cache_purge

newtype Fastcgi_Cache_Purge

Constructors

Fastcgi_Cache_Purge ConfStatement

fastcgi_cache_revalidate :: [Text] -> ConfStatement

Enables revalidation of expired cache items using conditional requests with the If-Modified-Since and If-None-Match header fields.

fastcgi_cache_revalidate_s :: [Text] -> Fastcgi_Cache_Revalidate

Safe version of fastcgi_cache_revalidate

newtype Fastcgi_Cache_Revalidate

Constructors

Fastcgi_Cache_Revalidate ConfStatement

fastcgi_cache_use_stale :: [Text] -> ConfStatement

Determines in which cases a stale cached response can be used when an error occurs during communication with the FastCGI server. The directive’s parameters match the parameters of the directive.

fastcgi_cache_use_stale_s :: [Text] -> Fastcgi_Cache_Use_Stale

Safe version of fastcgi_cache_use_stale

newtype Fastcgi_Cache_Use_Stale

Constructors

Fastcgi_Cache_Use_Stale ConfStatement

fastcgi_cache_valid :: [Text] -> ConfStatement

Sets caching time for different response codes. For example, the following directives fastcgi_cache_valid 200 302 10m; fastcgi_cache_valid 404 1m; set 10 minutes of caching for responses with codes 200 and 302 and 1 minute for responses with code 404.

fastcgi_cache_valid_s :: [Text] -> Fastcgi_Cache_Valid

Safe version of fastcgi_cache_valid

newtype Fastcgi_Cache_Valid

Constructors

Fastcgi_Cache_Valid ConfStatement

fastcgi_catch_stderr :: [Text] -> ConfStatement

Sets a string to search for in the error stream of a response received from a FastCGI server. If the string is found then it is considered that the FastCGI server has returned an invalid response. This allows handling application errors in nginx, for example: location /php { fastcgi_pass backend:9000; ... fastcgi_catch_stderr "PHP Fatal error"; fastcgi_next_upstream error timeout invalid_header; }

fastcgi_catch_stderr_s :: [Text] -> Fastcgi_Catch_Stderr

Safe version of fastcgi_catch_stderr

newtype Fastcgi_Catch_Stderr

Constructors

Fastcgi_Catch_Stderr ConfStatement

fastcgi_connect_timeout :: [Text] -> ConfStatement

Defines a timeout for establishing a connection with a FastCGI server. It should be noted that this timeout cannot usually exceed 75 seconds.

fastcgi_connect_timeout_s :: [Text] -> Fastcgi_Connect_Timeout

Safe version of fastcgi_connect_timeout

newtype Fastcgi_Connect_Timeout

Constructors

Fastcgi_Connect_Timeout ConfStatement

fastcgi_force_ranges :: [Text] -> ConfStatement

Enables byte-range support for both cached and uncached responses from the FastCGI server regardless of the Accept-Ranges field in these responses.

fastcgi_force_ranges_s :: [Text] -> Fastcgi_Force_Ranges

Safe version of fastcgi_force_ranges

newtype Fastcgi_Force_Ranges

Constructors

Fastcgi_Force_Ranges ConfStatement

fastcgi_hide_header :: [Text] -> ConfStatement

By default, nginx does not pass the header fields Status and X-Accel-... from the response of a FastCGI server to a client. The fastcgi_hide_header directive sets additional fields that will not be passed. If, on the contrary, the passing of fields needs to be permitted, the directive can be used.

fastcgi_hide_header_s :: [Text] -> Fastcgi_Hide_Header

Safe version of fastcgi_hide_header

newtype Fastcgi_Hide_Header

Constructors

Fastcgi_Hide_Header ConfStatement

fastcgi_ignore_client_abort :: [Text] -> ConfStatement

Determines whether the connection with a FastCGI server should be closed when a client closes the connection without waiting for a response.

fastcgi_ignore_client_abort_s :: [Text] -> Fastcgi_Ignore_Client_Abort

Safe version of fastcgi_ignore_client_abort

newtype Fastcgi_Ignore_Client_Abort

Constructors

Fastcgi_Ignore_Client_Abort ConfStatement

fastcgi_ignore_headers :: [Text] -> ConfStatement

Disables processing of certain response header fields from the FastCGI server. The following fields can be ignored: X-Accel-Redirect, X-Accel-Expires, X-Accel-Limit-Rate (1.1.6), X-Accel-Buffering (1.1.6), X-Accel-Charset (1.1.6), Expires, Cache-Control, Set-Cookie (0.8.44), and Vary (1.7.7).

fastcgi_ignore_headers_s :: [Text] -> Fastcgi_Ignore_Headers

Safe version of fastcgi_ignore_headers

newtype Fastcgi_Ignore_Headers

Constructors

Fastcgi_Ignore_Headers ConfStatement

fastcgi_index :: [Text] -> ConfStatement

Sets a file name that will be appended after a URI that ends with a slash, in the value of the $fastcgi_script_name variable. For example, with these settings fastcgi_index index.php; fastcgi_param SCRIPT_FILENAME homewwwscriptsphp$fastcgi_script_name; and the “page.php” request, the SCRIPT_FILENAME parameter will be equal to “homewwwscriptsphppage.php”, and with the “” request it will be equal to “homewwwscriptsphpindex.php”.

fastcgi_index_s :: [Text] -> Fastcgi_Index

Safe version of fastcgi_index

newtype Fastcgi_Index

Constructors

Fastcgi_Index ConfStatement

fastcgi_intercept_errors :: [Text] -> ConfStatement

Determines whether FastCGI server responses with codes greater than or equal to 300 should be passed to a client or be redirected to nginx for processing with the directive.

fastcgi_intercept_errors_s :: [Text] -> Fastcgi_Intercept_Errors

Safe version of fastcgi_intercept_errors

newtype Fastcgi_Intercept_Errors

Constructors

Fastcgi_Intercept_Errors ConfStatement

fastcgi_keep_conn :: [Text] -> ConfStatement

By default, a FastCGI server will close a connection right after sending the response. However, when this directive is set to the value on, nginx will instruct a FastCGI server to keep connections open. This is necessary, in particular, for connections to FastCGI servers to function.

fastcgi_keep_conn_s :: [Text] -> Fastcgi_Keep_Conn

Safe version of fastcgi_keep_conn

newtype Fastcgi_Keep_Conn

Constructors

Fastcgi_Keep_Conn ConfStatement

fastcgi_limit_rate :: [Text] -> ConfStatement

Limits the speed of reading the response from the FastCGI server. The rate is specified in bytes per second. The zero value disables rate limiting. The limit is set per a request, and so if nginx simultaneously opens two connections to the FastCFI server, the overall rate will be twice as much as the specified limit. The limitation works only if buffering of responses from the FastCGI server is enabled.

fastcgi_limit_rate_s :: [Text] -> Fastcgi_Limit_Rate

Safe version of fastcgi_limit_rate

newtype Fastcgi_Limit_Rate

Constructors

Fastcgi_Limit_Rate ConfStatement

fastcgi_max_temp_file_size :: [Text] -> ConfStatement

When buffering of responses from the FastCGI server is enabled, and the whole response does not fit into the buffers set by the and directives, a part of the response can be saved to a temporary file. This directive sets the maximum size of the temporary file. The size of data written to the temporary file at a time is set by the directive.

fastcgi_max_temp_file_size_s :: [Text] -> Fastcgi_Max_Temp_File_Size

Safe version of fastcgi_max_temp_file_size

newtype Fastcgi_Max_Temp_File_Size

Constructors

Fastcgi_Max_Temp_File_Size ConfStatement

fastcgi_next_upstream :: [Text] -> ConfStatement

Specifies in which cases a request should be passed to the next server: error an error occurred while establishing a connection with the server, passing a request to it, or reading the response header; timeout a timeout has occurred while establishing a connection with the server, passing a request to it, or reading the response header; invalid_header a server returned an empty or invalid response; http_500 a server returned a response with the code 500; http_503 a server returned a response with the code 503; http_403 a server returned a response with the code 403; http_404 a server returned a response with the code 404; non_idempotent normally, requests with a non-idempotent method (POST, LOCK, PATCH) are not passed to the next server if a request has been sent to an upstream server (1.9.13); enabling this option explicitly allows retrying such requests; off disables passing a request to the next server.

fastcgi_next_upstream_s :: [Text] -> Fastcgi_Next_Upstream

Safe version of fastcgi_next_upstream

newtype Fastcgi_Next_Upstream

Constructors

Fastcgi_Next_Upstream ConfStatement

fastcgi_next_upstream_timeout :: [Text] -> ConfStatement

Limits the time during which a request can be passed to the next server. The 0 value turns off this limitation.

fastcgi_next_upstream_timeout_s :: [Text] -> Fastcgi_Next_Upstream_Timeout

Safe version of fastcgi_next_upstream_timeout

newtype Fastcgi_Next_Upstream_Timeout

Constructors

Fastcgi_Next_Upstream_Timeout ConfStatement

fastcgi_next_upstream_tries :: [Text] -> ConfStatement

Limits the number of possible tries for passing a request to the next server. The 0 value turns off this limitation.

fastcgi_next_upstream_tries_s :: [Text] -> Fastcgi_Next_Upstream_Tries

Safe version of fastcgi_next_upstream_tries

newtype Fastcgi_Next_Upstream_Tries

Constructors

Fastcgi_Next_Upstream_Tries ConfStatement

fastcgi_no_cache :: [Text] -> ConfStatement

Defines conditions under which the response will not be saved to a cache. If at least one value of the string parameters is not empty and is not equal to “0” then the response will not be saved: fastcgi_no_cache $cookie_nocache $arg_nocache$arg_comment; fastcgi_no_cache $http_pragma $http_authorization; Can be used along with the directive.

fastcgi_no_cache_s :: [Text] -> Fastcgi_No_Cache

Safe version of fastcgi_no_cache

newtype Fastcgi_No_Cache

Constructors

Fastcgi_No_Cache ConfStatement

fastcgi_param :: [Text] -> ConfStatement

Sets a parameter that should be passed to the FastCGI server. The value can contain text, variables, and their combination. These directives are inherited from the previous level if and only if there are no fastcgi_param directives defined on the current level.

fastcgi_param_s :: [Text] -> Fastcgi_Param

Safe version of fastcgi_param

newtype Fastcgi_Param

Constructors

Fastcgi_Param ConfStatement

fastcgi_pass :: [Text] -> ConfStatement

Sets the address of a FastCGI server. The address can be specified as a domain name or IP address, and a port: fastcgi_pass localhost:9000; or as a UNIX-domain socket path: fastcgi_pass unix:tmpfastcgi.socket;

fastcgi_pass_s :: [Text] -> Fastcgi_Pass

Safe version of fastcgi_pass

newtype Fastcgi_Pass

Constructors

Fastcgi_Pass ConfStatement

fastcgi_pass_header :: [Text] -> ConfStatement

Permits passing otherwise disabled header fields from a FastCGI server to a client.

fastcgi_pass_header_s :: [Text] -> Fastcgi_Pass_Header

Safe version of fastcgi_pass_header

newtype Fastcgi_Pass_Header

Constructors

Fastcgi_Pass_Header ConfStatement

fastcgi_pass_request_body :: [Text] -> ConfStatement

Indicates whether the original request body is passed to the FastCGI server. See also the directive.

fastcgi_pass_request_body_s :: [Text] -> Fastcgi_Pass_Request_Body

Safe version of fastcgi_pass_request_body

newtype Fastcgi_Pass_Request_Body

Constructors

Fastcgi_Pass_Request_Body ConfStatement

fastcgi_pass_request_headers :: [Text] -> ConfStatement

Indicates whether the header fields of the original request are passed to the FastCGI server. See also the directive.

fastcgi_pass_request_headers_s :: [Text] -> Fastcgi_Pass_Request_Headers

Safe version of fastcgi_pass_request_headers

newtype Fastcgi_Pass_Request_Headers

Constructors

Fastcgi_Pass_Request_Headers ConfStatement

fastcgi_read_timeout :: [Text] -> ConfStatement

Defines a timeout for reading a response from the FastCGI server. The timeout is set only between two successive read operations, not for the transmission of the whole response. If the FastCGI server does not transmit anything within this time, the connection is closed.

fastcgi_read_timeout_s :: [Text] -> Fastcgi_Read_Timeout

Safe version of fastcgi_read_timeout

newtype Fastcgi_Read_Timeout

Constructors

Fastcgi_Read_Timeout ConfStatement

fastcgi_request_buffering :: [Text] -> ConfStatement

Enables or disables buffering of a client request body.

fastcgi_request_buffering_s :: [Text] -> Fastcgi_Request_Buffering

Safe version of fastcgi_request_buffering

newtype Fastcgi_Request_Buffering

Constructors

Fastcgi_Request_Buffering ConfStatement

fastcgi_send_lowat :: [Text] -> ConfStatement

If the directive is set to a non-zero value, nginx will try to minimize the number of send operations on outgoing connections to a FastCGI server by using either NOTE_LOWAT flag of the method, or the SO_SNDLOWAT socket option, with the specified size.

fastcgi_send_lowat_s :: [Text] -> Fastcgi_Send_Lowat

Safe version of fastcgi_send_lowat

newtype Fastcgi_Send_Lowat

Constructors

Fastcgi_Send_Lowat ConfStatement

fastcgi_send_timeout :: [Text] -> ConfStatement

Sets a timeout for transmitting a request to the FastCGI server. The timeout is set only between two successive write operations, not for the transmission of the whole request. If the FastCGI server does not receive anything within this time, the connection is closed.

fastcgi_send_timeout_s :: [Text] -> Fastcgi_Send_Timeout

Safe version of fastcgi_send_timeout

newtype Fastcgi_Send_Timeout

Constructors

Fastcgi_Send_Timeout ConfStatement

fastcgi_split_path_info :: [Text] -> ConfStatement

Defines a regular expression that captures a value for the $fastcgi_path_info variable. The regular expression should have two captures: the first becomes a value of the $fastcgi_script_name variable, the second becomes a value of the $fastcgi_path_info variable. For example, with these settings location ~ ^(.+.php)(.*)$ { fastcgi_split_path_info ^(.+.php)(.*)$; fastcgi_param SCRIPT_FILENAME pathtophp$fastcgi_script_name; fastcgi_param PATH_INFO $fastcgi_path_info; and the “show.phparticle0001” request, the SCRIPT_FILENAME parameter will be equal to “pathtophpshow.php”, and the PATH_INFO parameter will be equal to “article0001”.

fastcgi_split_path_info_s :: [Text] -> Fastcgi_Split_Path_Info

Safe version of fastcgi_split_path_info

newtype Fastcgi_Split_Path_Info

Constructors

Fastcgi_Split_Path_Info ConfStatement

fastcgi_store :: [Text] -> ConfStatement

Enables saving of files to a disk. The on parameter saves files with paths corresponding to the directives or . The off parameter disables saving of files. In addition, the file name can be set explicitly using the string with variables: fastcgi_store datawww$original_uri;

fastcgi_store_s :: [Text] -> Fastcgi_Store

Safe version of fastcgi_store

newtype Fastcgi_Store

Constructors

Fastcgi_Store ConfStatement

fastcgi_store_access :: [Text] -> ConfStatement

Sets access permissions for newly created files and directories, e.g.: fastcgi_store_access user:rw group:rw all:r;

fastcgi_store_access_s :: [Text] -> Fastcgi_Store_Access

Safe version of fastcgi_store_access

newtype Fastcgi_Store_Access

Constructors

Fastcgi_Store_Access ConfStatement

fastcgi_temp_file_write_size :: [Text] -> ConfStatement

Limits the size of data written to a temporary file at a time, when buffering of responses from the FastCGI server to temporary files is enabled. By default, size is limited by two buffers set by the and directives. The maximum size of a temporary file is set by the directive.

fastcgi_temp_file_write_size_s :: [Text] -> Fastcgi_Temp_File_Write_Size

Safe version of fastcgi_temp_file_write_size

newtype Fastcgi_Temp_File_Write_Size

Constructors

Fastcgi_Temp_File_Write_Size ConfStatement

fastcgi_temp_path :: [Text] -> ConfStatement

Defines a directory for storing temporary files with data received from FastCGI servers. Up to three-level subdirectory hierarchy can be used underneath the specified directory. For example, in the following configuration fastcgi_temp_path spoolnginxfastcgi_temp 1 2; a temporary file might look like this: spoolnginxfastcgi_temp745/00000123457

fastcgi_temp_path_s :: [Text] -> Fastcgi_Temp_Path

Safe version of fastcgi_temp_path

newtype Fastcgi_Temp_Path

Constructors

Fastcgi_Temp_Path ConfStatement