Nginx_Ngx_Http_Gunzip_Module

Documentation

gunzip :: [Text] -> ConfStatement

Enables or disables decompression of gzipped responses for clients that lack gzip support. If enabled, the following directives are also taken into account when determining if clients support gzip: , , and . See also the directive.

gunzip_s :: [Text] -> Gunzip

Safe version of gunzip

newtype Gunzip

Constructors

Gunzip ConfStatement

gunzip_buffers :: [Text] -> ConfStatement

Sets the number and size of buffers used to decompress a response. By default, the buffer size is equal to one memory page. This is either 4K or 8K, depending on a platform.

gunzip_buffers_s :: [Text] -> Gunzip_Buffers

Safe version of gunzip_buffers

newtype Gunzip_Buffers

Constructors

Gunzip_Buffers ConfStatement