Nginx_Ngx_Http_Gzip_Module

Documentation

gzip :: [Text] -> ConfStatement

Enables or disables gzipping of responses.

gzip_s :: [Text] -> Gzip

Safe version of gzip

newtype Gzip

Constructors

Gzip ConfStatement

gzip_buffers :: [Text] -> ConfStatement

Sets the number and size of buffers used to compress a response. By default, the buffer size is equal to one memory page. This is either 4K or 8K, depending on a platform. Until version 0.7.28, four 4K or 8K buffers were used by default.

gzip_buffers_s :: [Text] -> Gzip_Buffers

Safe version of gzip_buffers

newtype Gzip_Buffers

Constructors

Gzip_Buffers ConfStatement

gzip_comp_level :: [Text] -> ConfStatement

Sets a gzip compression level of a response. Acceptable values are in the range from 1 to 9.

gzip_comp_level_s :: [Text] -> Gzip_Comp_Level

Safe version of gzip_comp_level

newtype Gzip_Comp_Level

Constructors

Gzip_Comp_Level ConfStatement

gzip_disable :: [Text] -> ConfStatement

Disables gzipping of responses for requests with User-Agent header fields matching any of the specified regular expressions.

gzip_disable_s :: [Text] -> Gzip_Disable

Safe version of gzip_disable

newtype Gzip_Disable

Constructors

Gzip_Disable ConfStatement

gzip_min_length :: [Text] -> ConfStatement

Sets the minimum length of a response that will be gzipped. The length is determined only from the Content-Length response header field.

gzip_min_length_s :: [Text] -> Gzip_Min_Length

Safe version of gzip_min_length

newtype Gzip_Min_Length

Constructors

Gzip_Min_Length ConfStatement

gzip_http_version :: [Text] -> ConfStatement

Sets the minimum HTTP version of a request required to compress a response.

gzip_http_version_s :: [Text] -> Gzip_Http_Version

Safe version of gzip_http_version

newtype Gzip_Http_Version

Constructors

Gzip_Http_Version ConfStatement

gzip_proxied :: [Text] -> ConfStatement

Enables or disables gzipping of responses for proxied requests depending on the request and response. The fact that the request is proxied is determined by the presence of the Via request header field. The directive accepts multiple parameters: off disables compression for all proxied requests, ignoring other parameters; expired enables compression if a response header includes the Expires field with a value that disables caching; no-cache enables compression if a response header includes the Cache-Control field with the “no-cache” parameter; no-store enables compression if a response header includes the Cache-Control field with the “no-store” parameter; private enables compression if a response header includes the Cache-Control field with the “private” parameter; no_last_modified enables compression if a response header does not include the Last-Modified field; no_etag enables compression if a response header does not include the ETag field; auth enables compression if a request header includes the Authorization field; any enables compression for all proxied requests.

gzip_proxied_s :: [Text] -> Gzip_Proxied

Safe version of gzip_proxied

newtype Gzip_Proxied

Constructors

Gzip_Proxied ConfStatement

gzip_types :: [Text] -> ConfStatement

Enables gzipping of responses for the specified MIME types in addition to “texthtml”. The special value “*” matches any MIME type (0.8.29). Responses with the “texthtml” type are always compressed.

gzip_types_s :: [Text] -> Gzip_Types

Safe version of gzip_types

newtype Gzip_Types

Constructors

Gzip_Types ConfStatement

gzip_vary :: [Text] -> ConfStatement

Enables or disables inserting the Vary: Accept-Encoding response header field if the directives , , or are active.

gzip_vary_s :: [Text] -> Gzip_Vary

Safe version of gzip_vary

newtype Gzip_Vary

Constructors

Gzip_Vary ConfStatement