Nginx_Ngx_Http_Headers_Module

Documentation

add_header :: [Text] -> ConfStatement

Adds the specified field to a response header provided that the response code equals 200, 201, 204, 206, 301, 302, 303, 304, or 307. A value can contain variables.

add_header_s :: [Text] -> Add_Header

Safe version of add_header

newtype Add_Header

Constructors

Add_Header ConfStatement

expires :: [Text] -> ConfStatement

Enables or disables adding or modifying the Expires and Cache-Control response header fields provided that the response code equals 200, 201, 204, 206, 301, 302, 303, 304, or 307. A parameter can be a positive or negative time.

expires_s :: [Text] -> Expires

Safe version of expires

newtype Expires

Constructors

Expires ConfStatement