Nginx_Ngx_Http_Hls_Module

Documentation

hls :: [Text] -> ConfStatement

Turns on HLS streaming in the surrounding location.

hls_s :: [Text] -> Hls

Safe version of hls

newtype Hls

Constructors

Hls ConfStatement

hls_buffers :: [Text] -> ConfStatement

Sets the maximum number and size of buffers that are used for reading and writing data frames.

hls_buffers_s :: [Text] -> Hls_Buffers

Safe version of hls_buffers

newtype Hls_Buffers

Constructors

Hls_Buffers ConfStatement

hls_forward_args :: [Text] -> ConfStatement

Adds arguments from a playlist request to URIs of fragments. This may be useful for performing client authorization at the moment of requesting a fragment, or when protecting an HLS stream with the ngx_http_secure_link_module module.

hls_forward_args_s :: [Text] -> Hls_Forward_Args

Safe version of hls_forward_args

newtype Hls_Forward_Args

Constructors

Hls_Forward_Args ConfStatement

hls_fragment :: [Text] -> ConfStatement

Defines the default fragment length for playlist URIs requested without the “len” argument.

hls_fragment_s :: [Text] -> Hls_Fragment

Safe version of hls_fragment

newtype Hls_Fragment

Constructors

Hls_Fragment ConfStatement

hls_mp4_buffer_size :: [Text] -> ConfStatement

Sets the initial size of the buffer used for processing MP4 and MOV files.

hls_mp4_buffer_size_s :: [Text] -> Hls_Mp4_Buffer_Size

Safe version of hls_mp4_buffer_size

newtype Hls_Mp4_Buffer_Size

Constructors

Hls_Mp4_Buffer_Size ConfStatement

hls_mp4_max_buffer_size :: [Text] -> ConfStatement

During metadata processing, a larger buffer may become necessary. Its size cannot exceed the specified size, or else nginx will return the server error , and log the following message: "somemovie/file.mp4" mp4 moov atom is too large: 12583268, you may want to increase hls_mp4_max_buffer_size

hls_mp4_max_buffer_size_s :: [Text] -> Hls_Mp4_Max_Buffer_Size

Safe version of hls_mp4_max_buffer_size

newtype Hls_Mp4_Max_Buffer_Size

Constructors

Hls_Mp4_Max_Buffer_Size ConfStatement