Nginx_Ngx_Http_Image_Filter_Module

Documentation

image_filter :: [Text] -> ConfStatement

Sets the type of transformation to perform on images: off turns off module processing in a surrounding location. test ensures that responses are images in either JPEG, GIF, or PNG format. Otherwise, the error is returned. size outputs information about images in a JSON format, e.g.: { "img" : { "width": 100, "height": 100, "type": "gif" } } In case of an error, the output is as follows: {} rotate 90|180|270 rotates images counter-clockwise by the specified number of degrees. Parameter value can contain variables. This mode can be used either alone or along with the resize and crop transformations. resize width height proportionally reduces an image to the specified sizes. To reduce by only one dimension, another dimension can be specified as “-”. In case of an error, the server will return code . Parameter values can contain variables. When used along with the rotate parameter, the rotation happens after reduction. crop width height proportionally reduces an image to the larger side size and crops extraneous edges by another side. To reduce by only one dimension, another dimension can be specified as “-”. In case of an error, the server will return code . Parameter values can contain variables. When used along with the rotate parameter, the rotation happens before reduction.

image_filter_s :: [Text] -> Image_Filter

Safe version of image_filter

newtype Image_Filter

Constructors

Image_Filter ConfStatement

image_filter_buffer :: [Text] -> ConfStatement

Sets the maximum size of the buffer used for reading images. When the size is exceeded the server returns error .

image_filter_buffer_s :: [Text] -> Image_Filter_Buffer

Safe version of image_filter_buffer

newtype Image_Filter_Buffer

Constructors

Image_Filter_Buffer ConfStatement

image_filter_interlace :: [Text] -> ConfStatement

If enabled, final images will be interlaced. For JPEG, final images will be in “progressive JPEG” format.

image_filter_interlace_s :: [Text] -> Image_Filter_Interlace

Safe version of image_filter_interlace

newtype Image_Filter_Interlace

Constructors

Image_Filter_Interlace ConfStatement

image_filter_jpeg_quality :: [Text] -> ConfStatement

Sets the desired quality of the transformed JPEG images. Acceptable values are in the range from 1 to 100. Lesser values usually imply both lower image quality and less data to transfer. The maximum recommended value is 95. Parameter value can contain variables.

image_filter_jpeg_quality_s :: [Text] -> Image_Filter_Jpeg_Quality

Safe version of image_filter_jpeg_quality

newtype Image_Filter_Jpeg_Quality

Constructors

Image_Filter_Jpeg_Quality ConfStatement

image_filter_sharpen :: [Text] -> ConfStatement

Increases sharpness of the final image. The sharpness percentage can exceed 100. The zero value disables sharpening. Parameter value can contain variables.

image_filter_sharpen_s :: [Text] -> Image_Filter_Sharpen

Safe version of image_filter_sharpen

newtype Image_Filter_Sharpen

Constructors

Image_Filter_Sharpen ConfStatement

image_filter_transparency :: [Text] -> ConfStatement

Defines whether transparency should be preserved when transforming GIF images or PNG images with colors specified by a palette. The loss of transparency results in images of a better quality. The alpha channel transparency in PNG is always preserved.

image_filter_transparency_s :: [Text] -> Image_Filter_Transparency

Safe version of image_filter_transparency

newtype Image_Filter_Transparency

Constructors

Image_Filter_Transparency ConfStatement