Nginx_Ngx_Http_Index_Module

Documentation

index :: [Text] -> ConfStatement

Defines files that will be used as an index. The file name can contain variables. Files are checked in the specified order. The last element of the list can be a file with an absolute path. Example: index index.$geo.html index.0.html /index.html;

index_s :: [Text] -> Index

Safe version of index

newtype Index

Constructors

Index ConfStatement