Nginx_Ngx_Http_Limit_Req_Module

Documentation

limit_req :: [Text] -> ConfStatement

Sets the shared memory zone and the maximum burst size of requests. If the requests rate exceeds the rate configured for a zone, their processing is delayed such that requests are processed at a defined rate. Excessive requests are delayed until their number exceeds the maximum burst size in which case the request is terminated with an error . By default, the maximum burst size is equal to zero. For example, the directives limit_req_zone $binary_remote_addr zone=one:10m rate=1rs; server { location search/ { limit_req zone=one burst=5; } allow not more than 1 request per second at an average, with bursts not exceeding 5 requests.

limit_req_s :: [Text] -> Limit_Req

Safe version of limit_req

newtype Limit_Req

Constructors

Limit_Req ConfStatement

limit_req_log_level :: [Text] -> ConfStatement

Sets the desired logging level for cases when the server refuses to process requests due to rate exceeding, or delays request processing. Logging level for delays is one point less than for refusals; for example, if “limit_req_log_level notice” is specified, delays are logged with the info level.

limit_req_log_level_s :: [Text] -> Limit_Req_Log_Level

Safe version of limit_req_log_level

newtype Limit_Req_Log_Level

Constructors

Limit_Req_Log_Level ConfStatement

limit_req_status :: [Text] -> ConfStatement

Sets the status code to return in response to rejected requests.

limit_req_status_s :: [Text] -> Limit_Req_Status

Safe version of limit_req_status

newtype Limit_Req_Status

Constructors

Limit_Req_Status ConfStatement

limit_req_zone :: [Text] -> ConfStatement

Sets parameters for a shared memory zone that will keep states for various keys. In particular, the state stores the current number of excessive requests. The key can contain text, variables, and their combination. Requests with an empty key value are not accounted. Prior to version 1.7.6, a key could contain exactly one variable. Usage example: limit_req_zone $binary_remote_addr zone=one:10m rate=1r/s;

limit_req_zone_s :: [Text] -> Limit_Req_Zone

Safe version of limit_req_zone

newtype Limit_Req_Zone

Constructors

Limit_Req_Zone ConfStatement