Nginx_Ngx_Http_Log_Module

Documentation

access_log :: [Text] -> ConfStatement

Sets the path, format, and configuration for a buffered log write. Several logs can be specified on the same level. Logging to syslog can be configured by specifying the “syslog:” prefix in the first parameter. The special value off cancels all access_log directives on the current level. If the format is not specified then the predefined “combined” format is used.

access_log_s :: [Text] -> Access_Log

Safe version of access_log

newtype Access_Log

Constructors

Access_Log ConfStatement

log_format :: [Text] -> ConfStatement

Specifies log format.

log_format_s :: [Text] -> Log_Format

Safe version of log_format

newtype Log_Format

Constructors

Log_Format ConfStatement

open_log_file_cache :: [Text] -> ConfStatement

Defines a cache that stores the file descriptors of frequently used logs whose names contain variables. The directive has the following parameters: max sets the maximum number of descriptors in a cache; if the cache becomes full the least recently used (LRU) descriptors are closed inactive sets the time after which the cached descriptor is closed if there were no access during this time; by default, 10 seconds min_uses sets the minimum number of file uses during the time defined by the inactive parameter to let the descriptor stay open in a cache; by default, 1 valid sets the time after which it should be checked that the file still exists with the same name; by default, 60 seconds off disables caching

open_log_file_cache_s :: [Text] -> Open_Log_File_Cache

Safe version of open_log_file_cache

newtype Open_Log_File_Cache

Constructors

Open_Log_File_Cache ConfStatement