Nginx_Ngx_Http_Memcached_Module

Documentation

memcached_bind :: [Text] -> ConfStatement

Makes outgoing connections to a memcached server originate from the specified local IP address with an optional port (1.11.2). Parameter value can contain variables (1.3.12). The special value off (1.3.12) cancels the effect of the memcached_bind directive inherited from the previous configuration level, which allows the system to auto-assign the local IP address and port.

memcached_bind_s :: [Text] -> Memcached_Bind

Safe version of memcached_bind

newtype Memcached_Bind

Constructors

Memcached_Bind ConfStatement

memcached_buffer_size :: [Text] -> ConfStatement

Sets the size of the buffer used for reading the response received from the memcached server. The response is passed to the client synchronously, as soon as it is received.

memcached_buffer_size_s :: [Text] -> Memcached_Buffer_Size

Safe version of memcached_buffer_size

newtype Memcached_Buffer_Size

Constructors

Memcached_Buffer_Size ConfStatement

memcached_connect_timeout :: [Text] -> ConfStatement

Defines a timeout for establishing a connection with a memcached server. It should be noted that this timeout cannot usually exceed 75 seconds.

memcached_connect_timeout_s :: [Text] -> Memcached_Connect_Timeout

Safe version of memcached_connect_timeout

newtype Memcached_Connect_Timeout

Constructors

Memcached_Connect_Timeout ConfStatement

memcached_force_ranges :: [Text] -> ConfStatement

Enables byte-range support for both cached and uncached responses from the memcached server regardless of the Accept-Ranges field in these responses.

memcached_force_ranges_s :: [Text] -> Memcached_Force_Ranges

Safe version of memcached_force_ranges

newtype Memcached_Force_Ranges

Constructors

Memcached_Force_Ranges ConfStatement

memcached_gzip_flag :: [Text] -> ConfStatement

Enables the test for the flag presence in the memcached server response and sets the “Content-Encoding” response header field to “gzip” if the flag is set.

memcached_gzip_flag_s :: [Text] -> Memcached_Gzip_Flag

Safe version of memcached_gzip_flag

newtype Memcached_Gzip_Flag

Constructors

Memcached_Gzip_Flag ConfStatement

memcached_next_upstream :: [Text] -> ConfStatement

Specifies in which cases a request should be passed to the next server: error an error occurred while establishing a connection with the server, passing a request to it, or reading the response header; timeout a timeout has occurred while establishing a connection with the server, passing a request to it, or reading the response header; invalid_response a server returned an empty or invalid response; not_found a response was not found on the server; off disables passing a request to the next server.

memcached_next_upstream_s :: [Text] -> Memcached_Next_Upstream

Safe version of memcached_next_upstream

newtype Memcached_Next_Upstream

Constructors

Memcached_Next_Upstream ConfStatement

memcached_next_upstream_timeout :: [Text] -> ConfStatement

Limits the time during which a request can be passed to the next server. The 0 value turns off this limitation.

memcached_next_upstream_timeout_s :: [Text] -> Memcached_Next_Upstream_Timeout

Safe version of memcached_next_upstream_timeout

newtype Memcached_Next_Upstream_Timeout

Constructors

Memcached_Next_Upstream_Timeout ConfStatement

memcached_next_upstream_tries :: [Text] -> ConfStatement

Limits the number of possible tries for passing a request to the next server. The 0 value turns off this limitation.

memcached_next_upstream_tries_s :: [Text] -> Memcached_Next_Upstream_Tries

Safe version of memcached_next_upstream_tries

newtype Memcached_Next_Upstream_Tries

Constructors

Memcached_Next_Upstream_Tries ConfStatement

memcached_pass :: [Text] -> ConfStatement

Sets the memcached server address. The address can be specified as a domain name or IP address, and a port: memcached_pass localhost:11211; or as a UNIX-domain socket path: memcached_pass unix:tmpmemcached.socket;

memcached_pass_s :: [Text] -> Memcached_Pass

Safe version of memcached_pass

newtype Memcached_Pass

Constructors

Memcached_Pass ConfStatement

memcached_read_timeout :: [Text] -> ConfStatement

Defines a timeout for reading a response from the memcached server. The timeout is set only between two successive read operations, not for the transmission of the whole response. If the memcached server does not transmit anything within this time, the connection is closed.

memcached_read_timeout_s :: [Text] -> Memcached_Read_Timeout

Safe version of memcached_read_timeout

newtype Memcached_Read_Timeout

Constructors

Memcached_Read_Timeout ConfStatement

memcached_send_timeout :: [Text] -> ConfStatement

Sets a timeout for transmitting a request to the memcached server. The timeout is set only between two successive write operations, not for the transmission of the whole request. If the memcached server does not receive anything within this time, the connection is closed.

memcached_send_timeout_s :: [Text] -> Memcached_Send_Timeout

Safe version of memcached_send_timeout

newtype Memcached_Send_Timeout

Constructors

Memcached_Send_Timeout ConfStatement