Nginx_Ngx_Http_Mp4_Module

Documentation

mp4 :: [Text] -> ConfStatement

Turns on module processing in a surrounding location.

mp4_s :: [Text] -> Mp4

Safe version of mp4

newtype Mp4

Constructors

Mp4 ConfStatement

mp4_buffer_size :: [Text] -> ConfStatement

Sets the initial size of the buffer used for processing MP4 files.

mp4_buffer_size_s :: [Text] -> Mp4_Buffer_Size

Safe version of mp4_buffer_size

newtype Mp4_Buffer_Size

Constructors

Mp4_Buffer_Size ConfStatement

mp4_max_buffer_size :: [Text] -> ConfStatement

During metadata processing, a larger buffer may become necessary. Its size cannot exceed the specified size, or else nginx will return the server error, and log the following message: "somemovie/file.mp4" mp4 moov atom is too large: 12583268, you may want to increase mp4_max_buffer_size

mp4_max_buffer_size_s :: [Text] -> Mp4_Max_Buffer_Size

Safe version of mp4_max_buffer_size

newtype Mp4_Max_Buffer_Size

Constructors

Mp4_Max_Buffer_Size ConfStatement

mp4_limit_rate :: [Text] -> ConfStatement

Limits the rate of response transmission to a client. The rate is limited based on the average bitrate of the MP4 file served. To calculate the rate, the bitrate is multiplied by the specified factor. The special value “on” corresponds to the factor of 1.1. The special value “off” disables rate limiting. The limit is set per a request, and so if a client simultaneously opens two connections, the overall rate will be twice as much as the specified limit.

mp4_limit_rate_s :: [Text] -> Mp4_Limit_Rate

Safe version of mp4_limit_rate

newtype Mp4_Limit_Rate

Constructors

Mp4_Limit_Rate ConfStatement

mp4_limit_rate_after :: [Text] -> ConfStatement

Sets the initial amount of media data (measured in playback time) after which the further transmission of the response to a client will be rate limited.

mp4_limit_rate_after_s :: [Text] -> Mp4_Limit_Rate_After

Safe version of mp4_limit_rate_after

newtype Mp4_Limit_Rate_After

Constructors

Mp4_Limit_Rate_After ConfStatement