Nginx_Ngx_Http_Proxy_Module

Documentation

proxy_bind :: [Text] -> ConfStatement

Makes outgoing connections to a proxied server originate from the specified local IP address with an optional port (1.11.2). Parameter value can contain variables (1.3.12). The special value off (1.3.12) cancels the effect of the proxy_bind directive inherited from the previous configuration level, which allows the system to auto-assign the local IP address and port.

proxy_bind_s :: [Text] -> Proxy_Bind

Safe version of proxy_bind

newtype Proxy_Bind

Constructors

Proxy_Bind ConfStatement

proxy_buffer_size :: [Text] -> ConfStatement

Sets the size of the buffer used for reading the first part of the response received from the proxied server. This part usually contains a small response header. By default, the buffer size is equal to one memory page. This is either 4K or 8K, depending on a platform. It can be made smaller, however.

proxy_buffer_size_s :: [Text] -> Proxy_Buffer_Size

Safe version of proxy_buffer_size

newtype Proxy_Buffer_Size

Constructors

Proxy_Buffer_Size ConfStatement

proxy_buffering :: [Text] -> ConfStatement

Enables or disables buffering of responses from the proxied server.

proxy_buffering_s :: [Text] -> Proxy_Buffering

Safe version of proxy_buffering

newtype Proxy_Buffering

Constructors

Proxy_Buffering ConfStatement

proxy_buffers :: [Text] -> ConfStatement

Sets the number and size of the buffers used for reading a response from the proxied server, for a single connection. By default, the buffer size is equal to one memory page. This is either 4K or 8K, depending on a platform.

proxy_buffers_s :: [Text] -> Proxy_Buffers

Safe version of proxy_buffers

newtype Proxy_Buffers

Constructors

Proxy_Buffers ConfStatement

proxy_busy_buffers_size :: [Text] -> ConfStatement

When buffering of responses from the proxied server is enabled, limits the total size of buffers that can be busy sending a response to the client while the response is not yet fully read. In the meantime, the rest of the buffers can be used for reading the response and, if needed, buffering part of the response to a temporary file. By default, size is limited by the size of two buffers set by the and directives.

proxy_busy_buffers_size_s :: [Text] -> Proxy_Busy_Buffers_Size

Safe version of proxy_busy_buffers_size

newtype Proxy_Busy_Buffers_Size

Constructors

Proxy_Busy_Buffers_Size ConfStatement

proxy_cache :: [Text] -> ConfStatement

Defines a shared memory zone used for caching. The same zone can be used in several places. Parameter value can contain variables (1.7.9). The off parameter disables caching inherited from the previous configuration level.

proxy_cache_s :: [Text] -> Proxy_Cache

Safe version of proxy_cache

newtype Proxy_Cache

Constructors

Proxy_Cache ConfStatement

proxy_cache_bypass :: [Text] -> ConfStatement

Defines conditions under which the response will not be taken from a cache. If at least one value of the string parameters is not empty and is not equal to “0” then the response will not be taken from the cache: proxy_cache_bypass $cookie_nocache $arg_nocache$arg_comment; proxy_cache_bypass $http_pragma $http_authorization; Can be used along with the directive.

proxy_cache_bypass_s :: [Text] -> Proxy_Cache_Bypass

Safe version of proxy_cache_bypass

newtype Proxy_Cache_Bypass

Constructors

Proxy_Cache_Bypass ConfStatement

proxy_cache_convert_head :: [Text] -> ConfStatement

Enables or disables the conversion of the “HEAD” method to “GET” for caching. When the conversion is disabled, the cache key should be configured to include the $request_method.

proxy_cache_convert_head_s :: [Text] -> Proxy_Cache_Convert_Head

Safe version of proxy_cache_convert_head

newtype Proxy_Cache_Convert_Head

Constructors

Proxy_Cache_Convert_Head ConfStatement

proxy_cache_key :: [Text] -> ConfStatement

Defines a key for caching, for example proxy_cache_key "$host$request_uri $cookie_user"; By default, the directive’s value is close to the string proxy_cache_key $scheme$proxy_host$uri$is_args$args;

proxy_cache_key_s :: [Text] -> Proxy_Cache_Key

Safe version of proxy_cache_key

newtype Proxy_Cache_Key

Constructors

Proxy_Cache_Key ConfStatement

proxy_cache_lock :: [Text] -> ConfStatement

When enabled, only one request at a time will be allowed to populate a new cache element identified according to the directive by passing a request to a proxied server. Other requests of the same cache element will either wait for a response to appear in the cache or the cache lock for this element to be released, up to the time set by the directive.

proxy_cache_lock_s :: [Text] -> Proxy_Cache_Lock

Safe version of proxy_cache_lock

newtype Proxy_Cache_Lock

Constructors

Proxy_Cache_Lock ConfStatement

proxy_cache_lock_age :: [Text] -> ConfStatement

If the last request passed to the proxied server for populating a new cache element has not completed for the specified time, one more request may be passed to the proxied server.

proxy_cache_lock_age_s :: [Text] -> Proxy_Cache_Lock_Age

Safe version of proxy_cache_lock_age

newtype Proxy_Cache_Lock_Age

Constructors

Proxy_Cache_Lock_Age ConfStatement

proxy_cache_lock_timeout :: [Text] -> ConfStatement

Sets a timeout for . When the time expires, the request will be passed to the proxied server, however, the response will not be cached. Before 1.7.8, the response could be cached.

proxy_cache_lock_timeout_s :: [Text] -> Proxy_Cache_Lock_Timeout

Safe version of proxy_cache_lock_timeout

newtype Proxy_Cache_Lock_Timeout

Constructors

Proxy_Cache_Lock_Timeout ConfStatement

proxy_cache_methods :: [Text] -> ConfStatement

If the client request method is listed in this directive then the response will be cached. “GET” and “HEAD” methods are always added to the list, though it is recommended to specify them explicitly. See also the directive.

proxy_cache_methods_s :: [Text] -> Proxy_Cache_Methods

Safe version of proxy_cache_methods

newtype Proxy_Cache_Methods

Constructors

Proxy_Cache_Methods ConfStatement

proxy_cache_min_uses :: [Text] -> ConfStatement

Sets the number of requests after which the response will be cached.

proxy_cache_min_uses_s :: [Text] -> Proxy_Cache_Min_Uses

Safe version of proxy_cache_min_uses

newtype Proxy_Cache_Min_Uses

Constructors

Proxy_Cache_Min_Uses ConfStatement

proxy_cache_path :: [Text] -> ConfStatement

Sets the path and other parameters of a cache. Cache data are stored in files. The file name in a cache is a result of applying the MD5 function to the cache key. The levels parameter defines hierarchy levels of a cache. For example, in the following configuration proxy_cache_path datanginxcache levels=1:2 keys_zone=one:10m; file names in a cache will look like this: datanginxcachec29/b7f54b2df7773722d382f4809d65029c

proxy_cache_path_s :: [Text] -> Proxy_Cache_Path

Safe version of proxy_cache_path

newtype Proxy_Cache_Path

Constructors

Proxy_Cache_Path ConfStatement

proxy_cache_purge :: [Text] -> ConfStatement

Defines conditions under which the request will be considered a cache purge request. If at least one value of the string parameters is not empty and is not equal to “0” then the cache entry with a corresponding cache key is removed. The result of successful operation is indicated by returning the response.

proxy_cache_purge_s :: [Text] -> Proxy_Cache_Purge

Safe version of proxy_cache_purge

newtype Proxy_Cache_Purge

Constructors

Proxy_Cache_Purge ConfStatement

proxy_cache_revalidate :: [Text] -> ConfStatement

Enables revalidation of expired cache items using conditional requests with the If-Modified-Since and If-None-Match header fields.

proxy_cache_revalidate_s :: [Text] -> Proxy_Cache_Revalidate

Safe version of proxy_cache_revalidate

newtype Proxy_Cache_Revalidate

Constructors

Proxy_Cache_Revalidate ConfStatement

proxy_cache_use_stale :: [Text] -> ConfStatement

Determines in which cases a stale cached response can be used when an error occurs during communication with the proxied server. The directive’s parameters match the parameters of the directive.

proxy_cache_use_stale_s :: [Text] -> Proxy_Cache_Use_Stale

Safe version of proxy_cache_use_stale

newtype Proxy_Cache_Use_Stale

Constructors

Proxy_Cache_Use_Stale ConfStatement

proxy_cache_valid :: [Text] -> ConfStatement

Sets caching time for different response codes. For example, the following directives proxy_cache_valid 200 302 10m; proxy_cache_valid 404 1m; set 10 minutes of caching for responses with codes 200 and 302 and 1 minute for responses with code 404.

proxy_cache_valid_s :: [Text] -> Proxy_Cache_Valid

Safe version of proxy_cache_valid

newtype Proxy_Cache_Valid

Constructors

Proxy_Cache_Valid ConfStatement

proxy_connect_timeout :: [Text] -> ConfStatement

Defines a timeout for establishing a connection with a proxied server. It should be noted that this timeout cannot usually exceed 75 seconds.

proxy_connect_timeout_s :: [Text] -> Proxy_Connect_Timeout

Safe version of proxy_connect_timeout

newtype Proxy_Connect_Timeout

Constructors

Proxy_Connect_Timeout ConfStatement

proxy_cookie_domain :: [Text] -> ConfStatement

Sets a text that should be changed in the domain attribute of the Set-Cookie header fields of a proxied server response. Suppose a proxied server returned the Set-Cookie header field with the attribute “domain=localhost”. The directive proxy_cookie_domain localhost example.org; will rewrite this attribute to “domain=example.org”.

proxy_cookie_domain_s :: [Text] -> Proxy_Cookie_Domain

Safe version of proxy_cookie_domain

newtype Proxy_Cookie_Domain

Constructors

Proxy_Cookie_Domain ConfStatement

proxy_cookie_path :: [Text] -> ConfStatement

Sets a text that should be changed in the path attribute of the Set-Cookie header fields of a proxied server response. Suppose a proxied server returned the Set-Cookie header field with the attribute “path=twosomeuri”. The directive proxy_cookie_path two ; will rewrite this attribute to “path=someuri”.

proxy_cookie_path_s :: [Text] -> Proxy_Cookie_Path

Safe version of proxy_cookie_path

newtype Proxy_Cookie_Path

Constructors

Proxy_Cookie_Path ConfStatement

proxy_force_ranges :: [Text] -> ConfStatement

Enables byte-range support for both cached and uncached responses from the proxied server regardless of the Accept-Ranges field in these responses.

proxy_force_ranges_s :: [Text] -> Proxy_Force_Ranges

Safe version of proxy_force_ranges

newtype Proxy_Force_Ranges

Constructors

Proxy_Force_Ranges ConfStatement

proxy_headers_hash_bucket_size :: [Text] -> ConfStatement

Sets the bucket size for hash tables used by the and directives. The details of setting up hash tables are provided in a separate document.

proxy_headers_hash_bucket_size_s :: [Text] -> Proxy_Headers_Hash_Bucket_Size

Safe version of proxy_headers_hash_bucket_size

newtype Proxy_Headers_Hash_Bucket_Size

Constructors

Proxy_Headers_Hash_Bucket_Size ConfStatement

proxy_headers_hash_max_size :: [Text] -> ConfStatement

Sets the maximum size of hash tables used by the and directives. The details of setting up hash tables are provided in a separate document.

proxy_headers_hash_max_size_s :: [Text] -> Proxy_Headers_Hash_Max_Size

Safe version of proxy_headers_hash_max_size

newtype Proxy_Headers_Hash_Max_Size

Constructors

Proxy_Headers_Hash_Max_Size ConfStatement

proxy_hide_header :: [Text] -> ConfStatement

By default, nginx does not pass the header fields Date, Server, X-Pad, and X-Accel-... from the response of a proxied server to a client. The proxy_hide_header directive sets additional fields that will not be passed. If, on the contrary, the passing of fields needs to be permitted, the directive can be used.

proxy_hide_header_s :: [Text] -> Proxy_Hide_Header

Safe version of proxy_hide_header

newtype Proxy_Hide_Header

Constructors

Proxy_Hide_Header ConfStatement

proxy_http_version :: [Text] -> ConfStatement

Sets the HTTP protocol version for proxying. By default, version 1.0 is used. Version 1.1 is recommended for use with connections and NTLM authentication.

proxy_http_version_s :: [Text] -> Proxy_Http_Version

Safe version of proxy_http_version

newtype Proxy_Http_Version

Constructors

Proxy_Http_Version ConfStatement

proxy_ignore_client_abort :: [Text] -> ConfStatement

Determines whether the connection with a proxied server should be closed when a client closes the connection without waiting for a response.

proxy_ignore_client_abort_s :: [Text] -> Proxy_Ignore_Client_Abort

Safe version of proxy_ignore_client_abort

newtype Proxy_Ignore_Client_Abort

Constructors

Proxy_Ignore_Client_Abort ConfStatement

proxy_ignore_headers :: [Text] -> ConfStatement

Disables processing of certain response header fields from the proxied server. The following fields can be ignored: X-Accel-Redirect, X-Accel-Expires, X-Accel-Limit-Rate (1.1.6), X-Accel-Buffering (1.1.6), X-Accel-Charset (1.1.6), Expires, Cache-Control, Set-Cookie (0.8.44), and Vary (1.7.7).

proxy_ignore_headers_s :: [Text] -> Proxy_Ignore_Headers

Safe version of proxy_ignore_headers

newtype Proxy_Ignore_Headers

Constructors

Proxy_Ignore_Headers ConfStatement

proxy_intercept_errors :: [Text] -> ConfStatement

Determines whether proxied responses with codes greater than or equal to 300 should be passed to a client or be redirected to nginx for processing with the directive.

proxy_intercept_errors_s :: [Text] -> Proxy_Intercept_Errors

Safe version of proxy_intercept_errors

newtype Proxy_Intercept_Errors

Constructors

Proxy_Intercept_Errors ConfStatement

proxy_limit_rate :: [Text] -> ConfStatement

Limits the speed of reading the response from the proxied server. The rate is specified in bytes per second. The zero value disables rate limiting. The limit is set per a request, and so if nginx simultaneously opens two connections to the proxied server, the overall rate will be twice as much as the specified limit. The limitation works only if buffering of responses from the proxied server is enabled.

proxy_limit_rate_s :: [Text] -> Proxy_Limit_Rate

Safe version of proxy_limit_rate

newtype Proxy_Limit_Rate

Constructors

Proxy_Limit_Rate ConfStatement

proxy_max_temp_file_size :: [Text] -> ConfStatement

When buffering of responses from the proxied server is enabled, and the whole response does not fit into the buffers set by the and directives, a part of the response can be saved to a temporary file. This directive sets the maximum size of the temporary file. The size of data written to the temporary file at a time is set by the directive.

proxy_max_temp_file_size_s :: [Text] -> Proxy_Max_Temp_File_Size

Safe version of proxy_max_temp_file_size

newtype Proxy_Max_Temp_File_Size

Constructors

Proxy_Max_Temp_File_Size ConfStatement

proxy_method :: [Text] -> ConfStatement

Specifies the HTTP method to use in requests forwarded to the proxied server instead of the method from the client request.

proxy_method_s :: [Text] -> Proxy_Method

Safe version of proxy_method

newtype Proxy_Method

Constructors

Proxy_Method ConfStatement

proxy_next_upstream :: [Text] -> ConfStatement

Specifies in which cases a request should be passed to the next server: error an error occurred while establishing a connection with the server, passing a request to it, or reading the response header; timeout a timeout has occurred while establishing a connection with the server, passing a request to it, or reading the response header; invalid_header a server returned an empty or invalid response; http_500 a server returned a response with the code 500; http_502 a server returned a response with the code 502; http_503 a server returned a response with the code 503; http_504 a server returned a response with the code 504; http_403 a server returned a response with the code 403; http_404 a server returned a response with the code 404; non_idempotent normally, requests with a non-idempotent method (POST, LOCK, PATCH) are not passed to the next server if a request has been sent to an upstream server (1.9.13); enabling this option explicitly allows retrying such requests; off disables passing a request to the next server.

proxy_next_upstream_s :: [Text] -> Proxy_Next_Upstream

Safe version of proxy_next_upstream

newtype Proxy_Next_Upstream

Constructors

Proxy_Next_Upstream ConfStatement

proxy_next_upstream_timeout :: [Text] -> ConfStatement

Limits the time during which a request can be passed to the next server. The 0 value turns off this limitation.

proxy_next_upstream_timeout_s :: [Text] -> Proxy_Next_Upstream_Timeout

Safe version of proxy_next_upstream_timeout

newtype Proxy_Next_Upstream_Timeout

Constructors

Proxy_Next_Upstream_Timeout ConfStatement

proxy_next_upstream_tries :: [Text] -> ConfStatement

Limits the number of possible tries for passing a request to the next server. The 0 value turns off this limitation.

proxy_next_upstream_tries_s :: [Text] -> Proxy_Next_Upstream_Tries

Safe version of proxy_next_upstream_tries

newtype Proxy_Next_Upstream_Tries

Constructors

Proxy_Next_Upstream_Tries ConfStatement

proxy_no_cache :: [Text] -> ConfStatement

Defines conditions under which the response will not be saved to a cache. If at least one value of the string parameters is not empty and is not equal to “0” then the response will not be saved: proxy_no_cache $cookie_nocache $arg_nocache$arg_comment; proxy_no_cache $http_pragma $http_authorization; Can be used along with the directive.

proxy_no_cache_s :: [Text] -> Proxy_No_Cache

Safe version of proxy_no_cache

newtype Proxy_No_Cache

Constructors

Proxy_No_Cache ConfStatement

proxy_pass :: [Text] -> ConfStatement

Sets the protocol and address of a proxied server and an optional URI to which a location should be mapped. As a protocol, “http” or “https” can be specified. The address can be specified as a domain name or IP address, and an optional port: proxy_pass http://localhost:8000/uri/; or as a UNIX-domain socket path specified after the word “unix” and enclosed in colons: proxy_pass http://unix:/tmp/backend.socket:/uri/;

proxy_pass_s :: [Text] -> Proxy_Pass

Safe version of proxy_pass

newtype Proxy_Pass

Constructors

Proxy_Pass ConfStatement

proxy_pass_header :: [Text] -> ConfStatement

Permits passing otherwise disabled header fields from a proxied server to a client.

proxy_pass_header_s :: [Text] -> Proxy_Pass_Header

Safe version of proxy_pass_header

newtype Proxy_Pass_Header

Constructors

Proxy_Pass_Header ConfStatement

proxy_pass_request_body :: [Text] -> ConfStatement

Indicates whether the original request body is passed to the proxied server. location x-accel-redirect-here { proxy_method GET; proxy_pass_request_body off; proxy_set_header Content-Length ""; proxy_pass ... } See also the and directives.

proxy_pass_request_body_s :: [Text] -> Proxy_Pass_Request_Body

Safe version of proxy_pass_request_body

newtype Proxy_Pass_Request_Body

Constructors

Proxy_Pass_Request_Body ConfStatement

proxy_pass_request_headers :: [Text] -> ConfStatement

Indicates whether the header fields of the original request are passed to the proxied server. location x-accel-redirect-here { proxy_method GET; proxy_pass_request_headers off; proxy_pass_request_body off; proxy_pass ... } See also the and directives.

proxy_pass_request_headers_s :: [Text] -> Proxy_Pass_Request_Headers

Safe version of proxy_pass_request_headers

newtype Proxy_Pass_Request_Headers

Constructors

Proxy_Pass_Request_Headers ConfStatement

proxy_read_timeout :: [Text] -> ConfStatement

Defines a timeout for reading a response from the proxied server. The timeout is set only between two successive read operations, not for the transmission of the whole response. If the proxied server does not transmit anything within this time, the connection is closed.

proxy_read_timeout_s :: [Text] -> Proxy_Read_Timeout

Safe version of proxy_read_timeout

newtype Proxy_Read_Timeout

Constructors

Proxy_Read_Timeout ConfStatement

proxy_redirect :: [Text] -> ConfStatement

Sets the text that should be changed in the Location and Refresh header fields of a proxied server response. Suppose a proxied server returned the header field “Location: http://localhost:8000/two/some/uri/”. The directive proxy_redirect http://localhost:8000/two/ http://frontend/one/; will rewrite this string to “Location: http://frontend/one/some/uri/”.

proxy_redirect_s :: [Text] -> Proxy_Redirect

Safe version of proxy_redirect

newtype Proxy_Redirect

Constructors

Proxy_Redirect ConfStatement

proxy_request_buffering :: [Text] -> ConfStatement

Enables or disables buffering of a client request body.

proxy_request_buffering_s :: [Text] -> Proxy_Request_Buffering

Safe version of proxy_request_buffering

newtype Proxy_Request_Buffering

Constructors

Proxy_Request_Buffering ConfStatement

proxy_send_lowat :: [Text] -> ConfStatement

If the directive is set to a non-zero value, nginx will try to minimize the number of send operations on outgoing connections to a proxied server by using either NOTE_LOWAT flag of the method, or the SO_SNDLOWAT socket option, with the specified size.

proxy_send_lowat_s :: [Text] -> Proxy_Send_Lowat

Safe version of proxy_send_lowat

newtype Proxy_Send_Lowat

Constructors

Proxy_Send_Lowat ConfStatement

proxy_send_timeout :: [Text] -> ConfStatement

Sets a timeout for transmitting a request to the proxied server. The timeout is set only between two successive write operations, not for the transmission of the whole request. If the proxied server does not receive anything within this time, the connection is closed.

proxy_send_timeout_s :: [Text] -> Proxy_Send_Timeout

Safe version of proxy_send_timeout

newtype Proxy_Send_Timeout

Constructors

Proxy_Send_Timeout ConfStatement

proxy_set_body :: [Text] -> ConfStatement

Allows redefining the request body passed to the proxied server. The value can contain text, variables, and their combination.

proxy_set_body_s :: [Text] -> Proxy_Set_Body

Safe version of proxy_set_body

newtype Proxy_Set_Body

Constructors

Proxy_Set_Body ConfStatement

proxy_set_header :: [Text] -> ConfStatement

Allows redefining or appending fields to the request header passed to the proxied server. The value can contain text, variables, and their combinations. These directives are inherited from the previous level if and only if there are no proxy_set_header directives defined on the current level. By default, only two fields are redefined: proxy_set_header Host $proxy_host; proxy_set_header Connection close; If caching is enabled, the header fields If-Modified-Since, If-Unmodified-Since, If-None-Match, If-Match, Range, and If-Range from the original request are not passed to the proxied server.

proxy_set_header_s :: [Text] -> Proxy_Set_Header

Safe version of proxy_set_header

newtype Proxy_Set_Header

Constructors

Proxy_Set_Header ConfStatement

proxy_ssl_certificate :: [Text] -> ConfStatement

Specifies a file with the certificate in the PEM format used for authentication to a proxied HTTPS server.

proxy_ssl_certificate_s :: [Text] -> Proxy_Ssl_Certificate

Safe version of proxy_ssl_certificate

newtype Proxy_Ssl_Certificate

Constructors

Proxy_Ssl_Certificate ConfStatement

proxy_ssl_certificate_key :: [Text] -> ConfStatement

Specifies a file with the secret key in the PEM format used for authentication to a proxied HTTPS server.

proxy_ssl_certificate_key_s :: [Text] -> Proxy_Ssl_Certificate_Key

Safe version of proxy_ssl_certificate_key

newtype Proxy_Ssl_Certificate_Key

Constructors

Proxy_Ssl_Certificate_Key ConfStatement

proxy_ssl_ciphers :: [Text] -> ConfStatement

Specifies the enabled ciphers for requests to a proxied HTTPS server. The ciphers are specified in the format understood by the OpenSSL library.

proxy_ssl_ciphers_s :: [Text] -> Proxy_Ssl_Ciphers

Safe version of proxy_ssl_ciphers

newtype Proxy_Ssl_Ciphers

Constructors

Proxy_Ssl_Ciphers ConfStatement

proxy_ssl_crl :: [Text] -> ConfStatement

Specifies a file with revoked certificates (CRL) in the PEM format used to verify the certificate of the proxied HTTPS server.

proxy_ssl_crl_s :: [Text] -> Proxy_Ssl_Crl

Safe version of proxy_ssl_crl

newtype Proxy_Ssl_Crl

Constructors

Proxy_Ssl_Crl ConfStatement

proxy_ssl_name :: [Text] -> ConfStatement

Allows overriding the server name used to verify the certificate of the proxied HTTPS server and to be passed through SNI when establishing a connection with the proxied HTTPS server.

proxy_ssl_name_s :: [Text] -> Proxy_Ssl_Name

Safe version of proxy_ssl_name

newtype Proxy_Ssl_Name

Constructors

Proxy_Ssl_Name ConfStatement

proxy_ssl_password_file :: [Text] -> ConfStatement

Specifies a file with passphrases for secret keys where each passphrase is specified on a separate line. Passphrases are tried in turn when loading the key.

proxy_ssl_password_file_s :: [Text] -> Proxy_Ssl_Password_File

Safe version of proxy_ssl_password_file

newtype Proxy_Ssl_Password_File

Constructors

Proxy_Ssl_Password_File ConfStatement

proxy_ssl_server_name :: [Text] -> ConfStatement

Enables or disables passing of the server name through TLS Server Name Indication extension (SNI, RFC 6066) when establishing a connection with the proxied HTTPS server.

proxy_ssl_server_name_s :: [Text] -> Proxy_Ssl_Server_Name

Safe version of proxy_ssl_server_name

newtype Proxy_Ssl_Server_Name

Constructors

Proxy_Ssl_Server_Name ConfStatement

proxy_ssl_session_reuse :: [Text] -> ConfStatement

Determines whether SSL sessions can be reused when working with the proxied server. If the errors “SSL3_GET_FINISHED:digest check failed” appear in the logs, try disabling session reuse.

proxy_ssl_session_reuse_s :: [Text] -> Proxy_Ssl_Session_Reuse

Safe version of proxy_ssl_session_reuse

newtype Proxy_Ssl_Session_Reuse

Constructors

Proxy_Ssl_Session_Reuse ConfStatement

proxy_ssl_protocols :: [Text] -> ConfStatement

Enables the specified protocols for requests to a proxied HTTPS server.

proxy_ssl_protocols_s :: [Text] -> Proxy_Ssl_Protocols

Safe version of proxy_ssl_protocols

newtype Proxy_Ssl_Protocols

Constructors

Proxy_Ssl_Protocols ConfStatement

proxy_ssl_trusted_certificate :: [Text] -> ConfStatement

Specifies a file with trusted CA certificates in the PEM format used to verify the certificate of the proxied HTTPS server.

proxy_ssl_trusted_certificate_s :: [Text] -> Proxy_Ssl_Trusted_Certificate

Safe version of proxy_ssl_trusted_certificate

newtype Proxy_Ssl_Trusted_Certificate

Constructors

Proxy_Ssl_Trusted_Certificate ConfStatement

proxy_ssl_verify :: [Text] -> ConfStatement

Enables or disables verification of the proxied HTTPS server certificate.

proxy_ssl_verify_s :: [Text] -> Proxy_Ssl_Verify

Safe version of proxy_ssl_verify

newtype Proxy_Ssl_Verify

Constructors

Proxy_Ssl_Verify ConfStatement

proxy_ssl_verify_depth :: [Text] -> ConfStatement

Sets the verification depth in the proxied HTTPS server certificates chain.

proxy_ssl_verify_depth_s :: [Text] -> Proxy_Ssl_Verify_Depth

Safe version of proxy_ssl_verify_depth

newtype Proxy_Ssl_Verify_Depth

Constructors

Proxy_Ssl_Verify_Depth ConfStatement

proxy_store :: [Text] -> ConfStatement

Enables saving of files to a disk. The on parameter saves files with paths corresponding to the directives or . The off parameter disables saving of files. In addition, the file name can be set explicitly using the string with variables: proxy_store datawww$original_uri;

proxy_store_s :: [Text] -> Proxy_Store

Safe version of proxy_store

newtype Proxy_Store

Constructors

Proxy_Store ConfStatement

proxy_store_access :: [Text] -> ConfStatement

Sets access permissions for newly created files and directories, e.g.: proxy_store_access user:rw group:rw all:r;

proxy_store_access_s :: [Text] -> Proxy_Store_Access

Safe version of proxy_store_access

newtype Proxy_Store_Access

Constructors

Proxy_Store_Access ConfStatement

proxy_temp_file_write_size :: [Text] -> ConfStatement

Limits the size of data written to a temporary file at a time, when buffering of responses from the proxied server to temporary files is enabled. By default, size is limited by two buffers set by the and directives. The maximum size of a temporary file is set by the directive.

proxy_temp_file_write_size_s :: [Text] -> Proxy_Temp_File_Write_Size

Safe version of proxy_temp_file_write_size

newtype Proxy_Temp_File_Write_Size

Constructors

Proxy_Temp_File_Write_Size ConfStatement

proxy_temp_path :: [Text] -> ConfStatement

Defines a directory for storing temporary files with data received from proxied servers. Up to three-level subdirectory hierarchy can be used underneath the specified directory. For example, in the following configuration proxy_temp_path spoolnginxproxy_temp 1 2; a temporary file might look like this: spoolnginxproxy_temp745/00000123457

proxy_temp_path_s :: [Text] -> Proxy_Temp_Path

Safe version of proxy_temp_path

newtype Proxy_Temp_Path

Constructors

Proxy_Temp_Path ConfStatement