Nginx_Ngx_Http_Realip_Module

Documentation

set_real_ip_from :: [Text] -> ConfStatement

Defines trusted addresses that are known to send correct replacement addresses. If the special value unix: is specified, all UNIX-domain sockets will be trusted. IPv6 addresses are supported starting from versions 1.3.0 and 1.2.1.

set_real_ip_from_s :: [Text] -> Set_Real_Ip_From

Safe version of set_real_ip_from

newtype Set_Real_Ip_From

Constructors

Set_Real_Ip_From ConfStatement

real_ip_header :: [Text] -> ConfStatement

Defines the request header field whose value will be used to replace the client address.

real_ip_header_s :: [Text] -> Real_Ip_Header

Safe version of real_ip_header

newtype Real_Ip_Header

Constructors

Real_Ip_Header ConfStatement

real_ip_recursive :: [Text] -> ConfStatement

If recursive search is disabled, the original client address that matches one of the trusted addresses is replaced by the last address sent in the request header field defined by the directive. If recursive search is enabled, the original client address that matches one of the trusted addresses is replaced by the last non-trusted address sent in the request header field.

real_ip_recursive_s :: [Text] -> Real_Ip_Recursive

Safe version of real_ip_recursive

newtype Real_Ip_Recursive

Constructors

Real_Ip_Recursive ConfStatement