Nginx_Ngx_Http_Referer_Module

Documentation

referer_hash_bucket_size :: [Text] -> ConfStatement

Sets the bucket size for the valid referers hash tables. The details of setting up hash tables are provided in a separate document.

referer_hash_bucket_size_s :: [Text] -> Referer_Hash_Bucket_Size

Safe version of referer_hash_bucket_size

newtype Referer_Hash_Bucket_Size

Constructors

Referer_Hash_Bucket_Size ConfStatement

referer_hash_max_size :: [Text] -> ConfStatement

Sets the maximum size of the valid referers hash tables. The details of setting up hash tables are provided in a separate document.

referer_hash_max_size_s :: [Text] -> Referer_Hash_Max_Size

Safe version of referer_hash_max_size

newtype Referer_Hash_Max_Size

Constructors

Referer_Hash_Max_Size ConfStatement

valid_referers :: [Text] -> ConfStatement

Specifies the Referer request header field values that will cause the embedded $invalid_referer variable to be set to an empty string. Otherwise, the variable will be set to “1”. Search for a match is case-insensitive.

valid_referers_s :: [Text] -> Valid_Referers

Safe version of valid_referers

newtype Valid_Referers

Constructors

Valid_Referers ConfStatement