Nginx_Ngx_Http_Rewrite_Module

Documentation

break :: [Text] -> ConfStatement

Stops processing the current set of ngx_http_rewrite_module directives.

break_s :: [Text] -> Break

Safe version of break

newtype Break

Constructors

Break ConfStatement

if_ :: [Text] -> [ConfStatement] -> ConfStatement

The specified condition is evaluated. If true, this module directives specified inside the braces are executed, and the request is assigned the configuration inside the if directive. Configurations inside the if directives are inherited from the previous configuration level.

if_s :: (HasContextIf f, ToConfStatement f) => [Text] -> [f] -> ConfStatement

Safe version of 'if'

return :: [Text] -> ConfStatement

Stops processing and returns the specified code to a client. The non-standard code 444 closes a connection without sending a response header.

return_s :: [Text] -> Return

Safe version of return

newtype Return

Constructors

Return ConfStatement

rewrite :: [Text] -> ConfStatement

If the specified regular expression matches a request URI, URI is changed as specified in the replacement string. The rewrite directives are executed sequentially in order of their appearance in the configuration file. It is possible to terminate further processing of the directives using flags. If a replacement string starts with “http:/” or “https:/”, the processing stops and the redirect is returned to a client.

rewrite_s :: [Text] -> Rewrite

Safe version of rewrite

newtype Rewrite

Constructors

Rewrite ConfStatement

rewrite_log :: [Text] -> ConfStatement

Enables or disables logging of ngx_http_rewrite_module module directives processing results into the at the notice level.

rewrite_log_s :: [Text] -> Rewrite_Log

Safe version of rewrite_log

newtype Rewrite_Log

Constructors

Rewrite_Log ConfStatement

set :: [Text] -> ConfStatement

Sets a value for the specified variable. The value can contain text, variables, and their combination.

set_s :: [Text] -> Set

Safe version of set

newtype Set

Constructors

Set ConfStatement

uninitialized_variable_warn :: [Text] -> ConfStatement

Controls whether warnings about uninitialized variables are logged.

uninitialized_variable_warn_s :: [Text] -> Uninitialized_Variable_Warn

Safe version of uninitialized_variable_warn

newtype Uninitialized_Variable_Warn

Constructors

Uninitialized_Variable_Warn ConfStatement