Nginx_Ngx_Http_Scgi_Module

Documentation

scgi_bind :: [Text] -> ConfStatement

Makes outgoing connections to an SCGI server originate from the specified local IP address with an optional port (1.11.2). Parameter value can contain variables (1.3.12). The special value off (1.3.12) cancels the effect of the scgi_bind directive inherited from the previous configuration level, which allows the system to auto-assign the local IP address and port.

scgi_bind_s :: [Text] -> Scgi_Bind

Safe version of scgi_bind

newtype Scgi_Bind

Constructors

Scgi_Bind ConfStatement

scgi_buffer_size :: [Text] -> ConfStatement

Sets the size of the buffer used for reading the first part of the response received from the SCGI server. This part usually contains a small response header. By default, the buffer size is equal to one memory page. This is either 4K or 8K, depending on a platform. It can be made smaller, however.

scgi_buffer_size_s :: [Text] -> Scgi_Buffer_Size

Safe version of scgi_buffer_size

newtype Scgi_Buffer_Size

Constructors

Scgi_Buffer_Size ConfStatement

scgi_buffering :: [Text] -> ConfStatement

Enables or disables buffering of responses from the SCGI server.

scgi_buffering_s :: [Text] -> Scgi_Buffering

Safe version of scgi_buffering

newtype Scgi_Buffering

Constructors

Scgi_Buffering ConfStatement

scgi_buffers :: [Text] -> ConfStatement

Sets the number and size of the buffers used for reading a response from the SCGI server, for a single connection. By default, the buffer size is equal to one memory page. This is either 4K or 8K, depending on a platform.

scgi_buffers_s :: [Text] -> Scgi_Buffers

Safe version of scgi_buffers

newtype Scgi_Buffers

Constructors

Scgi_Buffers ConfStatement

scgi_busy_buffers_size :: [Text] -> ConfStatement

When buffering of responses from the SCGI server is enabled, limits the total size of buffers that can be busy sending a response to the client while the response is not yet fully read. In the meantime, the rest of the buffers can be used for reading the response and, if needed, buffering part of the response to a temporary file. By default, size is limited by the size of two buffers set by the and directives.

scgi_busy_buffers_size_s :: [Text] -> Scgi_Busy_Buffers_Size

Safe version of scgi_busy_buffers_size

newtype Scgi_Busy_Buffers_Size

Constructors

Scgi_Busy_Buffers_Size ConfStatement

scgi_cache :: [Text] -> ConfStatement

Defines a shared memory zone used for caching. The same zone can be used in several places. Parameter value can contain variables (1.7.9). The off parameter disables caching inherited from the previous configuration level.

scgi_cache_s :: [Text] -> Scgi_Cache

Safe version of scgi_cache

newtype Scgi_Cache

Constructors

Scgi_Cache ConfStatement

scgi_cache_bypass :: [Text] -> ConfStatement

Defines conditions under which the response will not be taken from a cache. If at least one value of the string parameters is not empty and is not equal to “0” then the response will not be taken from the cache: scgi_cache_bypass $cookie_nocache $arg_nocache$arg_comment; scgi_cache_bypass $http_pragma $http_authorization; Can be used along with the directive.

scgi_cache_bypass_s :: [Text] -> Scgi_Cache_Bypass

Safe version of scgi_cache_bypass

newtype Scgi_Cache_Bypass

Constructors

Scgi_Cache_Bypass ConfStatement

scgi_cache_key :: [Text] -> ConfStatement

Defines a key for caching, for example scgi_cache_key localhost:9000$request_uri;

scgi_cache_key_s :: [Text] -> Scgi_Cache_Key

Safe version of scgi_cache_key

newtype Scgi_Cache_Key

Constructors

Scgi_Cache_Key ConfStatement

scgi_cache_lock :: [Text] -> ConfStatement

When enabled, only one request at a time will be allowed to populate a new cache element identified according to the directive by passing a request to an SCGI server. Other requests of the same cache element will either wait for a response to appear in the cache or the cache lock for this element to be released, up to the time set by the directive.

scgi_cache_lock_s :: [Text] -> Scgi_Cache_Lock

Safe version of scgi_cache_lock

newtype Scgi_Cache_Lock

Constructors

Scgi_Cache_Lock ConfStatement

scgi_cache_lock_age :: [Text] -> ConfStatement

If the last request passed to the SCGI server for populating a new cache element has not completed for the specified time, one more request may be passed to the SCGI server.

scgi_cache_lock_age_s :: [Text] -> Scgi_Cache_Lock_Age

Safe version of scgi_cache_lock_age

newtype Scgi_Cache_Lock_Age

Constructors

Scgi_Cache_Lock_Age ConfStatement

scgi_cache_lock_timeout :: [Text] -> ConfStatement

Sets a timeout for . When the time expires, the request will be passed to the SCGI server, however, the response will not be cached. Before 1.7.8, the response could be cached.

scgi_cache_lock_timeout_s :: [Text] -> Scgi_Cache_Lock_Timeout

Safe version of scgi_cache_lock_timeout

newtype Scgi_Cache_Lock_Timeout

Constructors

Scgi_Cache_Lock_Timeout ConfStatement

scgi_cache_methods :: [Text] -> ConfStatement

If the client request method is listed in this directive then the response will be cached. “GET” and “HEAD” methods are always added to the list, though it is recommended to specify them explicitly. See also the directive.

scgi_cache_methods_s :: [Text] -> Scgi_Cache_Methods

Safe version of scgi_cache_methods

newtype Scgi_Cache_Methods

Constructors

Scgi_Cache_Methods ConfStatement

scgi_cache_min_uses :: [Text] -> ConfStatement

Sets the number of requests after which the response will be cached.

scgi_cache_min_uses_s :: [Text] -> Scgi_Cache_Min_Uses

Safe version of scgi_cache_min_uses

newtype Scgi_Cache_Min_Uses

Constructors

Scgi_Cache_Min_Uses ConfStatement

scgi_cache_path :: [Text] -> ConfStatement

Sets the path and other parameters of a cache. Cache data are stored in files. The file name in a cache is a result of applying the MD5 function to the cache key. The levels parameter defines hierarchy levels of a cache. For example, in the following configuration scgi_cache_path datanginxcache levels=1:2 keys_zone=one:10m; file names in a cache will look like this: datanginxcachec29/b7f54b2df7773722d382f4809d65029c

scgi_cache_path_s :: [Text] -> Scgi_Cache_Path

Safe version of scgi_cache_path

newtype Scgi_Cache_Path

Constructors

Scgi_Cache_Path ConfStatement

scgi_cache_purge :: [Text] -> ConfStatement

Defines conditions under which the request will be considered a cache purge request. If at least one value of the string parameters is not empty and is not equal to “0” then the cache entry with a corresponding cache key is removed. The result of successful operation is indicated by returning the response.

scgi_cache_purge_s :: [Text] -> Scgi_Cache_Purge

Safe version of scgi_cache_purge

newtype Scgi_Cache_Purge

Constructors

Scgi_Cache_Purge ConfStatement

scgi_cache_revalidate :: [Text] -> ConfStatement

Enables revalidation of expired cache items using conditional requests with the If-Modified-Since and If-None-Match header fields.

scgi_cache_revalidate_s :: [Text] -> Scgi_Cache_Revalidate

Safe version of scgi_cache_revalidate

newtype Scgi_Cache_Revalidate

Constructors

Scgi_Cache_Revalidate ConfStatement

scgi_cache_use_stale :: [Text] -> ConfStatement

Determines in which cases a stale cached response can be used when an error occurs during communication with the SCGI server. The directive’s parameters match the parameters of the directive.

scgi_cache_use_stale_s :: [Text] -> Scgi_Cache_Use_Stale

Safe version of scgi_cache_use_stale

newtype Scgi_Cache_Use_Stale

Constructors

Scgi_Cache_Use_Stale ConfStatement

scgi_cache_valid :: [Text] -> ConfStatement

Sets caching time for different response codes. For example, the following directives scgi_cache_valid 200 302 10m; scgi_cache_valid 404 1m; set 10 minutes of caching for responses with codes 200 and 302 and 1 minute for responses with code 404.

scgi_cache_valid_s :: [Text] -> Scgi_Cache_Valid

Safe version of scgi_cache_valid

newtype Scgi_Cache_Valid

Constructors

Scgi_Cache_Valid ConfStatement

scgi_connect_timeout :: [Text] -> ConfStatement

Defines a timeout for establishing a connection with an SCGI server. It should be noted that this timeout cannot usually exceed 75 seconds.

scgi_connect_timeout_s :: [Text] -> Scgi_Connect_Timeout

Safe version of scgi_connect_timeout

newtype Scgi_Connect_Timeout

Constructors

Scgi_Connect_Timeout ConfStatement

scgi_force_ranges :: [Text] -> ConfStatement

Enables byte-range support for both cached and uncached responses from the SCGI server regardless of the Accept-Ranges field in these responses.

scgi_force_ranges_s :: [Text] -> Scgi_Force_Ranges

Safe version of scgi_force_ranges

newtype Scgi_Force_Ranges

Constructors

Scgi_Force_Ranges ConfStatement

scgi_hide_header :: [Text] -> ConfStatement

By default, nginx does not pass the header fields Status and X-Accel-... from the response of an SCGI server to a client. The scgi_hide_header directive sets additional fields that will not be passed. If, on the contrary, the passing of fields needs to be permitted, the directive can be used.

scgi_hide_header_s :: [Text] -> Scgi_Hide_Header

Safe version of scgi_hide_header

newtype Scgi_Hide_Header

Constructors

Scgi_Hide_Header ConfStatement

scgi_ignore_client_abort :: [Text] -> ConfStatement

Determines whether the connection with an SCGI server should be closed when a client closes the connection without waiting for a response.

scgi_ignore_client_abort_s :: [Text] -> Scgi_Ignore_Client_Abort

Safe version of scgi_ignore_client_abort

newtype Scgi_Ignore_Client_Abort

Constructors

Scgi_Ignore_Client_Abort ConfStatement

scgi_ignore_headers :: [Text] -> ConfStatement

Disables processing of certain response header fields from the SCGI server. The following fields can be ignored: X-Accel-Redirect, X-Accel-Expires, X-Accel-Limit-Rate (1.1.6), X-Accel-Buffering (1.1.6), X-Accel-Charset (1.1.6), Expires, Cache-Control, Set-Cookie (0.8.44), and Vary (1.7.7).

scgi_ignore_headers_s :: [Text] -> Scgi_Ignore_Headers

Safe version of scgi_ignore_headers

newtype Scgi_Ignore_Headers

Constructors

Scgi_Ignore_Headers ConfStatement

scgi_intercept_errors :: [Text] -> ConfStatement

Determines whether an SCGI server responses with codes greater than or equal to 300 should be passed to a client or be redirected to nginx for processing with the directive.

scgi_intercept_errors_s :: [Text] -> Scgi_Intercept_Errors

Safe version of scgi_intercept_errors

newtype Scgi_Intercept_Errors

Constructors

Scgi_Intercept_Errors ConfStatement

scgi_limit_rate :: [Text] -> ConfStatement

Limits the speed of reading the response from the SCGI server. The rate is specified in bytes per second. The zero value disables rate limiting. The limit is set per a request, and so if nginx simultaneously opens two connections to the SCGI server, the overall rate will be twice as much as the specified limit. The limitation works only if buffering of responses from the SCGI server is enabled.

scgi_limit_rate_s :: [Text] -> Scgi_Limit_Rate

Safe version of scgi_limit_rate

newtype Scgi_Limit_Rate

Constructors

Scgi_Limit_Rate ConfStatement

scgi_max_temp_file_size :: [Text] -> ConfStatement

When buffering of responses from the SCGI server is enabled, and the whole response does not fit into the buffers set by the and directives, a part of the response can be saved to a temporary file. This directive sets the maximum size of the temporary file. The size of data written to the temporary file at a time is set by the directive.

scgi_max_temp_file_size_s :: [Text] -> Scgi_Max_Temp_File_Size

Safe version of scgi_max_temp_file_size

newtype Scgi_Max_Temp_File_Size

Constructors

Scgi_Max_Temp_File_Size ConfStatement

scgi_next_upstream :: [Text] -> ConfStatement

Specifies in which cases a request should be passed to the next server: error an error occurred while establishing a connection with the server, passing a request to it, or reading the response header; timeout a timeout has occurred while establishing a connection with the server, passing a request to it, or reading the response header; invalid_header a server returned an empty or invalid response; http_500 a server returned a response with the code 500; http_503 a server returned a response with the code 503; http_403 a server returned a response with the code 403; http_404 a server returned a response with the code 404; non_idempotent normally, requests with a non-idempotent method (POST, LOCK, PATCH) are not passed to the next server if a request has been sent to an upstream server (1.9.13); enabling this option explicitly allows retrying such requests; off disables passing a request to the next server.

scgi_next_upstream_s :: [Text] -> Scgi_Next_Upstream

Safe version of scgi_next_upstream

newtype Scgi_Next_Upstream

Constructors

Scgi_Next_Upstream ConfStatement

scgi_next_upstream_timeout :: [Text] -> ConfStatement

Limits the time during which a request can be passed to the next server. The 0 value turns off this limitation.

scgi_next_upstream_timeout_s :: [Text] -> Scgi_Next_Upstream_Timeout

Safe version of scgi_next_upstream_timeout

newtype Scgi_Next_Upstream_Timeout

Constructors

Scgi_Next_Upstream_Timeout ConfStatement

scgi_next_upstream_tries :: [Text] -> ConfStatement

Limits the number of possible tries for passing a request to the next server. The 0 value turns off this limitation.

scgi_next_upstream_tries_s :: [Text] -> Scgi_Next_Upstream_Tries

Safe version of scgi_next_upstream_tries

newtype Scgi_Next_Upstream_Tries

Constructors

Scgi_Next_Upstream_Tries ConfStatement

scgi_no_cache :: [Text] -> ConfStatement

Defines conditions under which the response will not be saved to a cache. If at least one value of the string parameters is not empty and is not equal to “0” then the response will not be saved: scgi_no_cache $cookie_nocache $arg_nocache$arg_comment; scgi_no_cache $http_pragma $http_authorization; Can be used along with the directive.

scgi_no_cache_s :: [Text] -> Scgi_No_Cache

Safe version of scgi_no_cache

newtype Scgi_No_Cache

Constructors

Scgi_No_Cache ConfStatement

scgi_param :: [Text] -> ConfStatement

Sets a parameter that should be passed to the SCGI server. The value can contain text, variables, and their combination. These directives are inherited from the previous level if and only if there are no scgi_param directives defined on the current level.

scgi_param_s :: [Text] -> Scgi_Param

Safe version of scgi_param

newtype Scgi_Param

Constructors

Scgi_Param ConfStatement

scgi_pass :: [Text] -> ConfStatement

Sets the address of an SCGI server. The address can be specified as a domain name or IP address, and a port: scgi_pass localhost:9000; or as a UNIX-domain socket path: scgi_pass unix:tmpscgi.socket;

scgi_pass_s :: [Text] -> Scgi_Pass

Safe version of scgi_pass

newtype Scgi_Pass

Constructors

Scgi_Pass ConfStatement

scgi_pass_header :: [Text] -> ConfStatement

Permits passing otherwise disabled header fields from an SCGI server to a client.

scgi_pass_header_s :: [Text] -> Scgi_Pass_Header

Safe version of scgi_pass_header

newtype Scgi_Pass_Header

Constructors

Scgi_Pass_Header ConfStatement

scgi_pass_request_body :: [Text] -> ConfStatement

Indicates whether the original request body is passed to the SCGI server. See also the directive.

scgi_pass_request_body_s :: [Text] -> Scgi_Pass_Request_Body

Safe version of scgi_pass_request_body

newtype Scgi_Pass_Request_Body

Constructors

Scgi_Pass_Request_Body ConfStatement

scgi_pass_request_headers :: [Text] -> ConfStatement

Indicates whether the header fields of the original request are passed to the SCGI server. See also the directive.

scgi_pass_request_headers_s :: [Text] -> Scgi_Pass_Request_Headers

Safe version of scgi_pass_request_headers

newtype Scgi_Pass_Request_Headers

Constructors

Scgi_Pass_Request_Headers ConfStatement

scgi_read_timeout :: [Text] -> ConfStatement

Defines a timeout for reading a response from the SCGI server. The timeout is set only between two successive read operations, not for the transmission of the whole response. If the SCGI server does not transmit anything within this time, the connection is closed.

scgi_read_timeout_s :: [Text] -> Scgi_Read_Timeout

Safe version of scgi_read_timeout

newtype Scgi_Read_Timeout

Constructors

Scgi_Read_Timeout ConfStatement

scgi_request_buffering :: [Text] -> ConfStatement

Enables or disables buffering of a client request body.

scgi_request_buffering_s :: [Text] -> Scgi_Request_Buffering

Safe version of scgi_request_buffering

newtype Scgi_Request_Buffering

Constructors

Scgi_Request_Buffering ConfStatement

scgi_send_timeout :: [Text] -> ConfStatement

Sets a timeout for transmitting a request to the SCGI server. The timeout is set only between two successive write operations, not for the transmission of the whole request. If the SCGI server does not receive anything within this time, the connection is closed.

scgi_send_timeout_s :: [Text] -> Scgi_Send_Timeout

Safe version of scgi_send_timeout

newtype Scgi_Send_Timeout

Constructors

Scgi_Send_Timeout ConfStatement

scgi_store :: [Text] -> ConfStatement

Enables saving of files to a disk. The on parameter saves files with paths corresponding to the directives or . The off parameter disables saving of files. In addition, the file name can be set explicitly using the string with variables: scgi_store datawww$original_uri;

scgi_store_s :: [Text] -> Scgi_Store

Safe version of scgi_store

newtype Scgi_Store

Constructors

Scgi_Store ConfStatement

scgi_store_access :: [Text] -> ConfStatement

Sets access permissions for newly created files and directories, e.g.: scgi_store_access user:rw group:rw all:r;

scgi_store_access_s :: [Text] -> Scgi_Store_Access

Safe version of scgi_store_access

newtype Scgi_Store_Access

Constructors

Scgi_Store_Access ConfStatement

scgi_temp_file_write_size :: [Text] -> ConfStatement

Limits the size of data written to a temporary file at a time, when buffering of responses from the SCGI server to temporary files is enabled. By default, size is limited by two buffers set by the and directives. The maximum size of a temporary file is set by the directive.

scgi_temp_file_write_size_s :: [Text] -> Scgi_Temp_File_Write_Size

Safe version of scgi_temp_file_write_size

newtype Scgi_Temp_File_Write_Size

Constructors

Scgi_Temp_File_Write_Size ConfStatement

scgi_temp_path :: [Text] -> ConfStatement

Defines a directory for storing temporary files with data received from SCGI servers. Up to three-level subdirectory hierarchy can be used underneath the specified directory. For example, in the following configuration scgi_temp_path spoolnginxscgi_temp 1 2; a temporary file might look like this: spoolnginxscgi_temp745/00000123457

scgi_temp_path_s :: [Text] -> Scgi_Temp_Path

Safe version of scgi_temp_path

newtype Scgi_Temp_Path

Constructors

Scgi_Temp_Path ConfStatement