Nginx_Ngx_Http_Secure_Link_Module

Documentation

secure_link :: [Text] -> ConfStatement

Defines a string with variables from which the checksum value and lifetime of a link will be extracted.

secure_link_s :: [Text] -> Secure_Link

Safe version of secure_link

newtype Secure_Link

Constructors

Secure_Link ConfStatement

secure_link_md5 :: [Text] -> ConfStatement

Defines an expression for which the MD5 hash value will be computed and compared with the value passed in a request.

secure_link_md5_s :: [Text] -> Secure_Link_Md5

Safe version of secure_link_md5

newtype Secure_Link_Md5

Constructors

Secure_Link_Md5 ConfStatement

secure_link_secret :: [Text] -> ConfStatement

Defines a secret word used to check authenticity of requested links.

secure_link_secret_s :: [Text] -> Secure_Link_Secret

Safe version of secure_link_secret

newtype Secure_Link_Secret

Constructors

Secure_Link_Secret ConfStatement