Nginx_Ngx_Http_Session_Log_Module

Documentation

session_log_format :: [Text] -> ConfStatement

Specifies the output format of a log. The value of the $body_bytes_sent variable is aggregated across all requests in a session. The values of all other variables available for logging correspond to the first request in a session.

session_log_format_s :: [Text] -> Session_Log_Format

Safe version of session_log_format

newtype Session_Log_Format

Constructors

Session_Log_Format ConfStatement

session_log_zone :: [Text] -> ConfStatement

Sets the path to a log file and configures the shared memory zone that is used to store currently active sessions.

session_log_zone_s :: [Text] -> Session_Log_Zone

Safe version of session_log_zone

newtype Session_Log_Zone

Constructors

Session_Log_Zone ConfStatement

session_log :: [Text] -> ConfStatement

Enables the use of the specified session log. The special value off cancels all session_log directives inherited from the previous configuration level.

session_log_s :: [Text] -> Session_Log

Safe version of session_log

newtype Session_Log

Constructors

Session_Log ConfStatement