Nginx_Ngx_Http_Spdy_Module

Documentation

spdy_chunk_size :: [Text] -> ConfStatement

Sets the maximum size of chunks into which the response body is sliced. A too low value results in higher overhead. A too high value impairs prioritization due to HOL blocking.

spdy_chunk_size_s :: [Text] -> Spdy_Chunk_Size

Safe version of spdy_chunk_size

newtype Spdy_Chunk_Size

Constructors

Spdy_Chunk_Size ConfStatement

spdy_headers_comp :: [Text] -> ConfStatement

Sets the header compression level of a response in a range from 1 (fastest, less compression) to 9 (slowest, best compression). The special value 0 turns off the header compression.

spdy_headers_comp_s :: [Text] -> Spdy_Headers_Comp

Safe version of spdy_headers_comp

newtype Spdy_Headers_Comp

Constructors

Spdy_Headers_Comp ConfStatement