Nginx_Ngx_Http_Ssi_Module

Documentation

ssi :: [Text] -> ConfStatement

Enables or disables processing of SSI commands in responses.

ssi_s :: [Text] -> Ssi

Safe version of ssi

newtype Ssi

Constructors

Ssi ConfStatement

ssi_last_modified :: [Text] -> ConfStatement

Allows preserving the Last-Modified header field from the original response during SSI processing to facilitate response caching.

ssi_last_modified_s :: [Text] -> Ssi_Last_Modified

Safe version of ssi_last_modified

newtype Ssi_Last_Modified

Constructors

Ssi_Last_Modified ConfStatement

ssi_min_file_chunk :: [Text] -> ConfStatement

Sets the minimum size for parts of a response stored on disk, starting from which it makes sense to send them using .

ssi_min_file_chunk_s :: [Text] -> Ssi_Min_File_Chunk

Safe version of ssi_min_file_chunk

newtype Ssi_Min_File_Chunk

Constructors

Ssi_Min_File_Chunk ConfStatement

ssi_silent_errors :: [Text] -> ConfStatement

If enabled, suppresses the output of the “[an error occurred while processing the directive]” string if an error occurred during SSI processing.

ssi_silent_errors_s :: [Text] -> Ssi_Silent_Errors

Safe version of ssi_silent_errors

newtype Ssi_Silent_Errors

Constructors

Ssi_Silent_Errors ConfStatement

ssi_types :: [Text] -> ConfStatement

Enables processing of SSI commands in responses with the specified MIME types in addition to “text/html”. The special value “*” matches any MIME type (0.8.29).

ssi_types_s :: [Text] -> Ssi_Types

Safe version of ssi_types

newtype Ssi_Types

Constructors

Ssi_Types ConfStatement

ssi_value_length :: [Text] -> ConfStatement

Sets the maximum length of parameter values in SSI commands.

ssi_value_length_s :: [Text] -> Ssi_Value_Length

Safe version of ssi_value_length

newtype Ssi_Value_Length

Constructors

Ssi_Value_Length ConfStatement