Nginx_Ngx_Http_Status_Module

Documentation

status :: [Text] -> ConfStatement

The status information will be accessible from the surrounding location. Access to this location should be limited.

status_s :: [Text] -> Status

Safe version of status

newtype Status

Constructors

Status ConfStatement

status_format :: [Text] -> ConfStatement

By default, status information is output in the JSON format.

status_format_s :: [Text] -> Status_Format

Safe version of status_format

newtype Status_Format

Constructors

Status_Format ConfStatement

status_zone :: [Text] -> ConfStatement

Enables collection of virtual http or stream (1.7.11) server status information in the specified zone. Several servers may share the same zone.

status_zone_s :: [Text] -> Status_Zone

Safe version of status_zone

newtype Status_Zone

Constructors

Status_Zone ConfStatement