Nginx_Ngx_Http_Sub_Module

Documentation

sub_filter :: [Text] -> ConfStatement

Sets a string to replace and a replacement string. The string to replace is matched ignoring the case. The string to replace (1.9.4) and replacement string can contain variables. Several sub_filter directives can be specified on one configuration level (1.9.4).

sub_filter_s :: [Text] -> Sub_Filter

Safe version of sub_filter

newtype Sub_Filter

Constructors

Sub_Filter ConfStatement

sub_filter_last_modified :: [Text] -> ConfStatement

Allows preserving the Last-Modified header field from the original response during replacement to facilitate response caching.

sub_filter_last_modified_s :: [Text] -> Sub_Filter_Last_Modified

Safe version of sub_filter_last_modified

newtype Sub_Filter_Last_Modified

Constructors

Sub_Filter_Last_Modified ConfStatement

sub_filter_once :: [Text] -> ConfStatement

Indicates whether to look for each string to replace once or repeatedly.

sub_filter_once_s :: [Text] -> Sub_Filter_Once

Safe version of sub_filter_once

newtype Sub_Filter_Once

Constructors

Sub_Filter_Once ConfStatement

sub_filter_types :: [Text] -> ConfStatement

Enables string replacement in responses with the specified MIME types in addition to “text/html”. The special value “*” matches any MIME type (0.8.29).

sub_filter_types_s :: [Text] -> Sub_Filter_Types

Safe version of sub_filter_types

newtype Sub_Filter_Types

Constructors

Sub_Filter_Types ConfStatement