Nginx_Ngx_Http_Upstream_Module

Documentation

upstream :: [Text] -> [ConfStatement] -> ConfStatement

Defines a group of servers. Servers can listen on different ports. In addition, servers listening on TCP and UNIX-domain sockets can be mixed.

upstream_s :: (HasContextUpstream f, ToConfStatement f) => [Text] -> [f] -> ConfStatement

Safe version of upstream

server :: [Text] -> ConfStatement

Defines the address and other parameters of a server. The address can be specified as a domain name or IP address, with an optional port, or as a UNIX-domain socket path specified after the “unix:” prefix. If a port is not specified, the port 80 is used. A domain name that resolves to several IP addresses defines multiple servers at once.

server_s :: [Text] -> Server

Safe version of server

newtype Server

Constructors

Server ConfStatement

zone :: [Text] -> ConfStatement

Defines the name and size of the shared memory zone that keeps the group’s configuration and run-time state that are shared between worker processes. Several groups may share the same zone. In this case, it is enough to specify the size only once.

zone_s :: [Text] -> Zone

Safe version of zone

newtype Zone

Constructors

Zone ConfStatement

state :: [Text] -> ConfStatement

Specifies a file that keeps the state of the dynamically configurable group. The state is currently limited to the list of servers with their parameters. The file is read when parsing the configuration and is updated each time the upstream configuration is changed. Changing the file content directly should be avoided. The directive cannot be used along with the directive.

state_s :: [Text] -> State

Safe version of state

newtype State

Constructors

State ConfStatement

hash :: [Text] -> ConfStatement

Specifies a load balancing method for a server group where the client-server mapping is based on the hashed key value. The key can contain text, variables, and their combinations. Note that adding or removing a server from the group may result in remapping most of the keys to different servers. The method is compatible with the Cache::Memcached Perl library.

hash_s :: [Text] -> Hash

Safe version of hash

newtype Hash

Constructors

Hash ConfStatement

ip_hash :: [Text] -> ConfStatement

Specifies that a group should use a load balancing method where requests are distributed between servers based on client IP addresses. The first three octets of the client IPv4 address, or the entire IPv6 address, are used as a hashing key. The method ensures that requests from the same client will always be passed to the same server except when this server is unavailable. In the latter case client requests will be passed to another server. Most probably, it will always be the same server as well. IPv6 addresses are supported starting from versions 1.3.2 and 1.2.2.

ip_hash_s :: [Text] -> Ip_Hash

Safe version of ip_hash

newtype Ip_Hash

Constructors

Ip_Hash ConfStatement

keepalive :: [Text] -> ConfStatement

Activates the cache for connections to upstream servers.

keepalive_s :: [Text] -> Keepalive

Safe version of keepalive

newtype Keepalive

Constructors

Keepalive ConfStatement

ntlm :: [Text] -> ConfStatement

Allows proxying requests with NTLM Authentication. The upstream connection is bound to the client connection once the client sends a request with the Authorization header field value starting with “Negotiate” or “NTLM”. Further client requests will be proxied through the same upstream connection, keeping the authentication context.

ntlm_s :: [Text] -> Ntlm

Safe version of ntlm

newtype Ntlm

Constructors

Ntlm ConfStatement

least_conn :: [Text] -> ConfStatement

Specifies that a group should use a load balancing method where a request is passed to the server with the least number of active connections, taking into account weights of servers. If there are several such servers, they are tried in turn using a weighted round-robin balancing method.

least_conn_s :: [Text] -> Least_Conn

Safe version of least_conn

newtype Least_Conn

Constructors

Least_Conn ConfStatement

least_time :: [Text] -> ConfStatement

Specifies that a group should use a load balancing method where a request is passed to the server with the least average response time and least number of active connections, taking into account weights of servers. If there are several such servers, they are tried in turn using a weighted round-robin balancing method.

least_time_s :: [Text] -> Least_Time

Safe version of least_time

newtype Least_Time

Constructors

Least_Time ConfStatement

health_check :: [Text] -> ConfStatement

Enables periodic health checks of the servers in a group referenced in the surrounding location.

health_check_s :: [Text] -> Health_Check

Safe version of health_check

newtype Health_Check

Constructors

Health_Check ConfStatement

match :: [Text] -> [ConfStatement] -> ConfStatement

Defines the named test set used to verify responses to health check requests.

match_s :: (HasContextMatch f, ToConfStatement f) => [Text] -> [f] -> ConfStatement

Safe version of match

queue :: [Text] -> ConfStatement

If an upstream server cannot be selected immediately while processing a request, and there are the servers in the group that have reached the limit, the request will be placed into the queue. The directive specifies the maximum number of requests that can be in the queue at the same time. If the queue is filled up, or the server to pass the request to cannot be selected within the time period specified in the timeout parameter, the error will be returned to the client.

queue_s :: [Text] -> Queue

Safe version of queue

newtype Queue

Constructors

Queue ConfStatement

sticky :: [Text] -> ConfStatement

Enables session affinity, which causes requests from the same client to be passed to the same server in a group of servers. Three methods are available: cookie When the cookie method is used, information about the designated server is passed in an HTTP cookie generated by nginx: upstream backend { server backend1.example.com; server backend2.example.com; sticky cookie srv_id expires=1h domain=.example.com path=/; }

sticky_s :: [Text] -> Sticky

Safe version of sticky

newtype Sticky

Constructors

Sticky ConfStatement

sticky_cookie_insert :: [Text] -> ConfStatement

This directive is obsolete since version 1.5.7. An equivalent directive with a new syntax should be used instead: sticky cookie name [expires=time] [domain=domain] [path=path];

sticky_cookie_insert_s :: [Text] -> Sticky_Cookie_Insert

Safe version of sticky_cookie_insert

newtype Sticky_Cookie_Insert

Constructors

Sticky_Cookie_Insert ConfStatement