Nginx_Ngx_Http_Userid_Module

Documentation

userid :: [Text] -> ConfStatement

Enables or disables setting cookies and logging the received cookies: on enables the setting of version 2 cookies and logging of the received cookies; v1 enables the setting of version 1 cookies and logging of the received cookies; log disables the setting of cookies, but enables logging of the received cookies; off disables the setting of cookies and logging of the received cookies.

userid_s :: [Text] -> Userid

Safe version of userid

newtype Userid

Constructors

Userid ConfStatement

userid_domain :: [Text] -> ConfStatement

Defines a domain for which the cookie is set. The none parameter disables setting of a domain for the cookie.

userid_domain_s :: [Text] -> Userid_Domain

Safe version of userid_domain

newtype Userid_Domain

Constructors

Userid_Domain ConfStatement

userid_expires :: [Text] -> ConfStatement

Sets a time during which a browser should keep the cookie. The parameter max will cause the cookie to expire on “31 Dec 2037 23:55:55 GMT”. The parameter off will cause the cookie to expire at the end of a browser session.

userid_expires_s :: [Text] -> Userid_Expires

Safe version of userid_expires

newtype Userid_Expires

Constructors

Userid_Expires ConfStatement

userid_mark :: [Text] -> ConfStatement

If the parameter is not off, enables the cookie marking mechanism and sets the character used as a mark. This mechanism is used to add or change and/or a cookie expiration time while preserving the client identifier. A mark can be any letter of the English alphabet (case-sensitive), digit, or the “=” character.

userid_mark_s :: [Text] -> Userid_Mark

Safe version of userid_mark

newtype Userid_Mark

Constructors

Userid_Mark ConfStatement

userid_name :: [Text] -> ConfStatement

Sets the cookie name.

userid_name_s :: [Text] -> Userid_Name

Safe version of userid_name

newtype Userid_Name

Constructors

Userid_Name ConfStatement

userid_p3p :: [Text] -> ConfStatement

Sets a value for the P3P header field that will be sent along with the cookie. If the directive is set to the special value none, the P3P header will not be sent in a response.

userid_p3p_s :: [Text] -> Userid_P3P

Safe version of userid_p3p

newtype Userid_P3P

Constructors

Userid_P3P ConfStatement

userid_path :: [Text] -> ConfStatement

Defines a path for which the cookie is set.

userid_path_s :: [Text] -> Userid_Path

Safe version of userid_path

newtype Userid_Path

Constructors

Userid_Path ConfStatement

userid_service :: [Text] -> ConfStatement

If identifiers are issued by multiple servers (services), each service should be assigned its own number to ensure that client identifiers are unique. For version 1 cookies, the default value is zero. For version 2 cookies, the default value is the number composed from the last four octets of the server’s IP address.

userid_service_s :: [Text] -> Userid_Service

Safe version of userid_service

newtype Userid_Service

Constructors

Userid_Service ConfStatement