Nginx_Ngx_Http_Uwsgi_Module

Documentation

uwsgi_bind :: [Text] -> ConfStatement

Makes outgoing connections to a uwsgi server originate from the specified local IP address with an optional port (1.11.2). Parameter value can contain variables (1.3.12). The special value off (1.3.12) cancels the effect of the uwsgi_bind directive inherited from the previous configuration level, which allows the system to auto-assign the local IP address and port.

uwsgi_bind_s :: [Text] -> Uwsgi_Bind

Safe version of uwsgi_bind

newtype Uwsgi_Bind

Constructors

Uwsgi_Bind ConfStatement

uwsgi_buffer_size :: [Text] -> ConfStatement

Sets the size of the buffer used for reading the first part of the response received from the uwsgi server. This part usually contains a small response header. By default, the buffer size is equal to one memory page. This is either 4K or 8K, depending on a platform. It can be made smaller, however.

uwsgi_buffer_size_s :: [Text] -> Uwsgi_Buffer_Size

Safe version of uwsgi_buffer_size

newtype Uwsgi_Buffer_Size

Constructors

Uwsgi_Buffer_Size ConfStatement

uwsgi_buffering :: [Text] -> ConfStatement

Enables or disables buffering of responses from the uwsgi server.

uwsgi_buffering_s :: [Text] -> Uwsgi_Buffering

Safe version of uwsgi_buffering

newtype Uwsgi_Buffering

Constructors

Uwsgi_Buffering ConfStatement

uwsgi_buffers :: [Text] -> ConfStatement

Sets the number and size of the buffers used for reading a response from the uwsgi server, for a single connection. By default, the buffer size is equal to one memory page. This is either 4K or 8K, depending on a platform.

uwsgi_buffers_s :: [Text] -> Uwsgi_Buffers

Safe version of uwsgi_buffers

newtype Uwsgi_Buffers

Constructors

Uwsgi_Buffers ConfStatement

uwsgi_busy_buffers_size :: [Text] -> ConfStatement

When buffering of responses from the uwsgi server is enabled, limits the total size of buffers that can be busy sending a response to the client while the response is not yet fully read. In the meantime, the rest of the buffers can be used for reading the response and, if needed, buffering part of the response to a temporary file. By default, size is limited by the size of two buffers set by the and directives.

uwsgi_busy_buffers_size_s :: [Text] -> Uwsgi_Busy_Buffers_Size

Safe version of uwsgi_busy_buffers_size

newtype Uwsgi_Busy_Buffers_Size

Constructors

Uwsgi_Busy_Buffers_Size ConfStatement

uwsgi_cache :: [Text] -> ConfStatement

Defines a shared memory zone used for caching. The same zone can be used in several places. Parameter value can contain variables (1.7.9). The off parameter disables caching inherited from the previous configuration level.

uwsgi_cache_s :: [Text] -> Uwsgi_Cache

Safe version of uwsgi_cache

newtype Uwsgi_Cache

Constructors

Uwsgi_Cache ConfStatement

uwsgi_cache_bypass :: [Text] -> ConfStatement

Defines conditions under which the response will not be taken from a cache. If at least one value of the string parameters is not empty and is not equal to “0” then the response will not be taken from the cache: uwsgi_cache_bypass $cookie_nocache $arg_nocache$arg_comment; uwsgi_cache_bypass $http_pragma $http_authorization; Can be used along with the directive.

uwsgi_cache_bypass_s :: [Text] -> Uwsgi_Cache_Bypass

Safe version of uwsgi_cache_bypass

newtype Uwsgi_Cache_Bypass

Constructors

Uwsgi_Cache_Bypass ConfStatement

uwsgi_cache_key :: [Text] -> ConfStatement

Defines a key for caching, for example uwsgi_cache_key localhost:9000$request_uri;

uwsgi_cache_key_s :: [Text] -> Uwsgi_Cache_Key

Safe version of uwsgi_cache_key

newtype Uwsgi_Cache_Key

Constructors

Uwsgi_Cache_Key ConfStatement

uwsgi_cache_lock :: [Text] -> ConfStatement

When enabled, only one request at a time will be allowed to populate a new cache element identified according to the directive by passing a request to a uwsgi server. Other requests of the same cache element will either wait for a response to appear in the cache or the cache lock for this element to be released, up to the time set by the directive.

uwsgi_cache_lock_s :: [Text] -> Uwsgi_Cache_Lock

Safe version of uwsgi_cache_lock

newtype Uwsgi_Cache_Lock

Constructors

Uwsgi_Cache_Lock ConfStatement

uwsgi_cache_lock_age :: [Text] -> ConfStatement

If the last request passed to the uwsgi server for populating a new cache element has not completed for the specified time, one more request may be passed to the uwsgi server.

uwsgi_cache_lock_age_s :: [Text] -> Uwsgi_Cache_Lock_Age

Safe version of uwsgi_cache_lock_age

newtype Uwsgi_Cache_Lock_Age

Constructors

Uwsgi_Cache_Lock_Age ConfStatement

uwsgi_cache_lock_timeout :: [Text] -> ConfStatement

Sets a timeout for . When the time expires, the request will be passed to the uwsgi server, however, the response will not be cached. Before 1.7.8, the response could be cached.

uwsgi_cache_lock_timeout_s :: [Text] -> Uwsgi_Cache_Lock_Timeout

Safe version of uwsgi_cache_lock_timeout

newtype Uwsgi_Cache_Lock_Timeout

Constructors

Uwsgi_Cache_Lock_Timeout ConfStatement

uwsgi_cache_methods :: [Text] -> ConfStatement

If the client request method is listed in this directive then the response will be cached. “GET” and “HEAD” methods are always added to the list, though it is recommended to specify them explicitly. See also the directive.

uwsgi_cache_methods_s :: [Text] -> Uwsgi_Cache_Methods

Safe version of uwsgi_cache_methods

newtype Uwsgi_Cache_Methods

Constructors

Uwsgi_Cache_Methods ConfStatement

uwsgi_cache_min_uses :: [Text] -> ConfStatement

Sets the number of requests after which the response will be cached.

uwsgi_cache_min_uses_s :: [Text] -> Uwsgi_Cache_Min_Uses

Safe version of uwsgi_cache_min_uses

newtype Uwsgi_Cache_Min_Uses

Constructors

Uwsgi_Cache_Min_Uses ConfStatement

uwsgi_cache_path :: [Text] -> ConfStatement

Sets the path and other parameters of a cache. Cache data are stored in files. The file name in a cache is a result of applying the MD5 function to the cache key. The levels parameter defines hierarchy levels of a cache. For example, in the following configuration uwsgi_cache_path datanginxcache levels=1:2 keys_zone=one:10m; file names in a cache will look like this: datanginxcachec29/b7f54b2df7773722d382f4809d65029c

uwsgi_cache_path_s :: [Text] -> Uwsgi_Cache_Path

Safe version of uwsgi_cache_path

newtype Uwsgi_Cache_Path

Constructors

Uwsgi_Cache_Path ConfStatement

uwsgi_cache_purge :: [Text] -> ConfStatement

Defines conditions under which the request will be considered a cache purge request. If at least one value of the string parameters is not empty and is not equal to “0” then the cache entry with a corresponding cache key is removed. The result of successful operation is indicated by returning the response.

uwsgi_cache_purge_s :: [Text] -> Uwsgi_Cache_Purge

Safe version of uwsgi_cache_purge

newtype Uwsgi_Cache_Purge

Constructors

Uwsgi_Cache_Purge ConfStatement

uwsgi_cache_revalidate :: [Text] -> ConfStatement

Enables revalidation of expired cache items using conditional requests with the If-Modified-Since and If-None-Match header fields.

uwsgi_cache_revalidate_s :: [Text] -> Uwsgi_Cache_Revalidate

Safe version of uwsgi_cache_revalidate

newtype Uwsgi_Cache_Revalidate

Constructors

Uwsgi_Cache_Revalidate ConfStatement

uwsgi_cache_use_stale :: [Text] -> ConfStatement

Determines in which cases a stale cached response can be used when an error occurs during communication with the uwsgi server. The directive’s parameters match the parameters of the directive.

uwsgi_cache_use_stale_s :: [Text] -> Uwsgi_Cache_Use_Stale

Safe version of uwsgi_cache_use_stale

newtype Uwsgi_Cache_Use_Stale

Constructors

Uwsgi_Cache_Use_Stale ConfStatement

uwsgi_cache_valid :: [Text] -> ConfStatement

Sets caching time for different response codes. For example, the following directives uwsgi_cache_valid 200 302 10m; uwsgi_cache_valid 404 1m; set 10 minutes of caching for responses with codes 200 and 302 and 1 minute for responses with code 404.

uwsgi_cache_valid_s :: [Text] -> Uwsgi_Cache_Valid

Safe version of uwsgi_cache_valid

newtype Uwsgi_Cache_Valid

Constructors

Uwsgi_Cache_Valid ConfStatement

uwsgi_connect_timeout :: [Text] -> ConfStatement

Defines a timeout for establishing a connection with a uwsgi server. It should be noted that this timeout cannot usually exceed 75 seconds.

uwsgi_connect_timeout_s :: [Text] -> Uwsgi_Connect_Timeout

Safe version of uwsgi_connect_timeout

newtype Uwsgi_Connect_Timeout

Constructors

Uwsgi_Connect_Timeout ConfStatement

uwsgi_force_ranges :: [Text] -> ConfStatement

Enables byte-range support for both cached and uncached responses from the uwsgi server regardless of the Accept-Ranges field in these responses.

uwsgi_force_ranges_s :: [Text] -> Uwsgi_Force_Ranges

Safe version of uwsgi_force_ranges

newtype Uwsgi_Force_Ranges

Constructors

Uwsgi_Force_Ranges ConfStatement

uwsgi_hide_header :: [Text] -> ConfStatement

By default, nginx does not pass the header fields Status and X-Accel-... from the response of a uwsgi server to a client. The uwsgi_hide_header directive sets additional fields that will not be passed. If, on the contrary, the passing of fields needs to be permitted, the directive can be used.

uwsgi_hide_header_s :: [Text] -> Uwsgi_Hide_Header

Safe version of uwsgi_hide_header

newtype Uwsgi_Hide_Header

Constructors

Uwsgi_Hide_Header ConfStatement

uwsgi_ignore_client_abort :: [Text] -> ConfStatement

Determines whether the connection with a uwsgi server should be closed when a client closes the connection without waiting for a response.

uwsgi_ignore_client_abort_s :: [Text] -> Uwsgi_Ignore_Client_Abort

Safe version of uwsgi_ignore_client_abort

newtype Uwsgi_Ignore_Client_Abort

Constructors

Uwsgi_Ignore_Client_Abort ConfStatement

uwsgi_ignore_headers :: [Text] -> ConfStatement

Disables processing of certain response header fields from the uwsgi server. The following fields can be ignored: X-Accel-Redirect, X-Accel-Expires, X-Accel-Limit-Rate (1.1.6), X-Accel-Buffering (1.1.6), X-Accel-Charset (1.1.6), Expires, Cache-Control, Set-Cookie (0.8.44), and Vary (1.7.7).

uwsgi_ignore_headers_s :: [Text] -> Uwsgi_Ignore_Headers

Safe version of uwsgi_ignore_headers

newtype Uwsgi_Ignore_Headers

Constructors

Uwsgi_Ignore_Headers ConfStatement

uwsgi_intercept_errors :: [Text] -> ConfStatement

Determines whether a uwsgi server responses with codes greater than or equal to 300 should be passed to a client or be redirected to nginx for processing with the directive.

uwsgi_intercept_errors_s :: [Text] -> Uwsgi_Intercept_Errors

Safe version of uwsgi_intercept_errors

newtype Uwsgi_Intercept_Errors

Constructors

Uwsgi_Intercept_Errors ConfStatement

uwsgi_limit_rate :: [Text] -> ConfStatement

Limits the speed of reading the response from the uwsgi server. The rate is specified in bytes per second. The zero value disables rate limiting. The limit is set per a request, and so if nginx simultaneously opens two connections to the uwsgi server, the overall rate will be twice as much as the specified limit. The limitation works only if buffering of responses from the uwsgi server is enabled.

uwsgi_limit_rate_s :: [Text] -> Uwsgi_Limit_Rate

Safe version of uwsgi_limit_rate

newtype Uwsgi_Limit_Rate

Constructors

Uwsgi_Limit_Rate ConfStatement

uwsgi_max_temp_file_size :: [Text] -> ConfStatement

When buffering of responses from the uwsgi server is enabled, and the whole response does not fit into the buffers set by the and directives, a part of the response can be saved to a temporary file. This directive sets the maximum size of the temporary file. The size of data written to the temporary file at a time is set by the directive.

uwsgi_max_temp_file_size_s :: [Text] -> Uwsgi_Max_Temp_File_Size

Safe version of uwsgi_max_temp_file_size

newtype Uwsgi_Max_Temp_File_Size

Constructors

Uwsgi_Max_Temp_File_Size ConfStatement

uwsgi_modifier1 :: [Text] -> ConfStatement

Sets the value of the modifier1 field in the uwsgi packet header.

uwsgi_modifier1_s :: [Text] -> Uwsgi_Modifier1

Safe version of uwsgi_modifier1

newtype Uwsgi_Modifier1

Constructors

Uwsgi_Modifier1 ConfStatement

uwsgi_modifier2 :: [Text] -> ConfStatement

Sets the value of the modifier2 field in the uwsgi packet header.

uwsgi_modifier2_s :: [Text] -> Uwsgi_Modifier2

Safe version of uwsgi_modifier2

newtype Uwsgi_Modifier2

Constructors

Uwsgi_Modifier2 ConfStatement

uwsgi_next_upstream :: [Text] -> ConfStatement

Specifies in which cases a request should be passed to the next server: error an error occurred while establishing a connection with the server, passing a request to it, or reading the response header; timeout a timeout has occurred while establishing a connection with the server, passing a request to it, or reading the response header; invalid_header a server returned an empty or invalid response; http_500 a server returned a response with the code 500; http_503 a server returned a response with the code 503; http_403 a server returned a response with the code 403; http_404 a server returned a response with the code 404; non_idempotent normally, requests with a non-idempotent method (POST, LOCK, PATCH) are not passed to the next server if a request has been sent to an upstream server (1.9.13); enabling this option explicitly allows retrying such requests; off disables passing a request to the next server.

uwsgi_next_upstream_s :: [Text] -> Uwsgi_Next_Upstream

Safe version of uwsgi_next_upstream

newtype Uwsgi_Next_Upstream

Constructors

Uwsgi_Next_Upstream ConfStatement

uwsgi_next_upstream_timeout :: [Text] -> ConfStatement

Limits the time during which a request can be passed to the next server. The 0 value turns off this limitation.

uwsgi_next_upstream_timeout_s :: [Text] -> Uwsgi_Next_Upstream_Timeout

Safe version of uwsgi_next_upstream_timeout

newtype Uwsgi_Next_Upstream_Timeout

Constructors

Uwsgi_Next_Upstream_Timeout ConfStatement

uwsgi_next_upstream_tries :: [Text] -> ConfStatement

Limits the number of possible tries for passing a request to the next server. The 0 value turns off this limitation.

uwsgi_next_upstream_tries_s :: [Text] -> Uwsgi_Next_Upstream_Tries

Safe version of uwsgi_next_upstream_tries

newtype Uwsgi_Next_Upstream_Tries

Constructors

Uwsgi_Next_Upstream_Tries ConfStatement

uwsgi_no_cache :: [Text] -> ConfStatement

Defines conditions under which the response will not be saved to a cache. If at least one value of the string parameters is not empty and is not equal to “0” then the response will not be saved: uwsgi_no_cache $cookie_nocache $arg_nocache$arg_comment; uwsgi_no_cache $http_pragma $http_authorization; Can be used along with the directive.

uwsgi_no_cache_s :: [Text] -> Uwsgi_No_Cache

Safe version of uwsgi_no_cache

newtype Uwsgi_No_Cache

Constructors

Uwsgi_No_Cache ConfStatement

uwsgi_param :: [Text] -> ConfStatement

Sets a parameter that should be passed to the uwsgi server. The value can contain text, variables, and their combination. These directives are inherited from the previous level if and only if there are no uwsgi_param directives defined on the current level.

uwsgi_param_s :: [Text] -> Uwsgi_Param

Safe version of uwsgi_param

newtype Uwsgi_Param

Constructors

Uwsgi_Param ConfStatement

uwsgi_pass :: [Text] -> ConfStatement

Sets the protocol and address of a uwsgi server. As a protocol, “uwsgi” or “suwsgi” (secured uwsgi, uwsgi over SSL) can be specified. The address can be specified as a domain name or IP address, and a port: uwsgi_pass localhost:9000; uwsgi_pass uwsgi:/localhost:9000; uwsgi_pass suwsgi:[2001:db8::1]:9090; or as a UNIX-domain socket path: uwsgi_pass unix:tmp/uwsgi.socket;

uwsgi_pass_s :: [Text] -> Uwsgi_Pass

Safe version of uwsgi_pass

newtype Uwsgi_Pass

Constructors

Uwsgi_Pass ConfStatement

uwsgi_pass_header :: [Text] -> ConfStatement

Permits passing otherwise disabled header fields from a uwsgi server to a client.

uwsgi_pass_header_s :: [Text] -> Uwsgi_Pass_Header

Safe version of uwsgi_pass_header

newtype Uwsgi_Pass_Header

Constructors

Uwsgi_Pass_Header ConfStatement

uwsgi_pass_request_body :: [Text] -> ConfStatement

Indicates whether the original request body is passed to the uwsgi server. See also the directive.

uwsgi_pass_request_body_s :: [Text] -> Uwsgi_Pass_Request_Body

Safe version of uwsgi_pass_request_body

newtype Uwsgi_Pass_Request_Body

Constructors

Uwsgi_Pass_Request_Body ConfStatement

uwsgi_pass_request_headers :: [Text] -> ConfStatement

Indicates whether the header fields of the original request are passed to the uwsgi server. See also the directive.

uwsgi_pass_request_headers_s :: [Text] -> Uwsgi_Pass_Request_Headers

Safe version of uwsgi_pass_request_headers

newtype Uwsgi_Pass_Request_Headers

Constructors

Uwsgi_Pass_Request_Headers ConfStatement

uwsgi_read_timeout :: [Text] -> ConfStatement

Defines a timeout for reading a response from the uwsgi server. The timeout is set only between two successive read operations, not for the transmission of the whole response. If the uwsgi server does not transmit anything within this time, the connection is closed.

uwsgi_read_timeout_s :: [Text] -> Uwsgi_Read_Timeout

Safe version of uwsgi_read_timeout

newtype Uwsgi_Read_Timeout

Constructors

Uwsgi_Read_Timeout ConfStatement

uwsgi_request_buffering :: [Text] -> ConfStatement

Enables or disables buffering of a client request body.

uwsgi_request_buffering_s :: [Text] -> Uwsgi_Request_Buffering

Safe version of uwsgi_request_buffering

newtype Uwsgi_Request_Buffering

Constructors

Uwsgi_Request_Buffering ConfStatement

uwsgi_send_timeout :: [Text] -> ConfStatement

Sets a timeout for transmitting a request to the uwsgi server. The timeout is set only between two successive write operations, not for the transmission of the whole request. If the uwsgi server does not receive anything within this time, the connection is closed.

uwsgi_send_timeout_s :: [Text] -> Uwsgi_Send_Timeout

Safe version of uwsgi_send_timeout

newtype Uwsgi_Send_Timeout

Constructors

Uwsgi_Send_Timeout ConfStatement

uwsgi_ssl_certificate :: [Text] -> ConfStatement

Specifies a file with the certificate in the PEM format used for authentication to a secured uwsgi server.

uwsgi_ssl_certificate_s :: [Text] -> Uwsgi_Ssl_Certificate

Safe version of uwsgi_ssl_certificate

newtype Uwsgi_Ssl_Certificate

Constructors

Uwsgi_Ssl_Certificate ConfStatement

uwsgi_ssl_certificate_key :: [Text] -> ConfStatement

Specifies a file with the secret key in the PEM format used for authentication to a secured uwsgi server.

uwsgi_ssl_certificate_key_s :: [Text] -> Uwsgi_Ssl_Certificate_Key

Safe version of uwsgi_ssl_certificate_key

newtype Uwsgi_Ssl_Certificate_Key

Constructors

Uwsgi_Ssl_Certificate_Key ConfStatement

uwsgi_ssl_ciphers :: [Text] -> ConfStatement

Specifies the enabled ciphers for requests to a secured uwsgi server. The ciphers are specified in the format understood by the OpenSSL library.

uwsgi_ssl_ciphers_s :: [Text] -> Uwsgi_Ssl_Ciphers

Safe version of uwsgi_ssl_ciphers

newtype Uwsgi_Ssl_Ciphers

Constructors

Uwsgi_Ssl_Ciphers ConfStatement

uwsgi_ssl_crl :: [Text] -> ConfStatement

Specifies a file with revoked certificates (CRL) in the PEM format used to verify the certificate of the secured uwsgi server.

uwsgi_ssl_crl_s :: [Text] -> Uwsgi_Ssl_Crl

Safe version of uwsgi_ssl_crl

newtype Uwsgi_Ssl_Crl

Constructors

Uwsgi_Ssl_Crl ConfStatement

uwsgi_ssl_name :: [Text] -> ConfStatement

Allows overriding the server name used to verify the certificate of the secured uwsgi server and to be passed through SNI when establishing a connection with the secured uwsgi server.

uwsgi_ssl_name_s :: [Text] -> Uwsgi_Ssl_Name

Safe version of uwsgi_ssl_name

newtype Uwsgi_Ssl_Name

Constructors

Uwsgi_Ssl_Name ConfStatement

uwsgi_ssl_password_file :: [Text] -> ConfStatement

Specifies a file with passphrases for secret keys where each passphrase is specified on a separate line. Passphrases are tried in turn when loading the key.

uwsgi_ssl_password_file_s :: [Text] -> Uwsgi_Ssl_Password_File

Safe version of uwsgi_ssl_password_file

newtype Uwsgi_Ssl_Password_File

Constructors

Uwsgi_Ssl_Password_File ConfStatement

uwsgi_ssl_protocols :: [Text] -> ConfStatement

Enables the specified protocols for requests to a secured uwsgi server.

uwsgi_ssl_protocols_s :: [Text] -> Uwsgi_Ssl_Protocols

Safe version of uwsgi_ssl_protocols

newtype Uwsgi_Ssl_Protocols

Constructors

Uwsgi_Ssl_Protocols ConfStatement

uwsgi_ssl_server_name :: [Text] -> ConfStatement

Enables or disables passing of the server name through TLS Server Name Indication extension (SNI, RFC 6066) when establishing a connection with the secured uwsgi server.

uwsgi_ssl_server_name_s :: [Text] -> Uwsgi_Ssl_Server_Name

Safe version of uwsgi_ssl_server_name

newtype Uwsgi_Ssl_Server_Name

Constructors

Uwsgi_Ssl_Server_Name ConfStatement

uwsgi_ssl_session_reuse :: [Text] -> ConfStatement

Determines whether SSL sessions can be reused when working with a secured uwsgi server. If the errors “SSL3_GET_FINISHED:digest check failed” appear in the logs, try disabling session reuse.

uwsgi_ssl_session_reuse_s :: [Text] -> Uwsgi_Ssl_Session_Reuse

Safe version of uwsgi_ssl_session_reuse

newtype Uwsgi_Ssl_Session_Reuse

Constructors

Uwsgi_Ssl_Session_Reuse ConfStatement

uwsgi_ssl_trusted_certificate :: [Text] -> ConfStatement

Specifies a file with trusted CA certificates in the PEM format used to verify the certificate of the secured uwsgi server.

uwsgi_ssl_trusted_certificate_s :: [Text] -> Uwsgi_Ssl_Trusted_Certificate

Safe version of uwsgi_ssl_trusted_certificate

newtype Uwsgi_Ssl_Trusted_Certificate

Constructors

Uwsgi_Ssl_Trusted_Certificate ConfStatement

uwsgi_ssl_verify :: [Text] -> ConfStatement

Enables or disables verification of the secured uwsgi server certificate.

uwsgi_ssl_verify_s :: [Text] -> Uwsgi_Ssl_Verify

Safe version of uwsgi_ssl_verify

newtype Uwsgi_Ssl_Verify

Constructors

Uwsgi_Ssl_Verify ConfStatement

uwsgi_ssl_verify_depth :: [Text] -> ConfStatement

Sets the verification depth in the secured uwsgi server certificates chain.

uwsgi_ssl_verify_depth_s :: [Text] -> Uwsgi_Ssl_Verify_Depth

Safe version of uwsgi_ssl_verify_depth

newtype Uwsgi_Ssl_Verify_Depth

Constructors

Uwsgi_Ssl_Verify_Depth ConfStatement

uwsgi_store :: [Text] -> ConfStatement

Enables saving of files to a disk. The on parameter saves files with paths corresponding to the directives or . The off parameter disables saving of files. In addition, the file name can be set explicitly using the string with variables: uwsgi_store datawww$original_uri;

uwsgi_store_s :: [Text] -> Uwsgi_Store

Safe version of uwsgi_store

newtype Uwsgi_Store

Constructors

Uwsgi_Store ConfStatement

uwsgi_store_access :: [Text] -> ConfStatement

Sets access permissions for newly created files and directories, e.g.: uwsgi_store_access user:rw group:rw all:r;

uwsgi_store_access_s :: [Text] -> Uwsgi_Store_Access

Safe version of uwsgi_store_access

newtype Uwsgi_Store_Access

Constructors

Uwsgi_Store_Access ConfStatement

uwsgi_temp_file_write_size :: [Text] -> ConfStatement

Limits the size of data written to a temporary file at a time, when buffering of responses from the uwsgi server to temporary files is enabled. By default, size is limited by two buffers set by the and directives. The maximum size of a temporary file is set by the directive.

uwsgi_temp_file_write_size_s :: [Text] -> Uwsgi_Temp_File_Write_Size

Safe version of uwsgi_temp_file_write_size

newtype Uwsgi_Temp_File_Write_Size

Constructors

Uwsgi_Temp_File_Write_Size ConfStatement

uwsgi_temp_path :: [Text] -> ConfStatement

Defines a directory for storing temporary files with data received from uwsgi servers. Up to three-level subdirectory hierarchy can be used underneath the specified directory. For example, in the following configuration uwsgi_temp_path spoolnginxuwsgi_temp 1 2; a temporary file might look like this: spoolnginxuwsgi_temp745/00000123457

uwsgi_temp_path_s :: [Text] -> Uwsgi_Temp_Path

Safe version of uwsgi_temp_path

newtype Uwsgi_Temp_Path

Constructors

Uwsgi_Temp_Path ConfStatement