Nginx_Ngx_Http_V2_Module

Documentation

http2_chunk_size :: [Text] -> ConfStatement

Sets the maximum size of chunks into which the response body is sliced. A too low value results in higher overhead. A too high value impairs prioritization due to HOL blocking.

http2_chunk_size_s :: [Text] -> Http2_Chunk_Size

Safe version of http2_chunk_size

newtype Http2_Chunk_Size

Constructors

Http2_Chunk_Size ConfStatement

http2_body_preread_size :: [Text] -> ConfStatement

Sets the size of the buffer per each request in which the request body may be saved before it is started to be processed.

http2_body_preread_size_s :: [Text] -> Http2_Body_Preread_Size

Safe version of http2_body_preread_size

newtype Http2_Body_Preread_Size

Constructors

Http2_Body_Preread_Size ConfStatement

http2_idle_timeout :: [Text] -> ConfStatement

Sets the timeout of inactivity after which the connection is closed.

http2_idle_timeout_s :: [Text] -> Http2_Idle_Timeout

Safe version of http2_idle_timeout

newtype Http2_Idle_Timeout

Constructors

Http2_Idle_Timeout ConfStatement

http2_max_concurrent_streams :: [Text] -> ConfStatement

Sets the maximum number of concurrent HTTP/2 streams in a connection.

http2_max_concurrent_streams_s :: [Text] -> Http2_Max_Concurrent_Streams

Safe version of http2_max_concurrent_streams

newtype Http2_Max_Concurrent_Streams

Constructors

Http2_Max_Concurrent_Streams ConfStatement

http2_max_field_size :: [Text] -> ConfStatement

Limits the maximum size of an HPACK-compressed request header field. The limit applies equally to both name and value. Note that if Huffman encoding is applied, the actual size of decompressed name and value strings may be larger. For most requests, the default limit should be enough.

http2_max_field_size_s :: [Text] -> Http2_Max_Field_Size

Safe version of http2_max_field_size

newtype Http2_Max_Field_Size

Constructors

Http2_Max_Field_Size ConfStatement

http2_max_header_size :: [Text] -> ConfStatement

Limits the maximum size of the entire request header list after HPACK decompression. For most requests, the default limit should be enough.

http2_max_header_size_s :: [Text] -> Http2_Max_Header_Size

Safe version of http2_max_header_size

newtype Http2_Max_Header_Size

Constructors

Http2_Max_Header_Size ConfStatement

http2_recv_buffer_size :: [Text] -> ConfStatement

Sets the size of the per worker input buffer.

http2_recv_buffer_size_s :: [Text] -> Http2_Recv_Buffer_Size

Safe version of http2_recv_buffer_size

newtype Http2_Recv_Buffer_Size

Constructors

Http2_Recv_Buffer_Size ConfStatement

http2_recv_timeout :: [Text] -> ConfStatement

Sets the timeout for expecting more data from the client, after which the connection is closed.

http2_recv_timeout_s :: [Text] -> Http2_Recv_Timeout

Safe version of http2_recv_timeout

newtype Http2_Recv_Timeout

Constructors

Http2_Recv_Timeout ConfStatement