Nginx_Ngx_Http_Xslt_Module

Documentation

xml_entities :: [Text] -> ConfStatement

Specifies the DTD file that declares character entities. This file is compiled at the configuration stage. For technical reasons, the module is unable to use the external subset declared in the processed XML, so it is ignored and a specially defined file is used instead. This file should not describe the XML structure. It is enough to declare just the required character entities, for example: nbsp "&#xa0;"

xml_entities_s :: [Text] -> Xml_Entities

Safe version of xml_entities

newtype Xml_Entities

Constructors

Xml_Entities ConfStatement

xslt_last_modified :: [Text] -> ConfStatement

Allows preserving the Last-Modified header field from the original response during XSLT transformations to facilitate response caching.

xslt_last_modified_s :: [Text] -> Xslt_Last_Modified

Safe version of xslt_last_modified

newtype Xslt_Last_Modified

Constructors

Xslt_Last_Modified ConfStatement

xslt_param :: [Text] -> ConfStatement

Defines the parameters for XSLT stylesheets. The value is treated as an XPath expression. The value can contain variables. To pass a string value to a stylesheet, the directive can be used.

xslt_param_s :: [Text] -> Xslt_Param

Safe version of xslt_param

newtype Xslt_Param

Constructors

Xslt_Param ConfStatement

xslt_string_param :: [Text] -> ConfStatement

Defines the string parameters for XSLT stylesheets. XPath expressions in the value are not interpreted. The value can contain variables.

xslt_string_param_s :: [Text] -> Xslt_String_Param

Safe version of xslt_string_param

newtype Xslt_String_Param

Constructors

Xslt_String_Param ConfStatement

xslt_stylesheet :: [Text] -> ConfStatement

Defines the XSLT stylesheet and its optional parameters. A stylesheet is compiled at the configuration stage.

xslt_stylesheet_s :: [Text] -> Xslt_Stylesheet

Safe version of xslt_stylesheet

newtype Xslt_Stylesheet

Constructors

Xslt_Stylesheet ConfStatement

xslt_types :: [Text] -> ConfStatement

Enables transformations in responses with the specified MIME types in addition to “textxml”. The special value “*” matches any MIME type (0.8.29). If the transformation result is an HTML response, its MIME type is changed to “texthtml”.

xslt_types_s :: [Text] -> Xslt_Types

Safe version of xslt_types

newtype Xslt_Types

Constructors

Xslt_Types ConfStatement