Nginx_Ngx_Mail_Auth_Http_Module

Documentation

auth_http :: [Text] -> ConfStatement

Sets the URL of the HTTP authentication server. The protocol is described below.

auth_http_s :: [Text] -> Auth_Http

Safe version of auth_http

newtype Auth_Http

Constructors

Auth_Http ConfStatement

auth_http_header :: [Text] -> ConfStatement

Appends the specified header to requests sent to the authentication server. This header can be used as the shared secret to verify that the request comes from nginx. For example: auth_http_header X-Auth-Key "secret_string";

auth_http_header_s :: [Text] -> Auth_Http_Header

Safe version of auth_http_header

newtype Auth_Http_Header

Constructors

Auth_Http_Header ConfStatement

auth_http_pass_client_cert :: [Text] -> ConfStatement

Appends the Auth-SSL-Cert header with the client certificate in the PEM format (urlencoded) to requests sent to the authentication server.

auth_http_pass_client_cert_s :: [Text] -> Auth_Http_Pass_Client_Cert

Safe version of auth_http_pass_client_cert

newtype Auth_Http_Pass_Client_Cert

Constructors

Auth_Http_Pass_Client_Cert ConfStatement

auth_http_timeout :: [Text] -> ConfStatement

Sets the timeout for communication with the authentication server.

auth_http_timeout_s :: [Text] -> Auth_Http_Timeout

Safe version of auth_http_timeout

newtype Auth_Http_Timeout

Constructors

Auth_Http_Timeout ConfStatement