Nginx_Ngx_Mail_Core_Module

Documentation

listen :: [Text] -> ConfStatement

Sets the address and port for the socket on which the server will accept requests. It is possible to specify just the port. The address can also be a hostname, for example: listen 127.0.0.1:110; listen *:110; listen 110; # same as *:110 listen localhost:110; IPv6 addresses (0.7.58) are specified in square brackets: listen [::1]:110; listen [::]:110; UNIX-domain sockets (1.3.5) are specified with the “unix:” prefix: listen unix:varrun/nginx.sock;

listen_s :: [Text] -> Listen

Safe version of listen

newtype Listen

Constructors

Listen ConfStatement

mail :: [Text] -> [ConfStatement] -> ConfStatement

Provides the configuration file context in which the mail server directives are specified.

mail_s :: (HasContextMail f, ToConfStatement f) => [Text] -> [f] -> ConfStatement

Safe version of mail

protocol :: [Text] -> ConfStatement

Sets the protocol for a proxied server. Supported protocols are IMAP, POP3, and SMTP.

protocol_s :: [Text] -> Protocol

Safe version of protocol

newtype Protocol

Constructors

Protocol ConfStatement

resolver :: [Text] -> ConfStatement

Configures name servers used to find the client’s hostname to pass it to the authentication server, and in the XCLIENT command when proxying SMTP. For example: resolver 127.0.0.1 [::1]:5353; An address can be specified as a domain name or IP address, and an optional port (1.3.1, 1.2.2). If port is not specified, the port 53 is used. Name servers are queried in a round-robin fashion. Before version 1.1.7, only a single name server could be configured. Specifying name servers using IPv6 addresses is supported starting from versions 1.3.1 and 1.2.2. By default, nginx caches answers using the TTL value of a response. An optional valid parameter allows overriding it: resolver 127.0.0.1 [::1]:5353 valid=30s; Before version 1.1.9, tuning of caching time was not possible, and nginx always cached answers for the duration of 5 minutes.

resolver_s :: [Text] -> Resolver

Safe version of resolver

newtype Resolver

Constructors

Resolver ConfStatement

resolver_timeout :: [Text] -> ConfStatement

Sets a timeout for DNS operations, for example: resolver_timeout 5s;

resolver_timeout_s :: [Text] -> Resolver_Timeout

Safe version of resolver_timeout

newtype Resolver_Timeout

Constructors

Resolver_Timeout ConfStatement

server :: [Text] -> [ConfStatement] -> ConfStatement

Sets the configuration for a server.

server_s :: (HasContextServer f, ToConfStatement f) => [Text] -> [f] -> ConfStatement

Safe version of server

server_name :: [Text] -> ConfStatement

Sets the server name that is used: in the initial POP3/SMTP server greeting; in the salt during the SASL CRAM-MD5 authentication; in the EHLO command when connecting to the SMTP backend, if the passing of the XCLIENT command is enabled.

server_name_s :: [Text] -> Server_Name

Safe version of server_name

newtype Server_Name

Constructors

Server_Name ConfStatement

timeout :: [Text] -> ConfStatement

Sets the timeout that is used before proxying to the backend starts.

timeout_s :: [Text] -> Timeout

Safe version of timeout

newtype Timeout

Constructors

Timeout ConfStatement