Nginx_Ngx_Mail_Imap_Module

Documentation

imap_auth :: [Text] -> ConfStatement

Sets permitted methods of authentication for IMAP clients. Supported methods are: login AUTH=LOGIN plain AUTH=PLAIN cram-md5 AUTH=CRAM-MD5. In order for this method to work, the password must be stored unencrypted.

imap_auth_s :: [Text] -> Imap_Auth

Safe version of imap_auth

newtype Imap_Auth

Constructors

Imap_Auth ConfStatement

imap_capabilities :: [Text] -> ConfStatement

Sets the IMAP protocol extensions list that is passed to the client in response to the CAPABILITY command. The authentication methods specified in the and STARTTLS directives are automatically added to this list if the directive is enabled.

imap_capabilities_s :: [Text] -> Imap_Capabilities

Safe version of imap_capabilities

newtype Imap_Capabilities

Constructors

Imap_Capabilities ConfStatement

imap_client_buffer :: [Text] -> ConfStatement

Sets the IMAP commands read buffer size. By default, the buffer size is equal to one memory page. This is either 4K or 8K, depending on a platform.

imap_client_buffer_s :: [Text] -> Imap_Client_Buffer

Safe version of imap_client_buffer

newtype Imap_Client_Buffer

Constructors

Imap_Client_Buffer ConfStatement