Nginx_Ngx_Mail_Pop3_Module

Documentation

pop3_auth :: [Text] -> ConfStatement

Sets permitted methods of authentication for POP3 clients. Supported methods are: plain USER/PASS, AUTH PLAIN, AUTH LOGIN. It is not possible to disable these methods. apop APOP. In order for this method to work, the password must be stored unencrypted. cram-md5 AUTH CRAM-MD5. In order for this method to work, the password must be stored unencrypted.

pop3_auth_s :: [Text] -> Pop3_Auth

Safe version of pop3_auth

newtype Pop3_Auth

Constructors

Pop3_Auth ConfStatement

pop3_capabilities :: [Text] -> ConfStatement

Sets the POP3 protocol extensions list that is passed to the client in response to the CAPA command. The authentication methods specified in the and (SASL extension) and STLS directives, are automatically added to this list if the directive is enabled.

pop3_capabilities_s :: [Text] -> Pop3_Capabilities

Safe version of pop3_capabilities

newtype Pop3_Capabilities

Constructors

Pop3_Capabilities ConfStatement