Nginx_Ngx_Mail_Proxy_Module

Documentation

proxy_buffer :: [Text] -> ConfStatement

Sets the size of the buffer used for proxying. By default, the buffer size is equal to one memory page. Depending on a platform, it is either 4K or 8K.

proxy_buffer_s :: [Text] -> Proxy_Buffer

Safe version of proxy_buffer

newtype Proxy_Buffer

Constructors

Proxy_Buffer ConfStatement

proxy_pass_error_message :: [Text] -> ConfStatement

Indicates whether to pass the error message obtained during the authentication on the backend to the client.

proxy_pass_error_message_s :: [Text] -> Proxy_Pass_Error_Message

Safe version of proxy_pass_error_message

newtype Proxy_Pass_Error_Message

Constructors

Proxy_Pass_Error_Message ConfStatement

proxy_timeout :: [Text] -> ConfStatement

Sets the timeout between two successive read or write operations on client or proxied server connections. If no data is transmitted within this time, the connection is closed.

proxy_timeout_s :: [Text] -> Proxy_Timeout

Safe version of proxy_timeout

newtype Proxy_Timeout

Constructors

Proxy_Timeout ConfStatement

xclient :: [Text] -> ConfStatement

Enables or disables the passing of the XCLIENT command with client parameters when connecting to the SMTP backend.

xclient_s :: [Text] -> Xclient

Safe version of xclient

newtype Xclient

Constructors

Xclient ConfStatement