Nginx_Ngx_Mail_Smtp_Module

Documentation

smtp_auth :: [Text] -> ConfStatement

Sets permitted methods of SASL authentication for SMTP clients. Supported methods are: login AUTH LOGIN plain AUTH PLAIN cram-md5 AUTH CRAM-MD5. In order for this method to work, the password must be stored unencrypted. none Authentication is not required.

smtp_auth_s :: [Text] -> Smtp_Auth

Safe version of smtp_auth

newtype Smtp_Auth

Constructors

Smtp_Auth ConfStatement

smtp_capabilities :: [Text] -> ConfStatement

Sets the SMTP protocol extensions list that is passed to the client in response to the EHLO command. Authentication methods specified in the directive are automatically added to this list.

smtp_capabilities_s :: [Text] -> Smtp_Capabilities

Safe version of smtp_capabilities

newtype Smtp_Capabilities

Constructors

Smtp_Capabilities ConfStatement