Nginx_Ngx_Mail_Ssl_Module

Documentation

ssl :: [Text] -> ConfStatement

Enables the SSL/TLS protocol for the given server.

ssl_s :: [Text] -> Ssl

Safe version of ssl

newtype Ssl

Constructors

Ssl ConfStatement

ssl_certificate :: [Text] -> ConfStatement

Specifies a file with the certificate in the PEM format for the given server. If intermediate certificates should be specified in addition to a primary certificate, they should be specified in the same file in the following order: the primary certificate comes first, then the intermediate certificates. A secret key in the PEM format may be placed in the same file.

ssl_certificate_s :: [Text] -> Ssl_Certificate

Safe version of ssl_certificate

newtype Ssl_Certificate

Constructors

Ssl_Certificate ConfStatement

ssl_certificate_key :: [Text] -> ConfStatement

Specifies a file with the secret key in the PEM format for the given server.

ssl_certificate_key_s :: [Text] -> Ssl_Certificate_Key

Safe version of ssl_certificate_key

newtype Ssl_Certificate_Key

Constructors

Ssl_Certificate_Key ConfStatement

ssl_ciphers :: [Text] -> ConfStatement

Specifies the enabled ciphers. The ciphers are specified in the format understood by the OpenSSL library, for example: ssl_ciphers ALL:!aNULL:!EXPORT56:RC4+RSA:+HIGH:+MEDIUM:+LOW:+SSLv2:+EXP;

ssl_ciphers_s :: [Text] -> Ssl_Ciphers

Safe version of ssl_ciphers

newtype Ssl_Ciphers

Constructors

Ssl_Ciphers ConfStatement

ssl_client_certificate :: [Text] -> ConfStatement

Specifies a file with trusted CA certificates in the PEM format used to verify client certificates.

ssl_client_certificate_s :: [Text] -> Ssl_Client_Certificate

Safe version of ssl_client_certificate

newtype Ssl_Client_Certificate

Constructors

Ssl_Client_Certificate ConfStatement

ssl_crl :: [Text] -> ConfStatement

Specifies a file with revoked certificates (CRL) in the PEM format used to verify client certificates.

ssl_crl_s :: [Text] -> Ssl_Crl

Safe version of ssl_crl

newtype Ssl_Crl

Constructors

Ssl_Crl ConfStatement

ssl_dhparam :: [Text] -> ConfStatement

Specifies a file with DH parameters for DHE ciphers.

ssl_dhparam_s :: [Text] -> Ssl_Dhparam

Safe version of ssl_dhparam

newtype Ssl_Dhparam

Constructors

Ssl_Dhparam ConfStatement

ssl_ecdh_curve :: [Text] -> ConfStatement

Specifies a curve for ECDHE ciphers.

ssl_ecdh_curve_s :: [Text] -> Ssl_Ecdh_Curve

Safe version of ssl_ecdh_curve

newtype Ssl_Ecdh_Curve

Constructors

Ssl_Ecdh_Curve ConfStatement

ssl_password_file :: [Text] -> ConfStatement

Specifies a file with passphrases for secret keys where each passphrase is specified on a separate line. Passphrases are tried in turn when loading the key.

ssl_password_file_s :: [Text] -> Ssl_Password_File

Safe version of ssl_password_file

newtype Ssl_Password_File

Constructors

Ssl_Password_File ConfStatement

ssl_prefer_server_ciphers :: [Text] -> ConfStatement

Specifies that server ciphers should be preferred over client ciphers when the SSLv3 and TLS protocols are used.

ssl_prefer_server_ciphers_s :: [Text] -> Ssl_Prefer_Server_Ciphers

Safe version of ssl_prefer_server_ciphers

newtype Ssl_Prefer_Server_Ciphers

Constructors

Ssl_Prefer_Server_Ciphers ConfStatement

ssl_protocols :: [Text] -> ConfStatement

Enables the specified protocols. The TLSv1.1 and TLSv1.2 parameters work only when the OpenSSL library of version 1.0.1 or higher is used. The TLSv1.1 and TLSv1.2 parameters are supported starting from versions 1.1.13 and 1.0.12 so when the OpenSSL version 1.0.1 or higher is used on older nginx versions, these protocols work, but cannot be disabled.

ssl_protocols_s :: [Text] -> Ssl_Protocols

Safe version of ssl_protocols

newtype Ssl_Protocols

Constructors

Ssl_Protocols ConfStatement

ssl_session_cache :: [Text] -> ConfStatement

Sets the types and sizes of caches that store session parameters. A cache can be of any of the following types: off the use of a session cache is strictly prohibited: nginx explicitly tells a client that sessions may not be reused. none the use of a session cache is gently disallowed: nginx tells a client that sessions may be reused, but does not actually store session parameters in the cache. builtin a cache built in OpenSSL; used by one worker process only. The cache size is specified in sessions. If size is not given, it is equal to 20480 sessions. Use of the built-in cache can cause memory fragmentation. shared a cache shared between all worker processes. The cache size is specified in bytes; one megabyte can store about 4000 sessions. Each shared cache should have an arbitrary name. A cache with the same name can be used in several servers.

ssl_session_cache_s :: [Text] -> Ssl_Session_Cache

Safe version of ssl_session_cache

newtype Ssl_Session_Cache

Constructors

Ssl_Session_Cache ConfStatement

ssl_session_ticket_key :: [Text] -> ConfStatement

Sets a file with the secret key used to encrypt and decrypt TLS session tickets. The directive is necessary if the same key has to be shared between multiple servers. By default, a randomly generated key is used.

ssl_session_ticket_key_s :: [Text] -> Ssl_Session_Ticket_Key

Safe version of ssl_session_ticket_key

newtype Ssl_Session_Ticket_Key

Constructors

Ssl_Session_Ticket_Key ConfStatement

ssl_session_tickets :: [Text] -> ConfStatement

Enables or disables session resumption through TLS session tickets.

ssl_session_tickets_s :: [Text] -> Ssl_Session_Tickets

Safe version of ssl_session_tickets

newtype Ssl_Session_Tickets

Constructors

Ssl_Session_Tickets ConfStatement

ssl_session_timeout :: [Text] -> ConfStatement

Specifies a time during which a client may reuse the session parameters stored in a cache.

ssl_session_timeout_s :: [Text] -> Ssl_Session_Timeout

Safe version of ssl_session_timeout

newtype Ssl_Session_Timeout

Constructors

Ssl_Session_Timeout ConfStatement

ssl_trusted_certificate :: [Text] -> ConfStatement

Specifies a file with trusted CA certificates in the PEM format used to verify client certificates.

ssl_trusted_certificate_s :: [Text] -> Ssl_Trusted_Certificate

Safe version of ssl_trusted_certificate

newtype Ssl_Trusted_Certificate

Constructors

Ssl_Trusted_Certificate ConfStatement

ssl_verify_client :: [Text] -> ConfStatement

Enables verification of client certificates. The verification result is passed in the Auth-SSL-Verify header of the authentication request.

ssl_verify_client_s :: [Text] -> Ssl_Verify_Client

Safe version of ssl_verify_client

newtype Ssl_Verify_Client

Constructors

Ssl_Verify_Client ConfStatement

ssl_verify_depth :: [Text] -> ConfStatement

Sets the verification depth in the client certificates chain.

ssl_verify_depth_s :: [Text] -> Ssl_Verify_Depth

Safe version of ssl_verify_depth

newtype Ssl_Verify_Depth

Constructors

Ssl_Verify_Depth ConfStatement

starttls :: [Text] -> ConfStatement

on allow usage of the STLS command for the POP3 and the STARTTLS command for the IMAP; off deny usage of the STLS and STARTTLS commands; only require preliminary TLS transition.

starttls_s :: [Text] -> Starttls

Safe version of starttls

newtype Starttls

Constructors

Starttls ConfStatement