Nginx_Ngx_Stream_Core_Module

Documentation

listen :: [Text] -> ConfStatement

Sets the address and port for the socket on which the server will accept connections. It is possible to specify just the port. The address can also be a hostname, for example: listen 127.0.0.1:12345; listen *:12345; listen 12345; # same as *:12345 listen localhost:12345; IPv6 addresses are specified in square brackets: listen [::1]:12345; listen [::]:12345; UNIX-domain sockets are specified with the “unix:” prefix: listen unix:varrun/nginx.sock;

listen_s :: [Text] -> Listen

Safe version of listen

newtype Listen

Constructors

Listen ConfStatement

resolver :: [Text] -> ConfStatement

Configures name servers used to resolve names of upstream servers into addresses, for example: resolver 127.0.0.1 [::1]:5353; An address can be specified as a domain name or IP address, and an optional port. If port is not specified, the port 53 is used. Name servers are queried in a round-robin fashion.

resolver_s :: [Text] -> Resolver

Safe version of resolver

newtype Resolver

Constructors

Resolver ConfStatement

resolver_timeout :: [Text] -> ConfStatement

Sets a timeout for name resolution, for example: resolver_timeout 5s; Before version 1.11.3, this directive was available as part of our commercial subscription.

resolver_timeout_s :: [Text] -> Resolver_Timeout

Safe version of resolver_timeout

newtype Resolver_Timeout

Constructors

Resolver_Timeout ConfStatement

server :: [Text] -> [ConfStatement] -> ConfStatement

Sets the configuration for a server.

server_s :: (HasContextServer f, ToConfStatement f) => [Text] -> [f] -> ConfStatement

Safe version of server

stream :: [Text] -> [ConfStatement] -> ConfStatement

Provides the configuration file context in which the stream server directives are specified.

stream_s :: (HasContextStream f, ToConfStatement f) => [Text] -> [f] -> ConfStatement

Safe version of stream

tcp_nodelay :: [Text] -> ConfStatement

Enables or disables the use of the TCP_NODELAY option. The option is enabled for both client and proxied server connections.

tcp_nodelay_s :: [Text] -> Tcp_Nodelay

Safe version of tcp_nodelay

newtype Tcp_Nodelay

Constructors

Tcp_Nodelay ConfStatement

variables_hash_bucket_size :: [Text] -> ConfStatement

Sets the bucket size for the variables hash table. The details of setting up hash tables are provided in a separate document.

variables_hash_bucket_size_s :: [Text] -> Variables_Hash_Bucket_Size

Safe version of variables_hash_bucket_size

newtype Variables_Hash_Bucket_Size

Constructors

Variables_Hash_Bucket_Size ConfStatement

variables_hash_max_size :: [Text] -> ConfStatement

Sets the maximum size of the variables hash table. The details of setting up hash tables are provided in a separate document.

variables_hash_max_size_s :: [Text] -> Variables_Hash_Max_Size

Safe version of variables_hash_max_size

newtype Variables_Hash_Max_Size

Constructors

Variables_Hash_Max_Size ConfStatement