Nginx_Ngx_Stream_Geo_Module

Documentation

geo :: [Text] -> [ConfStatement] -> ConfStatement

Describes the dependency of values of the specified variable on the client IP address. By default, the address is taken from the $remote_addr variable, but it can also be taken from another variable, for example: geo $arg_remote_addr $geo { ...; }

geo_s :: (HasContextGeo f, ToConfStatement f) => [Text] -> [f] -> ConfStatement

Safe version of geo