Nginx_Ngx_Stream_Limit_Conn_Module

Documentation

limit_conn :: [Text] -> ConfStatement

Sets the shared memory zone and the maximum allowed number of connections for a given key value. When this limit is exceeded, the server will close the connection. For example, the directives limit_conn_zone $binary_remote_addr zone=addr:10m; server { ... limit_conn addr 1; } allow only one connection per an IP address at a time.

limit_conn_s :: [Text] -> Limit_Conn

Safe version of limit_conn

newtype Limit_Conn

Constructors

Limit_Conn ConfStatement

limit_conn_log_level :: [Text] -> ConfStatement

Sets the desired logging level for cases when the server limits the number of connections.

limit_conn_log_level_s :: [Text] -> Limit_Conn_Log_Level

Safe version of limit_conn_log_level

newtype Limit_Conn_Log_Level

Constructors

Limit_Conn_Log_Level ConfStatement

limit_conn_zone :: [Text] -> ConfStatement

Sets parameters for a shared memory zone that will keep states for various keys. In particular, the state includes the current number of connections. The key can contain text, variables, and their combinations (1.11.2). Connections with an empty key value are not accounted. Usage example: limit_conn_zone $binary_remote_addr zone=addr:10m; Here, the key is a client IP address set by the $binary_remote_addr variable. The size of $binary_remote_addr is 4 bytes for IPv4 addresses or 16 bytes for IPv6 addresses. The stored state always occupies 32 or 64 bytes on 32-bit platforms and 64 bytes on 64-bit platforms. One megabyte zone can keep about 32 thousand 32-byte states or about 16 thousand 64-byte states. If the zone storage is exhausted, the server will close the connection.

limit_conn_zone_s :: [Text] -> Limit_Conn_Zone

Safe version of limit_conn_zone

newtype Limit_Conn_Zone

Constructors

Limit_Conn_Zone ConfStatement