Nginx_Ngx_Stream_Map_Module

Documentation

map :: [Text] -> [ConfStatement] -> ConfStatement

Creates a new variable whose value depends on values of one or more of the source variables specified in the first parameter.

map_s :: (HasContextMap f, ToConfStatement f) => [Text] -> [f] -> ConfStatement

Safe version of map

map_hash_bucket_size :: [Text] -> ConfStatement

Sets the bucket size for the variables hash tables. Default value depends on the processor’s cache line size. The details of setting up hash tables are provided in a separate document.

map_hash_bucket_size_s :: [Text] -> Map_Hash_Bucket_Size

Safe version of map_hash_bucket_size

newtype Map_Hash_Bucket_Size

Constructors

Map_Hash_Bucket_Size ConfStatement

map_hash_max_size :: [Text] -> ConfStatement

Sets the maximum size of the variables hash tables. The details of setting up hash tables are provided in a separate document.

map_hash_max_size_s :: [Text] -> Map_Hash_Max_Size

Safe version of map_hash_max_size

newtype Map_Hash_Max_Size

Constructors

Map_Hash_Max_Size ConfStatement