Nginx_Ngx_Stream_Proxy_Module

Documentation

proxy_bind :: [Text] -> ConfStatement

Makes outgoing connections to a proxied server originate from the specified local IP address. Parameter value can contain variables (1.11.2). The special value off cancels the effect of the proxy_bind directive inherited from the previous configuration level, which allows the system to auto-assign the local IP address.

proxy_bind_s :: [Text] -> Proxy_Bind

Safe version of proxy_bind

newtype Proxy_Bind

Constructors

Proxy_Bind ConfStatement

proxy_buffer_size :: [Text] -> ConfStatement

Sets the size of the buffer used for reading data from the proxied server. Also sets the size of the buffer used for reading data from the client.

proxy_buffer_size_s :: [Text] -> Proxy_Buffer_Size

Safe version of proxy_buffer_size

newtype Proxy_Buffer_Size

Constructors

Proxy_Buffer_Size ConfStatement

proxy_connect_timeout :: [Text] -> ConfStatement

Defines a timeout for establishing a connection with a proxied server.

proxy_connect_timeout_s :: [Text] -> Proxy_Connect_Timeout

Safe version of proxy_connect_timeout

newtype Proxy_Connect_Timeout

Constructors

Proxy_Connect_Timeout ConfStatement

proxy_download_rate :: [Text] -> ConfStatement

Limits the speed of reading the data from the proxied server. The rate is specified in bytes per second. The zero value disables rate limiting. The limit is set per a connection, so if nginx simultaneously opens two connections to the proxied server, the overall rate will be twice as much as the specified limit.

proxy_download_rate_s :: [Text] -> Proxy_Download_Rate

Safe version of proxy_download_rate

newtype Proxy_Download_Rate

Constructors

Proxy_Download_Rate ConfStatement

proxy_next_upstream :: [Text] -> ConfStatement

When a connection to the proxied server cannot be established, determines whether a client connection will be passed to the next server.

proxy_next_upstream_s :: [Text] -> Proxy_Next_Upstream

Safe version of proxy_next_upstream

newtype Proxy_Next_Upstream

Constructors

Proxy_Next_Upstream ConfStatement

proxy_next_upstream_timeout :: [Text] -> ConfStatement

Limits the time allowed to pass a connection to the next server. The 0 value turns off this limitation.

proxy_next_upstream_timeout_s :: [Text] -> Proxy_Next_Upstream_Timeout

Safe version of proxy_next_upstream_timeout

newtype Proxy_Next_Upstream_Timeout

Constructors

Proxy_Next_Upstream_Timeout ConfStatement

proxy_next_upstream_tries :: [Text] -> ConfStatement

Limits the number of possible tries for passing a connection to the next server. The 0 value turns off this limitation.

proxy_next_upstream_tries_s :: [Text] -> Proxy_Next_Upstream_Tries

Safe version of proxy_next_upstream_tries

newtype Proxy_Next_Upstream_Tries

Constructors

Proxy_Next_Upstream_Tries ConfStatement

proxy_pass :: [Text] -> ConfStatement

Sets the address of a proxied server. The address can be specified as a domain name or IP address, and a port: proxy_pass localhost:12345; or as a UNIX-domain socket path: proxy_pass unix:tmpstream.socket;

proxy_pass_s :: [Text] -> Proxy_Pass

Safe version of proxy_pass

newtype Proxy_Pass

Constructors

Proxy_Pass ConfStatement

proxy_protocol :: [Text] -> ConfStatement

Enables the PROXY protocol for connections to a proxied server.

proxy_protocol_s :: [Text] -> Proxy_Protocol

Safe version of proxy_protocol

newtype Proxy_Protocol

Constructors

Proxy_Protocol ConfStatement

proxy_responses :: [Text] -> ConfStatement

Sets the number of datagrams expected from the proxied server in response to the client request if the UDP protocol is used. By default, the number of datagrams is not limited: the response datagrams will be sent until the value expires.

proxy_responses_s :: [Text] -> Proxy_Responses

Safe version of proxy_responses

newtype Proxy_Responses

Constructors

Proxy_Responses ConfStatement

proxy_ssl :: [Text] -> ConfStatement

Enables the SSL/TLS protocol for connections to a proxied server.

proxy_ssl_s :: [Text] -> Proxy_Ssl

Safe version of proxy_ssl

newtype Proxy_Ssl

Constructors

Proxy_Ssl ConfStatement

proxy_ssl_certificate :: [Text] -> ConfStatement

Specifies a file with the certificate in the PEM format used for authentication to a proxied server.

proxy_ssl_certificate_s :: [Text] -> Proxy_Ssl_Certificate

Safe version of proxy_ssl_certificate

newtype Proxy_Ssl_Certificate

Constructors

Proxy_Ssl_Certificate ConfStatement

proxy_ssl_certificate_key :: [Text] -> ConfStatement

Specifies a file with the secret key in the PEM format used for authentication to a proxied server.

proxy_ssl_certificate_key_s :: [Text] -> Proxy_Ssl_Certificate_Key

Safe version of proxy_ssl_certificate_key

newtype Proxy_Ssl_Certificate_Key

Constructors

Proxy_Ssl_Certificate_Key ConfStatement

proxy_ssl_ciphers :: [Text] -> ConfStatement

Specifies the enabled ciphers for connections to a proxied server. The ciphers are specified in the format understood by the OpenSSL library.

proxy_ssl_ciphers_s :: [Text] -> Proxy_Ssl_Ciphers

Safe version of proxy_ssl_ciphers

newtype Proxy_Ssl_Ciphers

Constructors

Proxy_Ssl_Ciphers ConfStatement

proxy_ssl_crl :: [Text] -> ConfStatement

Specifies a file with revoked certificates (CRL) in the PEM format used to verify the certificate of the proxied server.

proxy_ssl_crl_s :: [Text] -> Proxy_Ssl_Crl

Safe version of proxy_ssl_crl

newtype Proxy_Ssl_Crl

Constructors

Proxy_Ssl_Crl ConfStatement

proxy_ssl_name :: [Text] -> ConfStatement

Allows overriding the server name used to verify the certificate of the proxied server and to be passed through SNI when establishing a connection with the proxied server. The server name can also be specified using variables (1.11.3).

proxy_ssl_name_s :: [Text] -> Proxy_Ssl_Name

Safe version of proxy_ssl_name

newtype Proxy_Ssl_Name

Constructors

Proxy_Ssl_Name ConfStatement

proxy_ssl_password_file :: [Text] -> ConfStatement

Specifies a file with passphrases for secret keys where each passphrase is specified on a separate line. Passphrases are tried in turn when loading the key.

proxy_ssl_password_file_s :: [Text] -> Proxy_Ssl_Password_File

Safe version of proxy_ssl_password_file

newtype Proxy_Ssl_Password_File

Constructors

Proxy_Ssl_Password_File ConfStatement

proxy_ssl_server_name :: [Text] -> ConfStatement

Enables or disables passing of the server name through TLS Server Name Indication extension (SNI, RFC 6066) when establishing a connection with the proxied server.

proxy_ssl_server_name_s :: [Text] -> Proxy_Ssl_Server_Name

Safe version of proxy_ssl_server_name

newtype Proxy_Ssl_Server_Name

Constructors

Proxy_Ssl_Server_Name ConfStatement

proxy_ssl_session_reuse :: [Text] -> ConfStatement

Determines whether SSL sessions can be reused when working with the proxied server. If the errors “SSL3_GET_FINISHED:digest check failed” appear in the logs, try disabling session reuse.

proxy_ssl_session_reuse_s :: [Text] -> Proxy_Ssl_Session_Reuse

Safe version of proxy_ssl_session_reuse

newtype Proxy_Ssl_Session_Reuse

Constructors

Proxy_Ssl_Session_Reuse ConfStatement

proxy_ssl_protocols :: [Text] -> ConfStatement

Enables the specified protocols for connections to a proxied server.

proxy_ssl_protocols_s :: [Text] -> Proxy_Ssl_Protocols

Safe version of proxy_ssl_protocols

newtype Proxy_Ssl_Protocols

Constructors

Proxy_Ssl_Protocols ConfStatement

proxy_ssl_trusted_certificate :: [Text] -> ConfStatement

Specifies a file with trusted CA certificates in the PEM format used to verify the certificate of the proxied server.

proxy_ssl_trusted_certificate_s :: [Text] -> Proxy_Ssl_Trusted_Certificate

Safe version of proxy_ssl_trusted_certificate

newtype Proxy_Ssl_Trusted_Certificate

Constructors

Proxy_Ssl_Trusted_Certificate ConfStatement

proxy_ssl_verify :: [Text] -> ConfStatement

Enables or disables verification of the proxied server certificate.

proxy_ssl_verify_s :: [Text] -> Proxy_Ssl_Verify

Safe version of proxy_ssl_verify

newtype Proxy_Ssl_Verify

Constructors

Proxy_Ssl_Verify ConfStatement

proxy_ssl_verify_depth :: [Text] -> ConfStatement

Sets the verification depth in the proxied server certificates chain.

proxy_ssl_verify_depth_s :: [Text] -> Proxy_Ssl_Verify_Depth

Safe version of proxy_ssl_verify_depth

newtype Proxy_Ssl_Verify_Depth

Constructors

Proxy_Ssl_Verify_Depth ConfStatement

proxy_timeout :: [Text] -> ConfStatement

Sets the timeout between two successive read or write operations on client or proxied server connections. If no data is transmitted within this time, the connection is closed.

proxy_timeout_s :: [Text] -> Proxy_Timeout

Safe version of proxy_timeout

newtype Proxy_Timeout

Constructors

Proxy_Timeout ConfStatement

proxy_upload_rate :: [Text] -> ConfStatement

Limits the speed of reading the data from the client. The rate is specified in bytes per second. The zero value disables rate limiting. The limit is set per a connection, so if the client simultaneously opens two connections, the overall rate will be twice as much as the specified limit.

proxy_upload_rate_s :: [Text] -> Proxy_Upload_Rate

Safe version of proxy_upload_rate

newtype Proxy_Upload_Rate

Constructors

Proxy_Upload_Rate ConfStatement