Nginx_Ngx_Stream_Upstream_Module

Documentation

upstream :: [Text] -> [ConfStatement] -> ConfStatement

Defines a group of servers. Servers can listen on different ports. In addition, servers listening on TCP and UNIX-domain sockets can be mixed.

upstream_s :: (HasContextUpstream f, ToConfStatement f) => [Text] -> [f] -> ConfStatement

Safe version of upstream

server :: [Text] -> ConfStatement

Defines the address and other parameters of a server. The address can be specified as a domain name or IP address with an obligatory port, or as a UNIX-domain socket path specified after the “unix:” prefix. A domain name that resolves to several IP addresses defines multiple servers at once.

server_s :: [Text] -> Server

Safe version of server

newtype Server

Constructors

Server ConfStatement

zone :: [Text] -> ConfStatement

Defines the name and size of the shared memory zone that keeps the group’s configuration and run-time state that are shared between worker processes. Several groups may share the same zone. In this case, it is enough to specify the zone size only once.

zone_s :: [Text] -> Zone

Safe version of zone

newtype Zone

Constructors

Zone ConfStatement

state :: [Text] -> ConfStatement

Specifies a file that keeps the state of the dynamically configurable group. The state is currently limited to the list of servers with their parameters. The file is read when parsing the configuration and is updated each time the upstream configuration is changed. Changing the file content directly should be avoided. The directive cannot be used along with the directive.

state_s :: [Text] -> State

Safe version of state

newtype State

Constructors

State ConfStatement

hash :: [Text] -> ConfStatement

Specifies a load balancing method for a server group where client-server mapping is based on the hashed key value. The key can contain text, variables, and their combinations (1.11.2). Usage example: hash $remote_addr; Note that adding or removing a server from the group may result in remapping most of the keys to different servers. The method is compatible with the Cache::Memcached Perl library.

hash_s :: [Text] -> Hash

Safe version of hash

newtype Hash

Constructors

Hash ConfStatement

least_conn :: [Text] -> ConfStatement

Specifies that a server group should use a load balancing method where a connection is passed to the server with the least number of active connections, taking into account weights of servers. If there are several such servers, they are tried in turn using a weighted round-robin balancing method.

least_conn_s :: [Text] -> Least_Conn

Safe version of least_conn

newtype Least_Conn

Constructors

Least_Conn ConfStatement

least_time :: [Text] -> ConfStatement

Specifies that a group should use a load balancing method where a connection is passed to the server with the least average time and least number of active connections, taking into account weights of servers. If there are several such servers, they are tried in turn using a weighted round-robin balancing method.

least_time_s :: [Text] -> Least_Time

Safe version of least_time

newtype Least_Time

Constructors

Least_Time ConfStatement

health_check :: [Text] -> ConfStatement

Enables periodic health checks of the servers in a group.

health_check_s :: [Text] -> Health_Check

Safe version of health_check

newtype Health_Check

Constructors

Health_Check ConfStatement

health_check_timeout :: [Text] -> ConfStatement

Overrides the value for health checks.

health_check_timeout_s :: [Text] -> Health_Check_Timeout

Safe version of health_check_timeout

newtype Health_Check_Timeout

Constructors

Health_Check_Timeout ConfStatement

match :: [Text] -> [ConfStatement] -> ConfStatement

Defines the named test set used to verify server responses to health checks.

match_s :: (HasContextMatch f, ToConfStatement f) => [Text] -> [f] -> ConfStatement

Safe version of match